fix cypress tests
