+++ package.json
@@ -59,7 +59,7 @@
     "@dropins/storefront-purchase-order": "~1.1.1",
     "@dropins/storefront-quote-management": "~1.1.2",
     "@dropins/storefront-recommendations": "^4.0.3",
-    "@dropins/storefront-requisition-list": "1.4.0-alpha-20260526092218",
+    "@dropins/storefront-requisition-list": "1.4.0-beta.0",
     "@dropins/storefront-wishlist": "~3.2.0",
     "@dropins/storefront-quick-order": "~1.0.1",
     "@dropins/tools": "~1.8.1"
