--- scripts/__dropins__/storefront-requisition-list/chunks/PaginationItemsCounter.js
@@ -1,4 +1,0 @@
-/*! Copyright 2026 Adobe
-All Rights Reserved. */
-import{jsxs as d,jsx as s}from"@dropins/tools/preact-jsx-runtime.js";import{Icon as p,Button as f,Picker as h}from"@dropins/tools/components.js";import{classes as P}from"@dropins/tools/lib.js";import"./RequisitionListView.js";import*as o from"@dropins/tools/preact-compat.js";import{useText as u}from"@dropins/tools/i18n.js";import{D as S}from"./RequisitionListForm.js";const E=e=>o.createElement("svg",{id:"Icon_Search_Base","data-name":"Icon \\u2013 Search \\u2013 Base",xmlns:"http://www.w3.org/2000/svg",width:24,height:24,fill:"none",viewBox:"0 0 24 24",...e},o.createElement("g",{id:"Large"},o.createElement("rect",{id:"Placement_area","data-name":"Placement area",width:24,height:24,fill:"#fff",opacity:0}),o.createElement("g",{id:"Search_icon","data-name":"Search icon",transform:"translate(3.75 3.75)"},o.createElement("circle",{vectorEffect:"non-scaling-stroke",id:"Ellipse_186","data-name":"Ellipse 186",cx:6,cy:6,r:6,fill:"none",stroke:"currentColor"}),o.createElement("line",{vectorEffect:"non-scaling-stroke",id:"Line_556","data-name":"Line 556",x2:6,y2:6,transform:"translate(10.5 10.5)",fill:"none",stroke:"currentColor"})))),C=({className:e,title:n="404 - Not Found",message:a="The requisition list you are looking for does not exist.",actionLabel:i,onAction:r,...c})=>d("div",{className:P(["not-found",e]),"data-testid":"not-found",...c,children:[s(p,{source:E,size:"64",stroke:"2"}),s("h2",{children:n}),a&&s("p",{children:a}),i&&r&&s(f,{variant:"primary",onClick:r,children:i})]}),y=({currentPageSize:e,onPageSizeChange:n,disabled:a=!1,pageSizeOptions:i=[10,25,50,100]})=>{const r=u({itemsPerPage:"RequisitionList.PageSizePicker.itemsPerPage"}),c=t=>{const m=t.target,g=parseInt(m.value,10);n(g)},l=i.map(t=>({value:t.toString(),text:t.toString()}));return s(h,{disabled:a,"data-testid":"page-size-picker",variant:"primary",size:"medium",value:e.toString(),options:l,handleSelect:c,"aria-label":r.itemsPerPage})},I=({pageInfo:e,totalCount:n,className:a=""})=>{const i=u({itemsCounter:"RequisitionList.PaginationItemsCounter.itemsCounter"});if(!e||!n)return null;const r=e.page_size??S,c=e.current_page??1,l=r*c,t=n,m=l-r+1,g=l>t?t:l;return s("span",{className:`pagination-items-counter ${a}`.trim(),children:i.itemsCounter.replace("{from}",m.toString()).replace("{to}",g.toString()).replace("{total}",t.toString())})};export{C as N,I as P,y as a};
-//# sourceMappingURL=PaginationItemsCounter.js.map
