--- scripts/__dropins__/storefront-requisition-list/chunks/RequisitionListActions.js
@@ -1,4 +1,0 @@
-/*! Copyright 2026 Adobe
-All Rights Reserved. */
-import{jsxs as o,jsx as n}from"@dropins/tools/preact-jsx-runtime.js";import{classes as r}from"@dropins/tools/lib.js";import{Icon as l}from"@dropins/tools/components.js";import"./RequisitionListView.js";import*as t from"@dropins/tools/preact-compat.js";import{useText as c}from"@dropins/tools/i18n.js";const d=e=>t.createElement("svg",{id:"Icon_Add_Base","data-name":"Icon \\u2013 Add \\u2013 Base",xmlns:"http://www.w3.org/2000/svg",width:24,height:24,viewBox:"0 0 24 24",...e},t.createElement("g",{id:"Large"},t.createElement("rect",{id:"Placement_area","data-name":"Placement area",width:24,height:24,fill:"#fff",opacity:0}),t.createElement("g",{id:"Add_icon","data-name":"Add icon",transform:"translate(9.734 9.737)"},t.createElement("line",{vectorEffect:"non-scaling-stroke",id:"Line_579","data-name":"Line 579",y2:12.7,transform:"translate(2.216 -4.087)",fill:"none",stroke:"currentColor"}),t.createElement("line",{vectorEffect:"non-scaling-stroke",id:"Line_580","data-name":"Line 580",x2:12.7,transform:"translate(-4.079 2.263)",fill:"none",stroke:"currentColor"})))),L=({selectable:e,className:a,onAddNew:s})=>{const i=c({addNewReqListBtn:"RequisitionList.AddNewReqList.addNewReqListBtn"});return o("button",{type:"button","aria-label":i.addNewReqListBtn,role:"button",className:r(["requisition-list-actions",["requisition-list-actions--selectable",e],a]),"data-testid":"requisition-list-actions-button",onClick:s,children:[n("span",{className:"requisition-list-actions__title","data-testid":"requisition-list-actions-button-text",children:i.addNewReqListBtn}),n(l,{source:d,size:"32"})]})};export{L as R};
-//# sourceMappingURL=RequisitionListActions.js.map
