--- scripts/__dropins__/storefront-requisition-list/chunks/RequisitionListForm.js
@@ -1,4 +1,0 @@
-/*! Copyright 2026 Adobe
-All Rights Reserved. */
-import{jsxs as o,jsx as e}from"@dropins/tools/preact-jsx-runtime.js";import{useState as m}from"@dropins/tools/preact-hooks.js";import{ProgressSpinner as M,InLineAlert as x,Field as f,Input as y,TextArea as D,Button as R}from"@dropins/tools/components.js";import{classes as c}from"@dropins/tools/lib.js";import{useText as B}from"@dropins/tools/i18n.js";const Z=10,g=3,G=40,H=255,P=/^[a-zA-Z0-9\s.,\-_!?'"()&]+$/,w=({className:l,mode:_,defaultValues:b={name:"",description:""},error:u=null,onSubmit:v,onCancel:F,...N})=>{const[s,T]=m(b),[E,d]=m({name:!1}),[r,p]=m(!1),i=B({actionCancel:"RequisitionList.RequisitionListForm.actionCancel",actionSave:"RequisitionList.RequisitionListForm.actionSave",requiredField:"RequisitionList.RequisitionListForm.requiredField",nameMinLength:"RequisitionList.RequisitionListForm.nameMinLength",nameInvalidCharacters:"RequisitionList.RequisitionListForm.nameInvalidCharacters",floatingLabel:"RequisitionList.RequisitionListForm.floatingLabel",placeholder:"RequisitionList.RequisitionListForm.placeholder",label:"RequisitionList.RequisitionListForm.label",updateTitle:"RequisitionList.RequisitionListForm.updateTitle",createTitle:"RequisitionList.RequisitionListForm.createTitle"}),q=n=>{const t=n.trim();return t?t.length<g?i.nameMinLength.replace("{min}",g.toString()):P.test(t)?"":i.nameInvalidCharacters:i.requiredField},L=n=>t=>{const a=t.target;T(I=>({...I,[n]:a.value}))},h=n=>()=>{d(t=>({...t,[n]:!0}))},S=async n=>{var a;if(n.preventDefault(),d({name:!0,description:!0}),!(q(s.name)||r)){p(!0);try{await v({name:s.name.trim(),description:((a=s.description)==null?void 0:a.trim())??""})}catch{p(!1)}}},A=E.name?q(s.name):"",C=_==="update"?i.updateTitle:i.createTitle;return o("div",{...N,className:c(["requisition-list-form",l]),children:[o("div",{className:"requisition-list-form__title",children:[C,r?e("div",{className:c(["requisition-list-form_progress-spinner",l]),"data-testid":"requisition-list-form-progress-spinner",children:e(M,{stroke:"4",size:"small"})}):null]}),u?e(x,{type:"error",className:"requisition-list-form__notification",variant:"secondary",heading:u,"data-testid":"requisition-list-alert"}):null,o("form",{className:c(["requisition-list-form__form",l]),onSubmit:S,children:[e(f,{error:A,disabled:r,children:e(y,{id:"requisition-list-form-name",name:"name",type:"text",floatingLabel:i.floatingLabel,placeholder:i.placeholder,maxLength:G,value:s.name,onChange:L("name"),onBlur:h("name")})}),e(f,{disabled:r,children:e(D,{id:"requisition-list-form-description",name:"description",label:i.label,placeholder:i.label,maxLength:H,value:s.description,onChange:L("description"),onBlur:h("description")})}),o("div",{className:"requisition-list-form__actions",children:[e(R,{type:"button",variant:"secondary",onClick:F,disabled:r,"data-testid":"requisition-list-form-cancel",children:i.actionCancel}),e(R,{type:"submit",disabled:r,"data-testid":"requisition-list-form-save",children:i.actionSave})]})]})]})};export{Z as D,w as R};
-//# sourceMappingURL=RequisitionListForm.js.map
