--- scripts/__dropins__/storefront-requisition-list/chunks/RequisitionListHeader.js
@@ -1,4 +1,0 @@
-/*! Copyright 2026 Adobe
-All Rights Reserved. */
-import{jsxs as u,jsx as s,Fragment as H}from"@dropins/tools/preact-jsx-runtime.js";import{useState as f,useCallback as h}from"@dropins/tools/preact-hooks.js";import{R as _}from"./RequisitionListModal.js";import{R as j}from"./RequisitionListForm.js";import{u as W}from"./updateRequisitionList.js";import{d as z}from"./deleteRequisitionList.js";import{s as J}from"./shareRequisitionListByEmail.js";import{S as P}from"./ShareRequisitionListContent.js";import{s as V}from"./state.js";import{useText as K}from"@dropins/tools/i18n.js";const Q=({name:i,description:o,backLink:n,actions:e,className:R,...L})=>u("div",{...L,className:`requisition-list-header ${R||""}`,children:[n&&s("div",{className:"requisition-list-header__back",children:u("a",{href:n.url,className:"requisition-list-header__back-link",onClick:n.onClick,children:[s("span",{className:"requisition-list-header__back-arrow",children:"<"}),n.label]})}),u("div",{className:"requisition-list-header__main",children:[u("div",{className:"requisition-list-header__title-section",children:[s("h1",{className:"requisition-list-header__title",children:i}),o&&s("p",{className:"requisition-list-header__description",children:o})]}),e&&(e.onRename||e.onDelete||e.onShare||e.shareDisabled)&&u("div",{className:"requisition-list-header__actions",children:[(e.onShare||e.shareDisabled)&&s("a",{href:"#",onClick:d=>{var t;d.preventDefault(),!e.shareDisabled&&((t=e.onShare)==null||t.call(e))},className:`requisition-list-header__action-link ${e.shareDisabled?"requisition-list-header__action-link--disabled":""}`,"data-testid":"share-list-btn","aria-disabled":e.shareDisabled?"true":"false","aria-label":e.shareDisabled&&e.shareDisabledReason?`${e.shareLabel} – ${e.shareDisabledReason}`:void 0,"data-disabled-reason":e.shareDisabled?e.shareDisabledReason:void 0,children:e.shareLabel}),e.onRename&&s("a",{href:"#",onClick:d=>{var t;d.preventDefault(),(t=e.onRename)==null||t.call(e)},className:"requisition-list-header__action-link","data-testid":"rename-list-btn",children:e.renameLabel}),e.onDelete&&s("a",{href:"#",onClick:d=>{var t;d.preventDefault(),(t=e.onDelete)==null||t.call(e)},className:"requisition-list-header__action-link","data-testid":"delete-list-btn",children:e.deleteLabel})]})]})]}),te=({requisitionList:i,routeRequisitionListGrid:o,onUpdate:n,onAlert:e,enrichConfigurableProducts:R,currentCustomerEmail:L,routeSharedRequisitionList:d})=>{var M,T;const[t,p]=f(!1),[S,b]=f(!1),[g,y]=f(!1),[x,C]=f(!1),[w,N]=f(!1),k=(M=V.config)==null?void 0:M.requisition_list_sharing_enabled,m=k!==!1&&k!=="0",v=Number(i.items_count??((T=i.items)==null?void 0:T.length)??0),q=v<=0||!V.isCompanyUser,a=K({actionBackToRequisitionLists:"RequisitionList.RequisitionListView.actionBackToRequisitionLists",actionRename:"RequisitionList.RequisitionListView.actionRename",actionDeleteList:"RequisitionList.RequisitionListView.actionDeleteList",actionShare:"RequisitionList.RequisitionListView.actionShare",shareDisabledReason:"RequisitionList.RequisitionListView.shareDisabledReason",shareDisabledNoCompany:"RequisitionList.RequisitionListView.shareDisabledNoCompany",shareListTitle:"RequisitionList.RequisitionListView.shareListTitle",deleteListTitle:"RequisitionList.RequisitionListView.deleteListTitle",deleteListMessage:"RequisitionList.RequisitionListView.deleteListMessage",confirmAction:"RequisitionList.RequisitionListWrapper.confirmAction",cancelAction:"RequisitionList.RequisitionListWrapper.cancelAction",updateTitle:"RequisitionList.RequisitionListForm.updateTitle"}),O=h(()=>{p(!0)},[]),B=h(async r=>{var l,c;try{const D=await W(i.uid,r.name,r.description,(l=i.page_info)==null?void 0:l.page_size,(c=i.page_info)==null?void 0:c.current_page,R);D?(n==null||n(D),p(!1),e==null||e({action:"update",type:"success",context:"requisitionList"})):e==null||e({action:"update",type:"error",context:"requisitionList"})}catch{e==null||e({action:"update",type:"error",context:"requisitionList"})}},[i.uid,i.page_info,n,e,R]),$=h(()=>{b(!0)},[]),E=h(async()=>{C(!0);try{if(await z(i.uid)){const l={action:"delete",type:"success",context:"requisitionList"};try{localStorage.setItem("requisitionListPendingAlert",JSON.stringify(l))}catch{}if(o){const c=o();c&&typeof c=="string"&&(window.location.href=c)}}else e==null||e({action:"delete",type:"error",context:"requisitionList"})}catch{e==null||e({action:"delete",type:"error",context:"requisitionList"})}finally{C(!1),b(!1)}},[i.uid,o,e]),F=h(()=>{y(!0)},[]),I=h(async r=>{N(!0);try{return await J(i.uid,r)}catch{return[{code:"SHARE_FAILED",message:"Unable to share list."}]}finally{N(!1)}},[i.uid]);return u(H,{children:[s(Q,{name:i.name,description:i.description,backLink:o?{url:"#",label:a.actionBackToRequisitionLists,onClick:r=>{r.preventDefault();const l=o();l&&typeof l=="string"&&(window.location.href=l)}}:void 0,actions:{onRename:O,onDelete:$,onShare:m&&!q?F:void 0,renameLabel:a.actionRename,deleteLabel:a.actionDeleteList,shareLabel:m?a.actionShare:void 0,shareDisabled:m&&q,shareDisabledReason:m&&q?v<=0?a.shareDisabledReason:a.shareDisabledNoCompany:void 0}}),t&&s(_,{isOpen:t,isLoading:!1,title:a.updateTitle,modalContent:s(j,{mode:"update",defaultValues:{name:i.name,description:i.description||""},onSubmit:B,onCancel:()=>p(!1)}),handleModalOnClose:()=>p(!1)}),S&&s(_,{isOpen:S,isLoading:x,title:a.deleteListTitle,modalContent:a.deleteListMessage,confirmBtnCaption:a.confirmAction,closeBtnCaption:a.cancelAction,handleModalOnClose:()=>b(!1),handleModalOnConfirm:E}),m&&g&&s(_,{isOpen:g,isLoading:w,title:a.shareListTitle,modalContent:s(P,{requisitionListUid:i.uid,isSubmitting:w,onSubmit:I,currentCustomerEmail:L,routeSharedRequisitionList:d}),handleModalOnClose:()=>y(!1)})]})};export{te as R};
-//# sourceMappingURL=RequisitionListHeader.js.map
