--- scripts/__dropins__/storefront-requisition-list/chunks/RequisitionListModal.js
@@ -1,4 +1,0 @@
-/*! Copyright 2026 Adobe
-All Rights Reserved. */
-import{jsx as i,jsxs as a}from"@dropins/tools/preact-jsx-runtime.js";import{Modal as c,ProgressSpinner as m,Button as d}from"@dropins/tools/components.js";const q=({isOpen:n,isLoading:s,title:t,modalContent:u,closeBtnCaption:r,confirmBtnCaption:l,handleModalOnClose:e,handleModalOnConfirm:o})=>n?i(c,{className:"requisition-list-modal--overlay","data-testid":"requisition-list-modal",size:"medium",centered:!1,title:t,onClose:e,backgroundDim:!0,clickToDismiss:!0,escapeToDismiss:!0,role:"dialog","aria-label":t,children:a("div",{className:"requisition-list-modal",children:[s?i("div",{className:"requisition-list-modal__spinner","data-testid":"progress-spinner",children:i(m,{stroke:"4",size:"large"})}):null,i("p",{children:u}),a("div",{className:"requisition-list-modal__buttons",children:[e&&r&&i(d,{"data-testid":"rl-modal-close-button",type:"button",onClick:e,variant:"secondary",disabled:s,children:r}),o&&l&&i(d,{"data-testid":"rl-modal-confirm-button",type:"button",onClick:o,disabled:s,children:l})]})]})}):null;export{q as R};
-//# sourceMappingURL=RequisitionListModal.js.map
