--- scripts/__dropins__/storefront-requisition-list/chunks/RequisitionListPicker.js
@@ -1,4 +1,0 @@
-/*! Copyright 2026 Adobe
-All Rights Reserved. */
-import{jsx as r,jsxs as q}from"@dropins/tools/preact-jsx-runtime.js";import{useState as l,useEffect as L}from"@dropins/tools/preact-compat.js";import{Card as p,Icon as v,Button as R}from"@dropins/tools/components.js";import{s as t,a as g,b as m,u as _}from"./state.js";import{g as b}from"./getRequisitionLists.js";import{events as h}from"@dropins/tools/event-bus.js";import{S as k}from"./RequisitionListView.js";let e=0;const S=()=>{const[d,a]=l(t.requisitionLists),[f,u]=l(t.requisitionListsLoading),[c,o]=l(e);return L(()=>{t.requisitionLists.length===0&&!t.requisitionListsLoading&&(g(!0),b(1,100).then(s=>{const n=(s==null?void 0:s.items)||[];m(n),e++}).catch(s=>{console.error("Error fetching requisition lists:",s),m([]),e++}).finally(()=>{g(!1)}))},[]),L(()=>{const s=h.on("requisitionLists/data",i=>{i&&(m(i),e++,o(e),a(t.requisitionLists),u(t.requisitionListsLoading))}),n=h.on("requisitionList/data",i=>{i&&(_(i),e++,o(e),a([...t.requisitionLists]),u(t.requisitionListsLoading))});return()=>{s==null||s.off(),n==null||n.off()}},[]),L(()=>{c!==e&&(a(t.requisitionLists),u(t.requisitionListsLoading),o(e))},[c]),{lists:d,loading:f}},w=({excludeUid:d,confirmLabel:a,disabled:f=!1,onConfirm:u})=>{const{lists:c}=S(),[o,s]=l(null),n=d?c.filter(i=>i.uid!==d):c;return r(p,{variant:"secondary",children:q("form",{onSubmit:i=>{i.preventDefault(),o&&u(o)},className:"requisition-list-picker__form",children:[r("div",{className:"requisition-list-picker__available-lists",children:n.map(i=>q(p,{variant:o===i.uid?"primary":"secondary",onClick:()=>s(i.uid),children:[r(v,{source:k}),r("span",{children:i.name})]},i.uid))}),r("div",{className:"requisition-list-picker__actions",children:r(R,{type:"submit",disabled:!o||f,children:a})})]})})};export{w as R,S as u};
-//# sourceMappingURL=RequisitionListPicker.js.map
