--- scripts/__dropins__/storefront-requisition-list/chunks/RequisitionListView.js
@@ -1,4 +1,0 @@
-/*! Copyright 2026 Adobe
-All Rights Reserved. */
-import*as i from"@dropins/tools/preact-compat.js";import{useState as m,useEffect as v,useMemo as q,useCallback as x}from"@dropins/tools/preact-compat.js";import{s as p}from"./state.js";import{events as h}from"@dropins/tools/event-bus.js";import{useText as E}from"@dropins/tools/i18n.js";import{jsxs as I,jsx as d}from"@dropins/tools/preact-jsx-runtime.js";import{Icon as M}from"@dropins/tools/components.js";import{classes as D}from"@dropins/tools/lib.js";function R(){const t=p.config;return t?t.is_requisition_list_active==="1"&&t.company_enabled===!0:!1}const j=()=>{const[t,r]=m(R);return v(()=>{const s=h.on("requisitionList/initialized",()=>{const o=R();r(e=>p.config!=null?o:e)});return()=>s==null?void 0:s.off()},[]),{isEnabled:t}},U=t=>i.createElement("svg",{width:24,height:24,viewBox:"0 0 24 24",fill:"none",xmlns:"http://www.w3.org/2000/svg",...t},i.createElement("rect",{x:4.89282,y:2.75,width:14.2143,height:18.5,rx:2.25,stroke:"currentColor",strokeWidth:1}),i.createElement("line",{x1:9.85718,y1:7.67871,x2:16.2857,y2:7.67871,stroke:"currentColor",strokeWidth:1}),i.createElement("line",{x1:9.85718,y1:11.9644,x2:16.2857,y2:11.9644,stroke:"currentColor",strokeWidth:1}),i.createElement("line",{x1:9.85718,y1:16.25,x2:16.2857,y2:16.25,stroke:"currentColor",strokeWidth:1}),i.createElement("circle",{cx:7.71429,cy:7.71429,r:.714286,fill:"currentColor"}),i.createElement("circle",{cx:7.71429,cy:11.9999,r:.714286,fill:"currentColor"}),i.createElement("circle",{cx:7.71429,cy:16.2856,r:.714286,fill:"currentColor"}));function N(t){const[r,s]=m(null),o=E({errorCreate:"RequisitionList.RequisitionListAlert.errorCreate",successCreate:"RequisitionList.RequisitionListAlert.successCreate",errorDeleteItem:"RequisitionList.RequisitionListAlert.errorDeleteItem",successDeleteItem:"RequisitionList.RequisitionListAlert.successDeleteItem",errorDeleteReqList:"RequisitionList.RequisitionListAlert.errorDeleteReqList",successDeleteReqList:"RequisitionList.RequisitionListAlert.successDeleteReqList",errorAddToCart:"RequisitionList.RequisitionListAlert.errorAddToCart",successAddToCart:"RequisitionList.RequisitionListAlert.successAddToCart",errorUpdateQuantity:"RequisitionList.RequisitionListAlert.errorUpdateQuantity",successUpdateQuantity:"RequisitionList.RequisitionListAlert.successUpdateQuantity",errorUpdate:"RequisitionList.RequisitionListAlert.errorUpdate",successUpdate:"RequisitionList.RequisitionListAlert.successUpdate",errorMove:"RequisitionList.RequisitionListAlert.errorMove",successMove:"RequisitionList.RequisitionListAlert.successMove",errorAddToRequisitionList:"RequisitionList.RequisitionListAlert.errorAddToRequisitionList",successAddToRequisitionList:"RequisitionList.RequisitionListAlert.successAddToRequisitionList",errorMoveToList:"RequisitionList.RequisitionListAlert.errorMoveToList",successMoveToList:"RequisitionList.RequisitionListAlert.successMoveToList",errorCopyToList:"RequisitionList.RequisitionListAlert.errorCopyToList",successCopyToList:"RequisitionList.RequisitionListAlert.successCopyToList",errorImport:"RequisitionList.RequisitionListAlert.errorImport",successImport:"RequisitionList.RequisitionListAlert.successImport"}),e=q(()=>({...o,...t}),[o,t]),l=q(()=>({create:{requisitionList:{success:e.successCreate,error:e.errorCreate}},add:{product:{success:e.successAddToRequisitionList,error:e.errorAddToRequisitionList}},update:{product:{success:e.successUpdateQuantity,error:e.errorUpdateQuantity},requisitionList:{success:e.successUpdate,error:e.errorUpdate}},delete:{product:{success:e.successDeleteItem,error:e.errorDeleteItem},requisitionList:{success:e.successDeleteReqList,error:e.errorDeleteReqList}},move:{product:{success:e.successMove,error:e.errorMove}},moveToList:{product:{success:e.successMoveToList,error:e.errorMoveToList}},copyToList:{product:{success:e.successCopyToList,error:e.errorCopyToList}},import:{requisitionList:{success:e.successImport,error:e.errorImport}}}),[e]),A=x(y=>{const{type:L,action:T,context:f,skus:c,message:u,listName:a}=y;let n=u&&u.length>0?u.join(". "):l[T][f][L];a&&(n=n.replace("{listName}",a)),s({type:L,description:n,sku:c==null?void 0:c[0]});const C=setTimeout(()=>{s(null)},5e3);return()=>clearTimeout(C)},[l]);return{alert:r,setAlert:s,handleRequisitionListAlert:A}}const S=({className:t,textContent:r,...s})=>I("div",{className:D(["empty-list",t]),"data-testid":"empty-list",...s,children:[d(M,{source:U,size:"64",stroke:"2"}),r&&d("h4",{children:r})]});export{S as E,U as S,N as a,j as u};
-//# sourceMappingURL=RequisitionListView.js.map
