--- scripts/__dropins__/storefront-requisition-list/chunks/ShareRequisitionListContent.js
@@ -1,4 +1,0 @@
-/*! Copyright 2026 Adobe
-All Rights Reserved. */
-import{jsxs as c,jsx as i,Fragment as P}from"@dropins/tools/preact-jsx-runtime.js";import{useState as n,useEffect as V,useCallback as A}from"@dropins/tools/preact-hooks.js";import{g as X,s as Y}from"./shareRequisitionListByToken.js";import{ProgressSpinner as F,Field as Z,MultiSelect as ii,Button as $,Divider as ei,Input as ti}from"@dropins/tools/components.js";/* empty css                            */import{useText as z}from"@dropins/tools/i18n.js";import{s as T}from"./state.js";const si=({loadingUsers:R,usersErrorMessage:u,loadingLink:_,selectedUserValues:d,multiSelectOptions:h,shareLink:m,linkErrorMessage:a,linkCopied:k,isSubmitting:S,canSubmit:g,onSubmitClick:b,onCopyLinkClick:y,onSelectedUsersChange:r,onUsersFieldInteract:f,selectionError:l,submitErrorMessage:L,isShareSuccess:v,sharedRecipientEmails:N})=>{const t=z({emailInstruction:"RequisitionList.ShareRequisitionListContent.emailInstruction",emailLabel:"RequisitionList.ShareRequisitionListContent.emailLabel",emailPlaceholder:"RequisitionList.ShareRequisitionListContent.emailPlaceholder",submitLabel:"RequisitionList.ShareRequisitionListContent.submitLabel",linkInstruction:"RequisitionList.ShareRequisitionListContent.linkInstruction",copyLink:"RequisitionList.ShareRequisitionListContent.copyLink",linkCopied:"RequisitionList.ShareRequisitionListContent.linkCopied",loadingUsers:"RequisitionList.ShareRequisitionListContent.loadingUsers",loadingLink:"RequisitionList.ShareRequisitionListContent.loadingLink",noUsersAvailable:"RequisitionList.ShareRequisitionListContent.noUsersAvailable",maxRecipientsValidation:"RequisitionList.ShareRequisitionListContent.maxRecipientsValidation",shareSuccessMessage:"RequisitionList.ShareRequisitionListContent.shareSuccessMessage"});return c("div",{className:"share-requisition-list-content",children:[v?c("div",{className:"share-requisition-list-content__success",children:[i("p",{className:"share-requisition-list-content__instruction",children:t.shareSuccessMessage}),i("div",{className:"share-requisition-list-content__recipient-list",children:N.map(C=>i("p",{className:"share-requisition-list-content__recipient",children:C},C))})]}):c(P,{children:[i("p",{className:"share-requisition-list-content__instruction",children:t.emailInstruction}),i("div",{className:"share-requisition-list-content__field",children:R?c("div",{className:"share-requisition-list-content__loading",children:[i(F,{size:"small",stroke:"3"}),i("span",{children:t.loadingUsers})]}):u?i("p",{className:"dropin-field__hint dropin-field__hint--medium dropin-field__hint--error",children:u}):i(Z,{label:t.emailLabel,error:l??void 0,disabled:S,onMouseDown:f,onKeyDown:f,children:i(ii,{options:h,value:d,onChange:r,placeholder:t.emailPlaceholder,noResultsText:t.noUsersAvailable,disabled:S,error:!!l,className:"share-requisition-list-content__multi-select"})})}),i("div",{className:"share-requisition-list-content__actions",children:i($,{variant:"primary",onClick:b,disabled:!g,type:"button","data-testid":"share-submit-btn",children:t.submitLabel})}),L&&i("p",{className:"dropin-field__hint dropin-field__hint--medium dropin-field__hint--error",children:L})]}),i(ei,{variant:"secondary",className:"share-requisition-list-content__divider-secondary"}),i("p",{className:"share-requisition-list-content__instruction",children:t.linkInstruction}),_?c("div",{className:"share-requisition-list-content__loading",children:[i(F,{size:"small",stroke:"3"}),i("span",{children:t.loadingLink})]}):m?c(P,{children:[i("div",{className:"share-requisition-list-content__link-row",children:i(ti,{readOnly:!0,value:m,"data-testid":"share-link-field"})}),i("div",{className:"share-requisition-list-content__actions",children:i($,{variant:"secondary",onClick:y,type:"button","data-testid":"copy-link-btn",children:k?t.linkCopied:t.copyLink})})]}):a?i("p",{className:"dropin-field__hint dropin-field__hint--medium dropin-field__hint--error",children:a}):null]})},di=({requisitionListUid:R,isSubmitting:u,onSubmit:_,currentCustomerEmail:d,routeSharedRequisitionList:h})=>{const m=z({emailInstruction:"RequisitionList.ShareRequisitionListContent.emailInstruction",emailLabel:"RequisitionList.ShareRequisitionListContent.emailLabel",emailPlaceholder:"RequisitionList.ShareRequisitionListContent.emailPlaceholder",submitLabel:"RequisitionList.ShareRequisitionListContent.submitLabel",linkInstruction:"RequisitionList.ShareRequisitionListContent.linkInstruction",copyLink:"RequisitionList.ShareRequisitionListContent.copyLink",linkCopied:"RequisitionList.ShareRequisitionListContent.linkCopied",loadingUsers:"RequisitionList.ShareRequisitionListContent.loadingUsers",loadingLink:"RequisitionList.ShareRequisitionListContent.loadingLink",noUsersAvailable:"RequisitionList.ShareRequisitionListContent.noUsersAvailable",usersLoadError:"RequisitionList.ShareRequisitionListContent.usersLoadError",maxRecipientsValidation:"RequisitionList.ShareRequisitionListContent.maxRecipientsValidation",shareSuccessMessage:"RequisitionList.ShareRequisitionListContent.shareSuccessMessage"}),[a,k]=n([]),[S,g]=n(!0),[b,y]=n(!1),[r,f]=n(new Set),[l,L]=n(null),[v,N]=n(!0),[t,C]=n(null),[D,w]=n(!1),[U,x]=n(null),[E,I]=n(null),[j,B]=n(!1),[O,K]=n([]),M=(()=>{var o;const e=(o=T.config)==null?void 0:o.requisition_list_share_max_recipients,s=Number(e);return Number.isFinite(s)&&s>0?s:null})();V(()=>{X().then(e=>{const s=e.filter(o=>!(d&&o.email.toLowerCase()===d.toLowerCase()));k(s)}).catch(()=>{y(!0)}).finally(()=>g(!1))},[d]),V(()=>{Y(R).then(e=>{var s;if(e.token){const p=`/${((s=T.config)==null?void 0:s.requisition_list_share_storefront_path)??""}?requisition_id=${e.token}`,q=h?h(p):p;L(q)}else L(null);C(e.errorMessage)}).finally(()=>N(!1))},[R,h]);const G=A(async()=>{var p;const e=Array.from(r),s=a.filter(q=>e.includes(String(q.id))).map(q=>q.email),o=await _(e);o?I(((p=o[0])==null?void 0:p.message)||null):(B(!0),K(s),I(null))},[r,a,_]),H=A(()=>{navigator.clipboard.writeText(l).then(()=>{w(!0),setTimeout(()=>w(!1),3e3)},e=>{console.error("Failed to copy share link to clipboard:",e)})},[l]),J=a.map(e=>({label:`${e.firstname} ${e.lastname} (${e.email})`,value:e.id})),Q=b?m.usersLoadError:null,W=!u&&r.size>0&&!U;return i(si,{loadingUsers:S,usersErrorMessage:Q,loadingLink:v,selectedUserValues:Array.from(r),multiSelectOptions:J,shareLink:l,linkErrorMessage:t,linkCopied:D,isSubmitting:u,canSubmit:W,onSubmitClick:G,onCopyLinkClick:H,onUsersFieldInteract:()=>{U&&x(null),E&&I(null)},onSelectedUsersChange:e=>{const s=e.map(String);if(M&&s.length>M){x(m.maxRecipientsValidation.replace("{max}",String(M)));return}x(null),f(new Set(s))},selectionError:U,submitErrorMessage:E,isShareSuccess:j,sharedRecipientEmails:O})};export{di as S};
-//# sourceMappingURL=ShareRequisitionListContent.js.map
