--- scripts/__dropins__/storefront-requisition-list/chunks/fetch-error.js
@@ -1,4 +1,0 @@
-/*! Copyright 2026 Adobe
-All Rights Reserved. */
-const s=r=>{const e=r.map(o=>o.message).join(" ");throw Error(e)};export{s as h};
-//# sourceMappingURL=fetch-error.js.map
