--- scripts/__dropins__/storefront-requisition-list/chunks/fetch-graphql.js
@@ -1,4 +1,0 @@
-/*! Copyright 2026 Adobe
-All Rights Reserved. */
-import{FetchGraphQL as e}from"@dropins/tools/fetch-graphql.js";const{setEndpoint:t,setFetchGraphQlHeader:r,removeFetchGraphQlHeader:s,setFetchGraphQlHeaders:h,fetchGraphQl:o,getConfig:p}=new e().getMethods();export{r as a,h as b,o as f,p as g,s as r,t as s};
-//# sourceMappingURL=fetch-graphql.js.map
