+++ scripts/__dropins__/storefront-requisition-list/containers/RequisitionListForm.js
@@ -1,4 +1,4 @@
 /*! Copyright 2026 Adobe
 All Rights Reserved. */
-import{R as n,R as q}from"../chunks/RequisitionListForm2.js";import"@dropins/tools/preact-jsx-runtime.js";import"../chunks/RequisitionListForm.js";import"@dropins/tools/preact-hooks.js";import"@dropins/tools/components.js";import"@dropins/tools/lib.js";import"@dropins/tools/i18n.js";import"../chunks/transform-requisition-list.js";import"../chunks/fetch-error.js";import"@dropins/tools/event-bus.js";import"../chunks/fetch-graphql.js";import"@dropins/tools/fetch-graphql.js";import"../chunks/updateRequisitionList.js";export{n as RequisitionListForm,q as default};
+import{R as d,R as l}from"./SharedRequisitionList.js";import"@dropins/tools/lib.js";import"@dropins/tools/i18n.js";import"@dropins/tools/preact-jsx-runtime.js";import"@dropins/tools/components.js";import"@dropins/tools/preact-compat.js";import"@dropins/tools/event-bus.js";import"../api.js";import"@dropins/tools/fetch-graphql.js";import"@dropins/tools/preact-hooks.js";import"@dropins/tools/preact.js";export{d as RequisitionListForm,l as default};
 //# sourceMappingURL=RequisitionListForm.js.map
