+++ scripts/__dropins__/storefront-requisition-list/containers/RequisitionListHeader.js
@@ -1,4 +1,4 @@
 /*! Copyright 2026 Adobe
 All Rights Reserved. */
-import{R as h}from"../chunks/RequisitionListHeader.js";import"@dropins/tools/preact-jsx-runtime.js";import"@dropins/tools/preact-hooks.js";import"../chunks/RequisitionListModal.js";import"@dropins/tools/components.js";import"../chunks/RequisitionListForm.js";import"@dropins/tools/lib.js";import"@dropins/tools/i18n.js";import"../chunks/updateRequisitionList.js";import"../chunks/transform-requisition-list.js";import"../chunks/fetch-error.js";import"@dropins/tools/event-bus.js";import"../chunks/fetch-graphql.js";import"@dropins/tools/fetch-graphql.js";import"../chunks/deleteRequisitionList.js";import"../chunks/shareRequisitionListByEmail.js";import"../chunks/ShareRequisitionListContent.js";import"../chunks/shareRequisitionListByToken.js";/* empty css                                    */import"../chunks/state.js";export{h as RequisitionListHeader};
+import{c as n}from"./SharedRequisitionList.js";import"@dropins/tools/lib.js";import"@dropins/tools/i18n.js";import"@dropins/tools/preact-jsx-runtime.js";import"@dropins/tools/components.js";import"@dropins/tools/preact-compat.js";import"@dropins/tools/event-bus.js";import"../api.js";import"@dropins/tools/fetch-graphql.js";import"@dropins/tools/preact-hooks.js";import"@dropins/tools/preact.js";export{n as RequisitionListHeader};
 //# sourceMappingURL=RequisitionListHeader.js.map
