+++ scripts/__dropins__/storefront-requisition-list/containers/RequisitionListSelector.js
@@ -1,23 +1,4 @@
 /*! Copyright 2026 Adobe
 All Rights Reserved. */
-import{jsxs as I,jsx as e,Fragment as D}from"@dropins/tools/preact-jsx-runtime.js";import*as b from"@dropins/tools/preact-compat.js";import{useState as v,useEffect as V,useMemo as H,useCallback as f}from"@dropins/tools/preact-compat.js";import{events as N}from"@dropins/tools/event-bus.js";import{Icon as y,InLineAlert as J,Button as X,Card as Y}from"@dropins/tools/components.js";import{h as Z}from"../chunks/fetch-error.js";import{a as K,R as tt,t as it}from"../chunks/transform-requisition-list.js";import{f as et}from"../chunks/fetch-graphql.js";import"@dropins/tools/preact-hooks.js";import"@dropins/tools/lib.js";import"../chunks/RequisitionListForm.js";import{R as st}from"../chunks/RequisitionListForm2.js";import{u as ot,a as nt,E as rt,S as P}from"../chunks/RequisitionListView.js";import"@dropins/tools/preact.js";import{u as at,R as ct}from"../chunks/RequisitionListPicker.js";import{g as lt}from"../chunks/state.js";/* empty css                                    *//* empty css                             */import{R as ut}from"../chunks/RequisitionListActions.js";import{R as dt}from"../chunks/RequisitionListModal.js";import{useText as mt}from"@dropins/tools/i18n.js";import"@dropins/tools/fetch-graphql.js";import"../chunks/updateRequisitionList.js";import"../chunks/getRequisitionLists.js";const Lt=`
-  mutation ADD_PRODUCTS_TO_REQUISITION_LIST_MUTATION(
-      $requisitionListUid: ID!, 
-      $requisitionListItems: [RequisitionListItemsInput!]!
-    ) {
-    addProductsToRequisitionList(
-      requisitionListUid: $requisitionListUid
-      requisitionListItems: $requisitionListItems
-    ) {
-      requisition_list {
-        ...REQUISITION_LIST_FRAGMENT
-        items {
-          ...REQUISITION_LIST_ITEMS_FRAGMENT
-        }
-      }
-    }
-  }
-${K}
-${tt}
-`,pt=async(c,i)=>{var n;const s=i.map(t=>{const r={sku:t.sku,quantity:t.quantity};return t.parent_sku&&(r.parent_sku=t.parent_sku),t.selected_options&&t.selected_options.length>0&&(r.selected_options=t.selected_options),t.entered_options&&t.entered_options.length>0&&(r.entered_options=t.entered_options),r}),{errors:u,data:o}=await et(Lt,{variables:{requisitionListUid:c,requisitionListItems:s}});if(u)return Z(u);if(!((n=o==null?void 0:o.addProductsToRequisitionList)!=null&&n.requisition_list))return null;const a=it(o.addProductsToRequisitionList.requisition_list);return N.emit("requisitionList/data",a),a},qt=c=>b.createElement("svg",{width:24,height:24,viewBox:"0 0 24 24",fill:"none",xmlns:"http://www.w3.org/2000/svg",...c},b.createElement("path",{vectorEffect:"non-scaling-stroke",d:"M7.74512 9.87701L12.0001 14.132L16.2551 9.87701",stroke:"currentColor",strokeWidth:1,strokeLinecap:"square",strokeLinejoin:"round"}));function ft(c,i,s){var n;const u=c.sku??((n=c.product)==null?void 0:n.sku);if(!u||u!==i.sku)return!1;if((s==null?void 0:s.matchBySkuOnly)!==!1)return!0;const o=(c.configurable_options??[]).map(t=>t.value_uid).filter(t=>!!t),a=(i.selectedOptions??[]).filter(t=>!!t);return o.length!==a.length?!1:(o.sort(),a.sort(),o.every((t,r)=>t===a[r]))}const $t=({canCreate:c=!0,sku:i,selectedOptions:s,quantity:u=1,matchBySKU:o,beforeAddProdToReqList:a})=>{const n=mt({createTitle:"RequisitionList.RequisitionListForm.createTitle",addToRequisitionList:"RequisitionList.RequisitionListForm.addToRequisitionList",emptyList:"RequisitionList.RequisitionListWrapper.emptyList",addToNewRequisitionList:"RequisitionList.RequisitionListSelector.addToNewRequisitionList",addToSelected:"RequisitionList.RequisitionListSelector.addToSelected"}),[t,r]=v(!1),{lists:R}=at(),{isEnabled:O}=ot(),[$,A]=v(null);V(()=>{const l=m=>{m&&A(m)},d=m=>{m&&A(W=>{const g=W??lt(),F=g.findIndex(E=>E.uid===m.uid);return F>=0?g.map((E,z)=>z===F?m:E):[...g,m]})},L=N.on("requisitionLists/data",l),q=N.on("requisitionList/data",d);return()=>{L==null||L.off(),q==null||q.off()}},[]);const p=$??R,w=H(()=>{if(!(p!=null&&p.length))return!1;const l={sku:i,selectedOptions:s};return p.some(d=>{var L;return(L=d.items)==null?void 0:L.some(q=>ft(q,l,{matchBySkuOnly:o}))})},[p,i,s,o]),{alert:_,setAlert:M,handleRequisitionListAlert:T}=nt(),[x,C]=v({isOpen:!1,isLoading:!1}),h=f(()=>{C({isOpen:!0,isLoading:!1})},[]),S=f(()=>{C({isOpen:!1,isLoading:!1}),r(!1),M(null)},[M]),U=f(async l=>{try{const d={sku:i,quantity:u,...s&&s.length>0?{selected_options:s}:{}};await pt(l,[d])}catch(d){throw console.error("Error adding product to list:",d),d}},[i,u,s]),k=f(async l=>{try{await U(l),T({action:"add",type:"success",context:"product",skus:[i]})}catch{T({action:"add",type:"error",context:"product",skus:[i]})}finally{setTimeout(()=>{S()},2e3)}},[i,U,S,T]),Q=f(()=>{if(!a){h();return}Promise.resolve(a()).then(()=>{h()}).catch(()=>{})},[a,h]),G=(R==null?void 0:R.length)>0?t?I("button",{type:"button","aria-label":"Select a requisition list",role:"button",className:"requisition-list-actions","data-testid":"requisition-list-actions-button",onClick:()=>r(!1),children:[e("span",{className:"requisition-list-actions__title","data-testid":"requisition-list-actions-button-text",children:n.addToRequisitionList}),e(y,{source:qt,size:"32"})]}):e(ct,{confirmLabel:n.addToSelected,onConfirm:k}):e(rt,{textContent:n.emptyList}),j=t?e(Y,{variant:"secondary",children:e(st,{mode:"create",onSuccess:async l=>{await k(l.uid)},onError:()=>{T({action:"add",type:"error",context:"product",skus:[i]})},onCancel:()=>{r(!1)}})}):e(ut,{onAddNew:()=>{r(!0)}}),B=I(D,{children:[_&&e("div",{className:"requisition-list__alert-wrapper",children:e(J,{id:`requisition-list-selector__alert__${i}`,heading:_.description,type:_.type,variant:"primary",className:"requisition-list-selector__alert"},`requisition-list-selector__alert__${i}`)}),!_&&I(D,{children:[G,c&&j]})]});return O===null||!O?null:I("div",{className:"requisition-list-selector",children:[e(X,{active:w,activeIcon:e(y,{source:P}),"aria-label":n.addToRequisitionList,className:w?"requisition-list-selector--active":void 0,"data-testid":"requisition-list-selector",size:"medium",variant:"tertiary",icon:e(y,{source:P}),onClick:Q}),x.isOpen&&e(dt,{isOpen:!0,isLoading:x.isLoading,title:n.addToRequisitionList,modalContent:B,handleModalOnClose:S})]})};export{$t as RequisitionListSelector,$t as default};
+import{b as u,b as c}from"./SharedRequisitionList.js";import"@dropins/tools/lib.js";import"@dropins/tools/i18n.js";import"@dropins/tools/preact-jsx-runtime.js";import"@dropins/tools/components.js";import"@dropins/tools/preact-compat.js";import"@dropins/tools/event-bus.js";import"../api.js";import"@dropins/tools/fetch-graphql.js";import"@dropins/tools/preact-hooks.js";import"@dropins/tools/preact.js";export{u as RequisitionListSelector,c as default};
 //# sourceMappingURL=RequisitionListSelector.js.map
