+++ scripts/__dropins__/storefront-requisition-list/containers/RequisitionListView.js
@@ -1,4 +1,4 @@
 /*! Copyright 2026 Adobe
 All Rights Reserved. */
-import{jsxs as E,jsx as i,Fragment as Ct}from"@dropins/tools/preact-jsx-runtime.js";import*as v from"@dropins/tools/preact-compat.js";import{useState as S,useCallback as D,useEffect as dt}from"@dropins/tools/preact-compat.js";import{m as Ht,c as Bt,a as $t,d as Ft,u as Ot,g as pt}from"../chunks/copyItemsBetweenRequisitionLists.js";import{Button as at,Icon as lt,Price as Rt,Field as zt,Input as Wt,Image as Qt,Checkbox as jt,Table as Zt,ProgressSpinner as Kt,InLineAlert as Xt,Pagination as Jt}from"@dropins/tools/components.js";import{events as et}from"@dropins/tools/event-bus.js";import{useState as rt,useCallback as mt}from"@dropins/tools/preact-hooks.js";import{VComponent as Yt,classes as Gt}from"@dropins/tools/lib.js";import{D as ft}from"../chunks/RequisitionListForm.js";import{i as Ut}from"../chunks/getRequisitionLists.js";import{a as te,u as ee,E as ie}from"../chunks/RequisitionListView.js";import{h as oe}from"@dropins/tools/preact.js";import{useText as ct}from"@dropins/tools/i18n.js";import{R as se}from"../chunks/RequisitionListHeader.js";import"../chunks/state.js";/* empty css                                    *//* empty css                             */import{R as Lt}from"../chunks/RequisitionListModal.js";import{N as St,P as ne,a as ae}from"../chunks/PaginationItemsCounter.js";import{R as Mt}from"../chunks/RequisitionListPicker.js";import"../chunks/transform-requisition-list.js";import"../chunks/fetch-error.js";import"../chunks/fetch-graphql.js";import"@dropins/tools/fetch-graphql.js";import"../chunks/updateRequisitionList.js";import"../chunks/deleteRequisitionList.js";import"../chunks/shareRequisitionListByEmail.js";import"../chunks/ShareRequisitionListContent.js";import"../chunks/shareRequisitionListByToken.js";const le=r=>v.createElement("svg",{width:24,height:24,viewBox:"0 0 24 24",fill:"none",xmlns:"http://www.w3.org/2000/svg",...r},v.createElement("g",{clipPath:"url(#clip0_102_196)"},v.createElement("path",{vectorEffect:"non-scaling-stroke",d:"M18.3601 18.16H6.5601L4.8801 3H2.3501M19.6701 19.59C19.6701 20.3687 19.0388 21 18.2601 21C17.4814 21 16.8501 20.3687 16.8501 19.59C16.8501 18.8113 17.4814 18.18 18.2601 18.18C19.0388 18.18 19.6701 18.8113 19.6701 19.59ZM7.42986 19.59C7.42986 20.3687 6.79858 21 6.01986 21C5.24114 21 4.60986 20.3687 4.60986 19.59C4.60986 18.8113 5.24114 18.18 6.01986 18.18C6.79858 18.18 7.42986 18.8113 7.42986 19.59Z",stroke:"currentColor",strokeLinejoin:"round"}),v.createElement("path",{vectorEffect:"non-scaling-stroke",d:"M5.25 6.37L20.89 8.06L20.14 14.8H6.19",stroke:"currentColor",strokeLinejoin:"round"})),v.createElement("defs",null,v.createElement("clipPath",{id:"clip0_102_196"},v.createElement("rect",{vectorEffect:"non-scaling-stroke",width:19.29,height:19.5,fill:"white",transform:"translate(2.3501 2.25)"})))),re=r=>v.createElement("svg",{width:24,height:24,viewBox:"0 0 24 24",fill:"none",xmlns:"http://www.w3.org/2000/svg",...r},v.createElement("path",{d:"M17.3332 11.75H6.6665",strokeWidth:1,strokeLinecap:"square",strokeLinejoin:"round",vectorEffect:"non-scaling-stroke",fill:"none",stroke:"currentColor"})),It=r=>v.createElement("svg",{xmlns:"http://www.w3.org/2000/svg",width:24,height:24,viewBox:"0 0 24 24",fill:"none",...r},v.createElement("path",{vectorEffect:"non-scaling-stroke",d:"M1 5H23",stroke:"currentColor",strokeWidth:1,strokeMiterlimit:10}),v.createElement("path",{vectorEffect:"non-scaling-stroke",d:"M17.3674 22H6.63446C5.67952 22 4.88992 21.2688 4.8379 20.3338L4 5H20L19.1621 20.3338C19.1119 21.2688 18.3223 22 17.3655 22H17.3674Z",stroke:"currentColor",strokeWidth:1,strokeMiterlimit:10}),v.createElement("path",{vectorEffect:"non-scaling-stroke",d:"M9.87189 2H14.1281C14.6085 2 15 2.39766 15 2.88889V5H9V2.88889C9 2.39912 9.39006 2 9.87189 2Z",stroke:"currentColor",strokeWidth:1,strokeMiterlimit:10}),v.createElement("path",{vectorEffect:"non-scaling-stroke",d:"M8.87402 8.58057L9.39348 17.682",stroke:"currentColor",strokeWidth:1,strokeMiterlimit:10}),v.createElement("path",{vectorEffect:"non-scaling-stroke",d:"M14.6673 8.58057L14.146 17.682",stroke:"currentColor",strokeWidth:1,strokeMiterlimit:10}));function ce(){const[r,b]=rt(null),[k,h]=rt(new Set),H=mt((y,f)=>{h(I=>{const M=new Set(I);return f?M.add(y):M.delete(y),M})},[]),P=mt(()=>{var f;const y=(f=r==null?void 0:r.items)==null?void 0:f.map(I=>I.uid);h(new Set(y))},[r]),d=mt(()=>{h(new Set)},[]);return{currentRequisitionList:r,setCurrentRequisitionList:b,selectedItems:k,setSelectedItems:h,handleItemSelection:H,handleSelectAll:P,handleSelectNone:d}}function ue({sourceListUid:r,selectedItems:b,currentPageSize:k,currentPage:h,enrichConfigurableProductsInList:H,fetchAndMergeProducts:P,setCurrentRequisitionList:d,setSelectedItems:y,handleRequisitionListAlert:f}){const[I,M]=S(!1),[O,N]=S(!1),[T,w]=S(!1),[p,B]=S(!1),z=ct({successMoveToList:"RequisitionList.RequisitionListAlert.successMoveToList",successCopyToList:"RequisitionList.RequisitionListAlert.successCopyToList"}),J=D(async $=>{var m;if(b.size!==0){N(!0);try{const x=await Ht(r,$,Array.from(b),k,h);if(x!=null&&x.sourceList){const it=await H(x.sourceList),W=await P(it);d(W),y(new Set);const Q=((m=x.destinationList)==null?void 0:m.name)||"";f({action:"moveToList",type:"success",context:"product",message:[z.successMoveToList.replace("{listName}",Q)]})}else f({action:"moveToList",type:"error",context:"product"})}catch{f({action:"moveToList",type:"error",context:"product"})}finally{N(!1),M(!1)}}},[b,r,k,h,H,P,d,y,f,z.successMoveToList]),Y=D(async $=>{if(b.size!==0){B(!0);try{const m=await Bt(r,$,Array.from(b));if(m!=null&&m.destinationList){y(new Set);const x=m.destinationList.name||"";f({action:"copyToList",type:"success",context:"product",message:[z.successCopyToList.replace("{listName}",x)]})}else f({action:"copyToList",type:"error",context:"product"})}catch{f({action:"copyToList",type:"error",context:"product"})}finally{B(!1),w(!1)}}},[b,r,y,f,z.successCopyToList]);return{showMoveToListModal:I,setShowMoveToListModal:M,movingToList:O,showCopyToListModal:T,setShowCopyToListModal:w,copyingToList:p,handleMoveToList:J,handleCopyToList:Y}}const de=({className:r,items:b,selectedItems:k,currentPage:h,pageSize:H,canEdit:P=!0,handleItemSelection:d,handleUpdateQuantity:y,onAddToCart:f,onDeleteItem:I,...M})=>{const[O,N]=rt({}),[T,w]=rt({}),p=ct({productNameHeader:"RequisitionList.RequisitionListView.productListTable.headers.productName",skuHeader:"RequisitionList.RequisitionListView.productListTable.headers.sku",priceHeader:"RequisitionList.RequisitionListView.productListTable.headers.price",quantityHeader:"RequisitionList.RequisitionListView.productListTable.headers.quantity",subtotalHeader:"RequisitionList.RequisitionListView.productListTable.headers.subtotal",actionsHeader:"RequisitionList.RequisitionListView.productListTable.headers.actions",actionAddToCart:"RequisitionList.RequisitionListView.actionAddToCart",actionDelete:"RequisitionList.RequisitionListView.actionDelete",actionSelect:"RequisitionList.RequisitionListView.actionSelect",itemQuantity:"RequisitionList.RequisitionListView.productListTable.itemQuantity",outOfStock:"RequisitionList.RequisitionListView.productListTable.outOfStock",onlyXLeftInStock:"RequisitionList.RequisitionListView.productListTable.onlyXLeftInStock"}),B=[{label:"#",key:"index"},{label:"item",key:"item"},{label:p.priceHeader,key:"price"},{label:p.quantityHeader,key:"quantity"},{label:p.subtotalHeader,key:"subtotal"}];P&&(B.unshift({label:p.actionSelect,key:"selector"}),B.push({label:p.actionsHeader,key:"actions"}));const z=(t,n)=>{const l=t.target.checked;d(n.uid,l)},J=t=>(h-1)*H+t+1,Y=t=>{var n,l;return((n=t.configured_product)==null?void 0:n.name)||((l=t.product)==null?void 0:l.name)||t.sku},$=t=>{var n,l,L,a,A,F;return(l=(n=t.configured_product)==null?void 0:n.images)!=null&&l.length?((L=t.configured_product.images[0])==null?void 0:L.url)||"":(A=(a=t.product)==null?void 0:a.images)!=null&&A.length&&((F=t.product.images[0])==null?void 0:F.url)||""},m=t=>{var n;return((n=t.product)==null?void 0:n.name)||t.sku},x=t=>{var n;return(n=t.configured_product)==null?void 0:n.name},it=t=>t.bundle_options?t.bundle_options.map((l,L)=>E("span",{className:"requisition-list-view-product-list-table__product-configurable-name",children:[l.values[0].label," x ",l.values[0].quantity]},l.values[0].label||L)):void 0,W=t=>{var n;return((n=t.configured_product)==null?void 0:n.sku)||t.sku},Q=t=>{var n,l,L,a,A,F,u,Z;return Array.isArray(t.bundle_options)&&t.bundle_options.length>0?t.bundle_options.reduce((K,_)=>{var st,nt,tt;return K+(((nt=(st=_.values[0])==null?void 0:st.priceV2)==null?void 0:nt.value)||0)*(((tt=_.values[0])==null?void 0:tt.quantity)||0)},0):((a=(L=(l=(n=t.configured_product)==null?void 0:n.price)==null?void 0:l.final)==null?void 0:L.amount)==null?void 0:a.value)||((Z=(u=(F=(A=t.product)==null?void 0:A.price)==null?void 0:F.final)==null?void 0:u.amount)==null?void 0:Z.value)||0},G=t=>{var n,l,L,a,A,F,u,Z;return((a=(L=(l=(n=t.configured_product)==null?void 0:n.price)==null?void 0:l.final)==null?void 0:L.amount)==null?void 0:a.currency)||((Z=(u=(F=(A=t.product)==null?void 0:A.price)==null?void 0:F.final)==null?void 0:u.amount)==null?void 0:Z.currency)},U=t=>Q(t)*t.quantity,ot=(t,n)=>{const l=+t.target.value;l>0&&!Number.isNaN(t.target.value)&&w(L=>({...L,[n.uid]:l}))},e=async(t,n)=>{let l=+t.target.value;if(!(l>0&&l!==n.quantity)){w(L=>({...L,[n.uid]:n.quantity}));return}N(L=>({...L,[n.uid]:!0}));try{await y(n.uid,l)}finally{N(L=>({...L,[n.uid]:!1}))}},g=b.map((t,n)=>({selector:i("label",{id:`item-selector-${t.sku}-label`,children:i(jt,{className:"requisition-list-view-product-list-table__checkbox",name:`item-selector-${t.sku}`,"aria-label":`${p.actionSelect} ${m(t)}`,"data-testid":`item-checkbox-${t.sku}`,onChange:l=>z(l,t),value:t.sku,checked:k.has(t.uid)})}),index:i("div",{className:"requisition-list-view-product-list-table__index-container",children:J(n)}),item:E("div",{className:"requisition-list-view-product-list-table__item-container",children:[i(Qt,{className:"requisition-list-view-product-list-table__thumbnail",alt:Y(t),src:$(t)}),E("div",{className:"requisition-list-view-product-list-table__item-details",children:[i("div",{className:"requisition-list-view-product-list-table__product-name",children:m(t)}),t.stock_status==="OUT_OF_STOCK"&&i("div",{className:"requisition-list-view-product-list-table__out-of-stock",children:p.outOfStock}),t.stock_status!=="OUT_OF_STOCK"&&t.only_x_left_in_stock!==null&&t.only_x_left_in_stock<t.quantity&&i("div",{className:"requisition-list-view-product-list-table__low-stock",children:p.onlyXLeftInStock.replace("{count}",String(t.only_x_left_in_stock))}),i("span",{className:"requisition-list-view-product-list-table__product-configurable-name",children:x(t)}),i("div",{className:"requisition-list-view-product-list-table__sku",children:W(t)}),it(t)]})]}),price:i(Rt,{className:"requisition-list-view-product-list-table__price",amount:Q(t),currency:G(t)}),quantity:i("span",{className:"requisition-list-view-product-list-table__quantity",children:i(zt,{disabled:!!O[t.uid],children:i(Wt,{id:`requisition-list-item-quantity-${t.sku}`,"data-testid":`requisition-list-item-quantity-${t.sku}`,name:"quantity",type:"text","aria-label":`${p.itemQuantity} - ${m(t)}`,value:T[t.uid]??t.quantity,onChange:l=>ot(l,t),onBlur:l=>e(l,t)})})}),subtotal:i(Rt,{className:"requisition-list-view-product-list-table__price",amount:U(t),currency:G(t)}),actions:E("div",{className:"requisition-list-view__bulk-actions",children:[i(at,{type:"button",onClick:()=>f([t.uid]),icon:i(lt,{source:le}),"aria-label":`${p.actionAddToCart} - ${m(t)}`,"data-testid":"product-list-table-add-to-cart-button"}),i(at,{type:"button",variant:"secondary",icon:i(lt,{source:It}),onClick:()=>I([t.uid]),"aria-label":`${p.actionDelete} - ${m(t)}`,"data-testid":"product-list-table-delete-button"})]})})),j=i(Zt,{columns:B,rowData:g,"data-testid":"product-list-table",mobileLayout:"stacked"});return i(Yt,{node:oe("div",{}),className:Gt(["requisition-list-view-product-list-table-container",r]),"data-testid":"product-list-table-container",...M,children:j})},pe=({selectedItems:r,deletingItemId:b,addingToCartItemId:k,bulkAddingToCart:h,updatingQuantityItemId:H,bulkMovingToList:P=!1,bulkCopyingToList:d=!1,onSelectAll:y,onSelectNone:f,onBulkAddToCart:I,onBulkDelete:M,onBulkMoveToList:O,onBulkCopyToList:N})=>{const T=ct({statusDeleting:"RequisitionList.RequisitionListView.statusDeleting",actionSelectAll:"RequisitionList.RequisitionListView.actionSelectAll",actionSelectNone:"RequisitionList.RequisitionListView.actionSelectNone",actionAddToCart:"RequisitionList.RequisitionListView.actionAddToCart",actionDeleteSelectedItems:"RequisitionList.RequisitionListView.actionDeleteSelectedItems",statusBulkAddingToCart:"RequisitionList.RequisitionListView.statusBulkAddingToCart",actionMoveToList:"RequisitionList.RequisitionListView.actionMoveToList",actionCopyToList:"RequisitionList.RequisitionListView.actionCopyToList"}),w=b!==null||k!==null||h||H!==null||P||d,p=r.size>0;return E("div",{className:"requisition-list-view__batch-actions",children:[E("div",{className:"requisition-list-view__batch-actions-left",children:[i("button",{"data-testid":"bulk-actions-select-toggle-btn",type:"button",className:`requisition-list-view__batch-actions-select-toggle ${p?"requisition-list-view__batch-actions-select-toggle--active":""}`,onClick:p?f:y,disabled:w,"aria-label":p?T.actionSelectNone:T.actionSelectAll,children:i(lt,{source:re})}),i("button",{type:"button",className:"requisition-list-view__batch-actions-select-label",onClick:p?f:y,disabled:w,children:T.actionSelectAll})]}),p&&E("div",{className:"requisition-list-view__batch-actions-buttons",children:[i("span",{className:"requisition-list-view__batch-actions-count-badge","aria-label":`${r.size} items selected`,children:r.size}),O&&i(at,{"data-testid":"bulk-actions-move-to-list-btn",type:"button",variant:"secondary",onClick:O,disabled:w,children:T.actionMoveToList}),N&&i(at,{"data-testid":"bulk-actions-copy-to-list-btn",type:"button",variant:"secondary",onClick:N,disabled:w,children:T.actionCopyToList}),i(at,{"data-testid":"bulk-actions-add-to-cart-btn",type:"button",variant:"secondary",onClick:I,disabled:w,children:h?T.statusBulkAddingToCart:T.actionAddToCart}),i("button",{"data-testid":"bulk-actions-delete-btn",type:"button",className:"requisition-list-view__batch-actions-delete-icon",onClick:M,disabled:w,"aria-label":T.actionDeleteSelectedItems,children:i(lt,{source:It})})]})]})},Fe=({requisitionListUid:r,skipProductLoading:b=!1,pageSize:k=ft,routeRequisitionListGrid:h,fallbackRoute:H="/customer/account",getProductData:P,enrichConfigurableProducts:d,currentCustomerEmail:y,routeSharedRequisitionList:f})=>{var _t,bt,qt,vt,kt,Tt;const[I,M]=S(!1),[O,N]=S(null),[T,w]=S(null),[p,B]=S(!1),[z,J]=S(null),[Y,$]=S(!1),[m,x]=S(k),[it,W]=S(!0),[Q,G]=S(!1),[U,ot]=S([]),{currentRequisitionList:e,setCurrentRequisitionList:g,selectedItems:j,setSelectedItems:t,handleItemSelection:n,handleSelectAll:l,handleSelectNone:L}=ce(),a=ct({emptyRequisitionList:"RequisitionList.RequisitionListView.emptyRequisitionList",errorLoadingProducts:"RequisitionList.RequisitionListView.errorLoadingProducts",errorLoadPage:"RequisitionList.RequisitionListView.errorLoadPage",notFoundTitle:"RequisitionList.RequisitionListView.notFoundTitle",notFoundMessage:"RequisitionList.RequisitionListView.notFoundMessage",notFoundActionLabel:"RequisitionList.RequisitionListView.notFoundActionLabel",notEnabledTitle:"RequisitionList.RequisitionListsNotEnabled.title",notEnabledMessage:"RequisitionList.RequisitionListsNotEnabled.message",partialMoveSuccess:"RequisitionList.RequisitionListAlert.partialMoveSuccess",notEnabledActionLabel:"RequisitionList.RequisitionListsNotEnabled.actionLabel",show:"RequisitionList.PageSizePicker.show",deleteItemsTitle:"RequisitionList.RequisitionListView.deleteItemsTitle",deleteItemsMessage:"RequisitionList.RequisitionListView.deleteItemsMessage",confirmAction:"RequisitionList.RequisitionListView.confirmAction",cancelAction:"RequisitionList.RequisitionListView.cancelAction",moveToListTitle:"RequisitionList.RequisitionListView.moveToListTitle",moveToListConfirm:"RequisitionList.RequisitionListView.moveToListConfirm",copyToListTitle:"RequisitionList.RequisitionListView.copyToListTitle",copyToListConfirm:"RequisitionList.RequisitionListView.copyToListConfirm"}),{alert:A,setAlert:F,handleRequisitionListAlert:u}=te(),{isEnabled:Z}=ee();dt(()=>{const o=et.on("requisitionList/data",c=>{(c==null?void 0:c.uid)===r&&(g(c),t(new Set))}),s=et.on("requisitionList/alert",u);return()=>{o==null||o.off(),s==null||s.off()}},[u,r,g,t]),dt(()=>{x(k)},[k]);const K=D(async o=>{var c;if(!((c=o.items)!=null&&c.length)||typeof d!="function")return o;const s=await d(o.items);return{...o,items:s}},[d]),_=D(async o=>{var c,q;const s=((c=o.items)==null?void 0:c.map(V=>V.sku))||[];if(s.length===0)return o;M(!0);try{const V=await P(s);if(V){const C=new Map;return V.forEach(R=>{C.set(R.sku,R)}),{...o,items:(q=o.items)==null?void 0:q.map(R=>{const X=C.get(R.sku);return{...R,product:X||R.product,stock_status:R.stock_status||(X==null?void 0:X.stock_status)||"IN_STOCK",only_x_left_in_stock:R.only_x_left_in_stock??(X==null?void 0:X.only_x_left_in_stock)??null}})}}return console.warn("No products found"),o}catch(V){return console.warn(V instanceof Error?V.message:a.errorLoadingProducts),o}finally{M(!1)}},[P,a.errorLoadingProducts]),{showMoveToListModal:st,setShowMoveToListModal:nt,movingToList:tt,showCopyToListModal:gt,setShowCopyToListModal:ht,copyingToList:ut,handleMoveToList:Nt,handleCopyToList:xt}=ue({sourceListUid:e==null?void 0:e.uid,selectedItems:j,currentPageSize:m,currentPage:((_t=e==null?void 0:e.page_info)==null?void 0:_t.current_page)||1,enrichConfigurableProductsInList:K,fetchAndMergeProducts:_,setCurrentRequisitionList:g,setSelectedItems:t,handleRequisitionListAlert:u});dt(()=>{if(!r){W(!1);return}if(e&&e.uid===r)return;W(!0),(async()=>{try{const s=await pt(r,1,k,d);if(s)if(b)g(s);else{const c=await _(s);g(c)}}catch(s){console.error("Failed to initialize requisition list from UID:",s)}finally{W(!1)}})()},[r,k,b,d,_,g,e]);const yt=D(async o=>{o&&o.length===1?(B(!1),w(o)):(B(!0),w(null));try{const s=await $t(e.uid,o),c=(o==null?void 0:o.length)||0,q=(s==null?void 0:s.length)||0,V=c-q;if(s&&s.length>0&&V>0){const C={action:"move",type:"error",context:"product",message:[a.partialMoveSuccess.replace("{successCount}",String(V)).replace("{failedCount}",String(q))]};u(C),et.emit("requisitionList/alert",C)}else if(s&&s.length>0){const C={action:"move",type:"error",context:"product",message:s.map(R=>R.message)};u(C),et.emit("requisitionList/alert",C)}else{const C={action:"move",type:"success",context:"product"};u(C),et.emit("requisitionList/alert",C)}}catch{const c={action:"move",type:"error",context:"product"};u(c),et.emit("requisitionList/alert",c)}finally{w(null),B(!1)}},[e,u,a]),wt=D(o=>{o&&o.length>0&&(ot(o),G(!0))},[]),At=D(async()=>{U.length===1?N(U[0]):N("bulk");try{const o=await Ft(e.uid,U,m,e.page_info.current_page,d);if(o){const s=await K(o),c=await _(s);g(c),u({action:"delete",type:"success",context:"product"})}else u({action:"delete",type:"error",context:"product"})}catch{u({action:"delete",type:"error",context:"product"})}finally{N(null),G(!1),ot([])}},[U,e,m,_,g,u,K,d]),Vt=D(async(o,s)=>{var c;J(o);try{const q=(c=e.items)==null?void 0:c.find(R=>R.uid===o);if(!q){u({action:"update",type:"error",context:"product"});return}const V={item_id:q.uid,entered_options:q.entered_options,selected_options:q.selected_options,quantity:s},C=await Ot(e.uid,[V],m,e.page_info.current_page,d);if(C){const R=await _(C);g(R),u({action:"update",type:"success",context:"product"})}else u({action:"update",type:"error",context:"product"})}catch{u({action:"update",type:"error",context:"product"})}finally{J(null)}},[e,_,g,m,u,d]),Et=D(async o=>{$(!0);try{const s=await pt(e.uid,o,m,d);if(s){const q=await _(s);g(q)}else console.warn(a.errorLoadPage)}catch(s){console.warn(s instanceof Error?s.message:a.errorLoadPage)}finally{$(!1)}},[e==null?void 0:e.uid,_,m,a,g,d]),Pt=D(async o=>{x(o),$(!0);try{const s=await pt(e.uid,1,o,d);if(s){const q=await _(s);g(q)}else console.warn(a.errorLoadPage)}catch{console.warn(a.errorLoadPage)}finally{$(!1)}},[e==null?void 0:e.uid,_,a,g,d]),Dt=D(async o=>{const s=await K(o),c=await _(s);g(c)},[g,_,K]);return E("div",{className:"requisition-list-view__container",children:[it?i("div",{className:"requisition-list-view__loading",children:i(Kt,{stroke:"4",size:"large"})}):Z===!1?i(St,{title:a.notEnabledTitle,message:a.notEnabledMessage,actionLabel:a.notEnabledActionLabel,onAction:()=>{window.location.href=H}}):!e||!Ut(e==null?void 0:e.uid)?i(St,{title:a.notFoundTitle,message:a.notFoundMessage,actionLabel:h?a.notFoundActionLabel:void 0,onAction:h?()=>{const o=h();o&&typeof o=="string"&&(window.location.href=o)}:void 0}):E(Ct,{children:[i(se,{requisitionList:e,routeRequisitionListGrid:h,onUpdate:Dt,onAlert:u,enrichConfigurableProducts:d,currentCustomerEmail:y,routeSharedRequisitionList:f}),A&&i("div",{className:"requisition-list__alert-wrapper",children:i(Xt,{heading:A.description,type:A.type,variant:"primary",onDismiss:()=>F(null)})}),e.items_count===0?i(ie,{textContent:`${e.name} ${a.emptyRequisitionList}`}):E(Ct,{children:[i(pe,{selectedItems:j,deletingItemId:O,addingToCartItemId:T,bulkAddingToCart:p,updatingQuantityItemId:z,bulkMovingToList:tt,bulkCopyingToList:ut,onSelectAll:l,onSelectNone:L,onBulkAddToCart:()=>yt(Array.from(j)),onBulkDelete:()=>wt(Array.from(j)),onBulkMoveToList:()=>nt(!0),onBulkCopyToList:()=>ht(!0)}),i(de,{items:e.items,selectedItems:j,currentPage:((bt=e.page_info)==null?void 0:bt.current_page)||1,pageSize:((qt=e.page_info)==null?void 0:qt.page_size)||ft,handleItemSelection:n,handleUpdateQuantity:Vt,onAddToCart:yt,onDeleteItem:wt}),e.page_info&&E("div",{className:"requisition-list-view__pagination",children:[i(ne,{pageInfo:e.page_info,totalCount:e.items_count}),(((vt=e.page_info)==null?void 0:vt.total_pages)||0)>1&&i(Jt,{totalPages:e.page_info.total_pages,currentPage:((kt=e.page_info)==null?void 0:kt.current_page)||1,onChange:Et,disabled:Y||I}),E("div",{className:"requisition-list-view__pagination-picker",children:[i("span",{children:a.show}),i(ae,{currentPageSize:((Tt=e.page_info)==null?void 0:Tt.page_size)||ft,onPageSizeChange:Pt,disabled:Y||I})]})]})]})]}),Q&&i(Lt,{isOpen:Q,isLoading:O!==null,title:a.deleteItemsTitle,modalContent:a.deleteItemsMessage,confirmBtnCaption:a.confirmAction,closeBtnCaption:a.cancelAction,handleModalOnClose:()=>{G(!1),ot([])},handleModalOnConfirm:At}),st&&i(Lt,{isOpen:st,isLoading:tt,title:a.moveToListTitle,modalContent:i(Mt,{excludeUid:e==null?void 0:e.uid,confirmLabel:a.moveToListConfirm,disabled:tt,onConfirm:Nt}),handleModalOnClose:()=>nt(!1)}),gt&&i(Lt,{isOpen:gt,isLoading:ut,title:a.copyToListTitle,modalContent:i(Mt,{excludeUid:e==null?void 0:e.uid,confirmLabel:a.copyToListConfirm,disabled:ut,onConfirm:xt}),handleModalOnClose:()=>ht(!1)})]})};export{Fe as RequisitionListView,Fe as default};
+import{d as l,d as n}from"./SharedRequisitionList.js";import"@dropins/tools/lib.js";import"@dropins/tools/i18n.js";import"@dropins/tools/preact-jsx-runtime.js";import"@dropins/tools/components.js";import"@dropins/tools/preact-compat.js";import"@dropins/tools/event-bus.js";import"../api.js";import"@dropins/tools/fetch-graphql.js";import"@dropins/tools/preact-hooks.js";import"@dropins/tools/preact.js";export{l as RequisitionListView,n as default};
 //# sourceMappingURL=RequisitionListView.js.map
