+++ scripts/__dropins__/storefront-requisition-list/containers/ShareRequisitionListContent.js
@@ -1,4 +1,4 @@
 /*! Copyright 2026 Adobe
 All Rights Reserved. */
-import{S as u,S as d}from"../chunks/ShareRequisitionListContent.js";import"@dropins/tools/preact-jsx-runtime.js";import"@dropins/tools/preact-hooks.js";import"../chunks/shareRequisitionListByToken.js";import"@dropins/tools/event-bus.js";import"../chunks/fetch-graphql.js";import"@dropins/tools/fetch-graphql.js";import"@dropins/tools/components.js";/* empty css                                    */import"@dropins/tools/i18n.js";import"../chunks/state.js";export{u as ShareRequisitionListContent,d as default};
+import{S as u,S as d}from"./SharedRequisitionList.js";import"@dropins/tools/lib.js";import"@dropins/tools/i18n.js";import"@dropins/tools/preact-jsx-runtime.js";import"@dropins/tools/components.js";import"@dropins/tools/preact-compat.js";import"@dropins/tools/event-bus.js";import"../api.js";import"@dropins/tools/fetch-graphql.js";import"@dropins/tools/preact-hooks.js";import"@dropins/tools/preact.js";export{u as ShareRequisitionListContent,d as default};
 //# sourceMappingURL=ShareRequisitionListContent.js.map
