+++ scripts/__dropins__/storefront-requisition-list/containers/SharedRequisitionList.js
@@ -1,4 +1,4 @@
 /*! Copyright 2026 Adobe
 All Rights Reserved. */
-import{jsxs as d,jsx as e}from"@dropins/tools/preact-jsx-runtime.js";import{useState as _,useRef as R,useEffect as g,useCallback as f}from"@dropins/tools/preact-compat.js";import{events as w}from"@dropins/tools/event-bus.js";import{g as N,i as S}from"../chunks/importSharedRequisitionList.js";import{ProgressSpinner as y,InLineAlert as q,Header as I,Table as H,Button as k}from"@dropins/tools/components.js";/* empty css                             */import{useText as B}from"@dropins/tools/i18n.js";import"../chunks/fetch-error.js";import"../chunks/transform-requisition-list.js";import"../chunks/fetch-graphql.js";import"@dropins/tools/fetch-graphql.js";const T=r=>{var o,n;return(o=r.configurable_options)!=null&&o.length?r.configurable_options.map(s=>`${s.option_label}: ${s.value_label}`).join(", "):(n=r.bundle_options)!=null&&n.length?r.bundle_options.map(s=>s.label).join(", "):""},x=({status:r,previewData:o,errorMessage:n,onImport:s,translations:t})=>{if(r==="preview_loading")return d("div",{className:"shared-requisition-list__loading",children:[e(y,{}),e("span",{children:t.loading})]});if(r==="preview_error")return e("div",{className:"shared-requisition-list__container",children:e(q,{heading:n,type:"error",variant:"primary"})});if(!o)return null;const c=o.requisitionList.name,m=o.requisitionList.items??[],a=r==="importing",l=r==="import_success",u=[{label:t.skuHeader,key:"sku"},{label:t.qtyHeader,key:"qty"},{label:t.optionsHeader,key:"options"}],h=m.map(p=>({sku:p.sku,qty:p.quantity,options:T(p)}));return d("div",{className:"shared-requisition-list__preview",children:[e(I,{title:t.previewTitle,"aria-label":t.previewTitle}),r==="import_success"&&e("div",{className:"shared-requisition-list__alert-wrapper",children:e(q,{heading:t.successImport.replace("{listName}",c),type:"success",variant:"primary"})}),r==="import_error"&&e("div",{className:"shared-requisition-list__alert-wrapper",children:e(q,{heading:n,type:"error",variant:"primary"})}),d("div",{className:"shared-requisition-list__preview-details",children:[d("div",{className:"shared-requisition-list__preview-row",children:[e("span",{className:"shared-requisition-list__preview-label",children:t.senderLabel}),e("span",{className:"shared-requisition-list__preview-value",children:o.senderName})]}),d("div",{className:"shared-requisition-list__preview-row",children:[e("span",{className:"shared-requisition-list__preview-label",children:t.listNameLabel}),e("span",{className:"shared-requisition-list__preview-value",children:c})]}),o.requisitionList.description&&d("div",{className:"shared-requisition-list__preview-row",children:[e("span",{className:"shared-requisition-list__preview-label",children:t.descriptionLabel}),e("span",{className:"shared-requisition-list__preview-value",children:o.requisitionList.description})]})]}),m.length>0&&e("div",{className:"shared-requisition-list__table-wrapper",children:e(H,{columns:u,rowData:h,"data-testid":"shared-list-items-table"})}),e("div",{className:"shared-requisition-list__actions",children:e(k,{type:"button",variant:"primary",onClick:s,disabled:a||l,"data-testid":"import-shared-list-btn",children:a?t.importingButton:t.importButton})})]})},J=({token:r,routeRequisitionList:o})=>{const[n,s]=_("preview_loading"),[t,c]=_(null),[m,a]=_(""),l=R(!0),u=B({loading:"RequisitionList.SharedRequisitionList.loading",previewTitle:"RequisitionList.SharedRequisitionList.previewTitle",senderLabel:"RequisitionList.SharedRequisitionList.senderLabel",listNameLabel:"RequisitionList.SharedRequisitionList.listNameLabel",descriptionLabel:"RequisitionList.SharedRequisitionList.descriptionLabel",itemsCountLabel:"RequisitionList.SharedRequisitionList.itemsCountLabel",importButton:"RequisitionList.SharedRequisitionList.importButton",importingButton:"RequisitionList.SharedRequisitionList.importingButton",errorPreview:"RequisitionList.SharedRequisitionList.errorPreview",successImport:"RequisitionList.RequisitionListAlert.successImport",errorImport:"RequisitionList.RequisitionListAlert.errorImport",skuHeader:"RequisitionList.SharedRequisitionList.skuHeader",qtyHeader:"RequisitionList.SharedRequisitionList.qtyHeader",optionsHeader:"RequisitionList.SharedRequisitionList.optionsHeader"});g(()=>()=>{l.current=!1},[]),g(()=>{N(r).then(i=>{if(l.current){if(!i){a(""),s("preview_error");return}c(i),s("preview_loaded")}}).catch(i=>{l.current&&(a(i instanceof Error&&i.message?i.message:""),s("preview_error"))})},[r]);const h=f(()=>{s("importing"),a(""),S(r).then(({requisitionList:i,userErrors:v})=>{if(!l.current)return;if(v.length>0){a(v[0].message),s("import_error");return}const L=(i==null?void 0:i.name)??"",b=(i==null?void 0:i.uid)??"";if(w.emit("requisitionList/alert",{action:"import",type:"success",context:"requisitionList",listName:L}),o){o(b,L);return}s("import_success")}).catch(i=>{l.current&&(a(i instanceof Error&&i.message?i.message:""),s("import_error"))})},[r,o]),p=m||(n==="import_error"?u.errorImport:u.errorPreview);return e(x,{status:n,previewData:t,errorMessage:p,onImport:h,translations:u})};export{J as SharedRequisitionList,J as default};
+import{classes as re,Slot as Li,VComponent as Ge}from"@dropins/tools/lib.js";import{useText as j}from"@dropins/tools/i18n.js";import{jsxs as R,jsx as e,Fragment as de}from"@dropins/tools/preact-jsx-runtime.js";import{ProgressSpinner as fe,InLineAlert as he,Header as Ye,Table as Me,Button as Q,Field as ye,MultiSelect as qi,Divider as fi,Input as Ae,TextArea as gi,Icon as le,Card as be,Modal as Ri,Pagination as ei,Price as je,Image as _i,Checkbox as yi,Picker as bi}from"@dropins/tools/components.js";import*as A from"@dropins/tools/preact-compat.js";import{useState as I,useRef as vi,useEffect as ne,useCallback as V,useMemo as _e}from"@dropins/tools/preact-compat.js";import{events as U}from"@dropins/tools/event-bus.js";import{getSharedRequisitionList as wi,importSharedRequisitionList as Ci,s as J,getCompanyUsers as ki,shareRequisitionListByToken as Si,updateRequisitionList as Ee,c as Ti,getRequisitionLists as Se,deleteRequisitionList as ii,a as Ze,b as Ce,u as Ni,moveItemsBetweenRequisitionLists as Mi,copyItemsBetweenRequisitionLists as Ai,d as Ei,g as xi,shareRequisitionListByEmail as Ii,addRequisitionListItemsToCart as Di,deleteRequisitionListItems as Pi,updateRequisitionListItems as Vi,getRequisitionList as ke,i as Oi}from"../api.js";import{useState as z,useEffect as Xe,useCallback as ae}from"@dropins/tools/preact-hooks.js";import{h as Bi}from"@dropins/tools/preact.js";import"@dropins/tools/fetch-graphql.js";const Fi=t=>{var n,r;return(n=t.configurable_options)!=null&&n.length?t.configurable_options.map(i=>`${i.option_label}: ${i.value_label}`).join(", "):(r=t.bundle_options)!=null&&r.length?t.bundle_options.map(i=>i.label).join(", "):""},zi=({status:t,previewData:n,errorMessage:r,onImport:i,translations:s})=>{if(t==="preview_loading")return R("div",{className:"shared-requisition-list__loading",children:[e(fe,{}),e("span",{children:s.loading})]});if(t==="preview_error")return e("div",{className:"shared-requisition-list__container",children:e(he,{heading:r,type:"error",variant:"primary"})});if(!n)return null;const u=n.requisitionList.name,l=n.requisitionList.items??[],a=t==="importing",c=t==="import_success",q=[{label:s.skuHeader,key:"sku"},{label:s.qtyHeader,key:"qty"},{label:s.optionsHeader,key:"options"}],T=l.map(k=>({sku:k.sku,qty:k.quantity,options:Fi(k)}));return R("div",{className:"shared-requisition-list__preview",children:[e(Ye,{title:s.previewTitle,"aria-label":s.previewTitle}),t==="import_success"&&e("div",{className:"shared-requisition-list__alert-wrapper",children:e(he,{heading:s.successImport.replace("{listName}",u),type:"success",variant:"primary"})}),t==="import_error"&&e("div",{className:"shared-requisition-list__alert-wrapper",children:e(he,{heading:r,type:"error",variant:"primary"})}),R("div",{className:"shared-requisition-list__preview-details",children:[R("div",{className:"shared-requisition-list__preview-row",children:[e("span",{className:"shared-requisition-list__preview-label",children:s.senderLabel}),e("span",{className:"shared-requisition-list__preview-value",children:n.senderName})]}),R("div",{className:"shared-requisition-list__preview-row",children:[e("span",{className:"shared-requisition-list__preview-label",children:s.listNameLabel}),e("span",{className:"shared-requisition-list__preview-value",children:u})]}),n.requisitionList.description&&R("div",{className:"shared-requisition-list__preview-row",children:[e("span",{className:"shared-requisition-list__preview-label",children:s.descriptionLabel}),e("span",{className:"shared-requisition-list__preview-value",children:n.requisitionList.description})]})]}),l.length>0&&e("div",{className:"shared-requisition-list__table-wrapper",children:e(Me,{columns:q,rowData:T,"data-testid":"shared-list-items-table"})}),e("div",{className:"shared-requisition-list__actions",children:e(Q,{type:"button",variant:"primary",onClick:i,disabled:a||c,"data-testid":"import-shared-list-btn",children:a?s.importingButton:s.importButton})})]})},gt=({token:t,routeRequisitionList:n})=>{const[r,i]=I("preview_loading"),[s,u]=I(null),[l,a]=I(""),c=vi(!0),q=j({loading:"RequisitionList.SharedRequisitionList.loading",previewTitle:"RequisitionList.SharedRequisitionList.previewTitle",senderLabel:"RequisitionList.SharedRequisitionList.senderLabel",listNameLabel:"RequisitionList.SharedRequisitionList.listNameLabel",descriptionLabel:"RequisitionList.SharedRequisitionList.descriptionLabel",itemsCountLabel:"RequisitionList.SharedRequisitionList.itemsCountLabel",importButton:"RequisitionList.SharedRequisitionList.importButton",importingButton:"RequisitionList.SharedRequisitionList.importingButton",errorPreview:"RequisitionList.SharedRequisitionList.errorPreview",successImport:"RequisitionList.RequisitionListAlert.successImport",errorImport:"RequisitionList.RequisitionListAlert.errorImport",skuHeader:"RequisitionList.SharedRequisitionList.skuHeader",qtyHeader:"RequisitionList.SharedRequisitionList.qtyHeader",optionsHeader:"RequisitionList.SharedRequisitionList.optionsHeader"});ne(()=>()=>{c.current=!1},[]),ne(()=>{wi(t).then(p=>{if(c.current){if(!p){a(""),i("preview_error");return}u(p),i("preview_loaded")}}).catch(p=>{c.current&&(a(p instanceof Error&&p.message?p.message:""),i("preview_error"))})},[t]);const T=V(()=>{i("importing"),a(""),Ci(t).then(({requisitionList:p,userErrors:h})=>{if(!c.current)return;if(h.length>0){a(h[0].message),i("import_error");return}const y=(p==null?void 0:p.name)??"",b=(p==null?void 0:p.uid)??"";if(U.emit("requisitionList/alert",{action:"import",type:"success",context:"requisitionList",listName:y}),n){n(b,y);return}i("import_success")}).catch(p=>{c.current&&(a(p instanceof Error&&p.message?p.message:""),i("import_error"))})},[t,n]),k=l||(r==="import_error"?q.errorImport:q.errorPreview);return e(zi,{status:r,previewData:s,errorMessage:k,onImport:T,translations:q})},Hi=({loadingUsers:t,usersErrorMessage:n,loadingLink:r,selectedUserValues:i,multiSelectOptions:s,shareLink:u,linkErrorMessage:l,linkCopied:a,isSubmitting:c,canSubmit:q,onSubmitClick:T,onCopyLinkClick:k,onSelectedUsersChange:p,onUsersFieldInteract:h,selectionError:y,submitErrorMessage:b,isShareSuccess:D,sharedRecipientEmails:_})=>{const f=j({emailInstruction:"RequisitionList.ShareRequisitionListContent.emailInstruction",emailLabel:"RequisitionList.ShareRequisitionListContent.emailLabel",emailPlaceholder:"RequisitionList.ShareRequisitionListContent.emailPlaceholder",submitLabel:"RequisitionList.ShareRequisitionListContent.submitLabel",linkInstruction:"RequisitionList.ShareRequisitionListContent.linkInstruction",copyLink:"RequisitionList.ShareRequisitionListContent.copyLink",linkCopied:"RequisitionList.ShareRequisitionListContent.linkCopied",loadingUsers:"RequisitionList.ShareRequisitionListContent.loadingUsers",loadingLink:"RequisitionList.ShareRequisitionListContent.loadingLink",noUsersAvailable:"RequisitionList.ShareRequisitionListContent.noUsersAvailable",maxRecipientsValidation:"RequisitionList.ShareRequisitionListContent.maxRecipientsValidation",shareSuccessMessage:"RequisitionList.ShareRequisitionListContent.shareSuccessMessage"});return R("div",{className:"share-requisition-list-content",children:[D?R("div",{className:"share-requisition-list-content__success",children:[e("p",{className:"share-requisition-list-content__instruction",children:f.shareSuccessMessage}),e("div",{className:"share-requisition-list-content__recipient-list",children:_.map(E=>e("p",{className:"share-requisition-list-content__recipient",children:E},E))})]}):R(de,{children:[e("p",{className:"share-requisition-list-content__instruction",children:f.emailInstruction}),e("div",{className:"share-requisition-list-content__field",children:t?R("div",{className:"share-requisition-list-content__loading",children:[e(fe,{size:"small",stroke:"3"}),e("span",{children:f.loadingUsers})]}):n?e("p",{className:"dropin-field__hint dropin-field__hint--medium dropin-field__hint--error",children:n}):e(ye,{label:f.emailLabel,error:y??void 0,disabled:c,onMouseDown:h,onKeyDown:h,children:e(qi,{options:s,value:i,onChange:p,placeholder:f.emailPlaceholder,noResultsText:f.noUsersAvailable,disabled:c,error:!!y,className:"share-requisition-list-content__multi-select"})})}),e("div",{className:"share-requisition-list-content__actions",children:e(Q,{variant:"primary",onClick:T,disabled:!q,type:"button","data-testid":"share-submit-btn",children:f.submitLabel})}),b&&e("p",{className:"dropin-field__hint dropin-field__hint--medium dropin-field__hint--error",children:b})]}),e(fi,{variant:"secondary",className:"share-requisition-list-content__divider-secondary"}),e("p",{className:"share-requisition-list-content__instruction",children:f.linkInstruction}),r?R("div",{className:"share-requisition-list-content__loading",children:[e(fe,{size:"small",stroke:"3"}),e("span",{children:f.loadingLink})]}):u?R(de,{children:[e("div",{className:"share-requisition-list-content__link-row",children:e(Ae,{readOnly:!0,value:u,"data-testid":"share-link-field"})}),e("div",{className:"share-requisition-list-content__actions",children:e(Q,{variant:"secondary",onClick:k,type:"button","data-testid":"copy-link-btn",children:a?f.linkCopied:f.copyLink})})]}):l?e("p",{className:"dropin-field__hint dropin-field__hint--medium dropin-field__hint--error",children:l}):null]})},$i=({requisitionListUid:t,isSubmitting:n,onSubmit:r,currentCustomerEmail:i,routeSharedRequisitionList:s})=>{const u=j({emailInstruction:"RequisitionList.ShareRequisitionListContent.emailInstruction",emailLabel:"RequisitionList.ShareRequisitionListContent.emailLabel",emailPlaceholder:"RequisitionList.ShareRequisitionListContent.emailPlaceholder",submitLabel:"RequisitionList.ShareRequisitionListContent.submitLabel",linkInstruction:"RequisitionList.ShareRequisitionListContent.linkInstruction",copyLink:"RequisitionList.ShareRequisitionListContent.copyLink",linkCopied:"RequisitionList.ShareRequisitionListContent.linkCopied",loadingUsers:"RequisitionList.ShareRequisitionListContent.loadingUsers",loadingLink:"RequisitionList.ShareRequisitionListContent.loadingLink",noUsersAvailable:"RequisitionList.ShareRequisitionListContent.noUsersAvailable",usersLoadError:"RequisitionList.ShareRequisitionListContent.usersLoadError",maxRecipientsValidation:"RequisitionList.ShareRequisitionListContent.maxRecipientsValidation",shareSuccessMessage:"RequisitionList.ShareRequisitionListContent.shareSuccessMessage"}),[l,a]=z([]),[c,q]=z(!0),[T,k]=z(!1),[p,h]=z(new Set),[y,b]=z(null),[D,_]=z(!0),[f,E]=z(null),[g,m]=z(!1),[w,H]=z(null),[Z,X]=z(null),[Y,ee]=z(!1),[$,d]=z([]),N=(()=>{var W;const L=(W=J.config)==null?void 0:W.requisition_list_share_max_recipients,P=Number(L);return Number.isFinite(P)&&P>0?P:null})();Xe(()=>{ki().then(L=>{const P=L.filter(W=>!(i&&W.email.toLowerCase()===i.toLowerCase()));a(P)}).catch(()=>{k(!0)}).finally(()=>q(!1))},[i]),Xe(()=>{Si(t).then(L=>{var P;if(L.token){const x=`/${((P=J.config)==null?void 0:P.requisition_list_share_storefront_path)??""}?requisition_id=${L.token}`,se=s?s(x):x;b(se)}else b(null);E(L.errorMessage)}).finally(()=>_(!1))},[t,s]);const B=ae(async()=>{var x;const L=Array.from(p),P=l.filter(se=>L.includes(String(se.id))).map(se=>se.email),W=await r(L);W?X(((x=W[0])==null?void 0:x.message)||null):(ee(!0),d(P),X(null))},[p,l,r]),o=ae(()=>{navigator.clipboard.writeText(y).then(()=>{m(!0),setTimeout(()=>m(!1),3e3)},L=>{console.error("Failed to copy share link to clipboard:",L)})},[y]),C=l.map(L=>({label:`${L.firstname} ${L.lastname} (${L.email})`,value:L.id})),M=T?u.usersLoadError:null,F=!n&&p.size>0&&!w;return e(Hi,{loadingUsers:c,usersErrorMessage:M,loadingLink:D,selectedUserValues:Array.from(p),multiSelectOptions:C,shareLink:y,linkErrorMessage:f,linkCopied:g,isSubmitting:n,canSubmit:F,onSubmitClick:B,onCopyLinkClick:o,onUsersFieldInteract:()=>{w&&H(null),Z&&X(null)},onSelectedUsersChange:L=>{const P=L.map(String);if(N&&P.length>N){H(u.maxRecipientsValidation.replace("{max}",String(N)));return}H(null),h(new Set(P))},selectionError:w,submitErrorMessage:Z,isShareSuccess:Y,sharedRecipientEmails:$})},qe=10,Ke=3,Wi=40,Ui=255,Qi=/^[a-zA-Z0-9\s.,\-_!?'"()&]+$/,xe=({className:t,mode:n,defaultValues:r={name:"",description:""},error:i=null,onSubmit:s,onCancel:u,...l})=>{const[a,c]=z(r),[q,T]=z({name:!1}),[k,p]=z(!1),h=j({actionCancel:"RequisitionList.RequisitionListForm.actionCancel",actionSave:"RequisitionList.RequisitionListForm.actionSave",requiredField:"RequisitionList.RequisitionListForm.requiredField",nameMinLength:"RequisitionList.RequisitionListForm.nameMinLength",nameInvalidCharacters:"RequisitionList.RequisitionListForm.nameInvalidCharacters",floatingLabel:"RequisitionList.RequisitionListForm.floatingLabel",placeholder:"RequisitionList.RequisitionListForm.placeholder",label:"RequisitionList.RequisitionListForm.label",updateTitle:"RequisitionList.RequisitionListForm.updateTitle",createTitle:"RequisitionList.RequisitionListForm.createTitle"}),y=g=>{const m=g.trim();return m?m.length<Ke?h.nameMinLength.replace("{min}",Ke.toString()):Qi.test(m)?"":h.nameInvalidCharacters:h.requiredField},b=g=>m=>{const w=m.target;c(H=>({...H,[g]:w.value}))},D=g=>()=>{T(m=>({...m,[g]:!0}))},_=async g=>{var w;if(g.preventDefault(),T({name:!0,description:!0}),!(y(a.name)||k)){p(!0);try{await s({name:a.name.trim(),description:((w=a.description)==null?void 0:w.trim())??""})}catch{p(!1)}}},f=q.name?y(a.name):"",E=n==="update"?h.updateTitle:h.createTitle;return R("div",{...l,className:re(["requisition-list-form",t]),children:[R("div",{className:"requisition-list-form__title",children:[E,k?e("div",{className:re(["requisition-list-form_progress-spinner",t]),"data-testid":"requisition-list-form-progress-spinner",children:e(fe,{stroke:"4",size:"small"})}):null]}),i?e(he,{type:"error",className:"requisition-list-form__notification",variant:"secondary",heading:i,"data-testid":"requisition-list-alert"}):null,R("form",{className:re(["requisition-list-form__form",t]),onSubmit:_,children:[e(ye,{error:f,disabled:k,children:e(Ae,{id:"requisition-list-form-name",name:"name",type:"text",floatingLabel:h.floatingLabel,placeholder:h.placeholder,maxLength:Wi,value:a.name,onChange:b("name"),onBlur:D("name")})}),e(ye,{disabled:k,children:e(gi,{id:"requisition-list-form-description",name:"description",label:h.label,placeholder:h.label,maxLength:Ui,value:a.description,onChange:b("description"),onBlur:D("description")})}),R("div",{className:"requisition-list-form__actions",children:[e(Q,{type:"button",variant:"secondary",onClick:u,disabled:k,"data-testid":"requisition-list-form-cancel",children:h.actionCancel}),e(Q,{type:"submit",disabled:k,"data-testid":"requisition-list-form-save",children:h.actionSave})]})]})]})};function ji(t,n,r,i){const[s,u]=z(null);return{error:s,submit:async a=>{u(null);try{const c=a.description??"",q=t==="update"&&n?await Ee(n,a.name,c):await Ti(a.name,c);return q&&(r==null||r(q)),q}catch(c){const q=(c==null?void 0:c.message)||"Unexpected error";return u(q),i==null||i(q),null}}}}const ti=({mode:t,requisitionListUid:n,defaultValues:r={name:"",description:""},onSuccess:i,onError:s,onCancel:u})=>{const{error:l,submit:a}=ji(t,n,i,s);return e(xe,{mode:t,defaultValues:r,error:l,onSubmit:async q=>{await a(q)},onCancel:u})};function Zi(t,n,r){const i=j({actionRename:"RequisitionList.RequisitionListView.actionRename",actionDeleteList:"RequisitionList.RequisitionListView.actionDeleteList"}),[s,u]=I(null),[l,a]=I(!1),[c,q]=I(!1),T=V(()=>{r&&r(),a(!0)},[r]),k=V(()=>a(!1),[]),p=_e(()=>{if(t)return{handleOpenRenameModal:_=>{a(f=>f&&!1),t.handleOpenRenameModal(_)},handleOpenDeleteModal:_=>{a(f=>f&&!1),t.handleOpenDeleteModal(_)}}},[t]),h=V(async(_,f)=>{var E,g,m;q(!0);try{const w=await Se(_,f),H=((E=w==null?void 0:w.page_info)==null?void 0:E.current_page)??1,Z=((g=w==null?void 0:w.page_info)==null?void 0:g.total_pages)??0;if(!((((m=w==null?void 0:w.items)==null?void 0:m.length)??0)>0)&&H>1&&Z>=H-1){const Y=H-1,ee=await Se(Y,f);u(ee)}else u(w)}finally{q(!1)}},[]);ne(()=>{const _=U.on("requisitionLists/data",f=>{f&&f.items&&u(f)},{eager:!0});return()=>{_==null||_.off()}},[]),ne(()=>{s||h(1,qe)},[s,h]);const y=V(_=>{var g,m;const f=_??((g=s==null?void 0:s.page_info)==null?void 0:g.current_page)??1,E=((m=s==null?void 0:s.page_info)==null?void 0:m.page_size)??qe;return h(f,E)},[h,s]),b=V(async _=>{await h(1,_)},[h]);return{rows:_e(()=>((s==null?void 0:s.items)??[]).map(_=>({name:R("div",{className:"requisition-list-grid-wrapper__name",children:[e("div",{className:"requisition-list-grid-wrapper__name__title",children:e("a",{href:"#",onClick:f=>{if(f.preventDefault(),n){const E=n(_.uid);typeof E=="string"&&(window.location.href=E)}},children:_.name})}),_.description&&e("div",{className:"requisition-list-grid-wrapper__name__description",children:_.description})]}),items_count:_.items_count,last_updated:new Date(_.updated_at).toLocaleString(),actions:R("div",{className:"requisition-list-grid-wrapper__actions",children:[e(Q,{variant:"tertiary",type:"button","data-testid":"rename-button",onClick:()=>p==null?void 0:p.handleOpenRenameModal(_),children:i.actionRename}),e(Q,{variant:"tertiary",type:"button","data-testid":"delete-button",onClick:()=>p==null?void 0:p.handleOpenDeleteModal(_),children:i.actionDeleteList})]})})),[s==null?void 0:s.items,i.actionRename,i.actionDeleteList,p,n]),isLoading:c||!s,pageInfo:s==null?void 0:s.page_info,totalCount:s==null?void 0:s.total_count,handlePageChange:y,handlePageSizeChange:b,isAdding:l,handleAddNew:T,handleCancelCreate:k}}function Je(){const t=J.config;return t?t.is_requisition_list_active==="1"&&t.company_enabled===!0:!1}const Ie=()=>{const[t,n]=I(Je);return ne(()=>{const r=U.on("requisitionList/initialized",()=>{const i=Je();n(s=>J.config!=null?i:s)});return()=>r==null?void 0:r.off()},[]),{isEnabled:t}},Rt=({routeRequisitionListDetails:t,fallbackRoute:n="/customer/account",slots:r})=>{const{isEnabled:i}=Ie(),[s,u]=I({type:null,isOpen:!1,isLoading:!1,requisitionList:null}),l=V(()=>{u({type:null,isOpen:!1,isLoading:!1,requisitionList:null})},[]),a=V(w=>{u({type:"rename",isOpen:!0,isLoading:!1,requisitionList:w})},[]),c=V(w=>{u({type:"delete",isOpen:!0,isLoading:!1,requisitionList:w})},[]),{rows:q,isLoading:T,pageInfo:k,totalCount:p,handlePageChange:h,handlePageSizeChange:y,isAdding:b,handleAddNew:D,handleCancelCreate:_}=Zi({handleOpenRenameModal:a,handleOpenDeleteModal:c},t,l),f=V(async w=>{if(s.requisitionList)try{await Ee(s.requisitionList.uid,w.name,w.description),U.emit("requisitionList/alert",{action:"update",type:"success",context:"requisitionList"}),await h(),l()}catch{U.emit("requisitionList/alert",{action:"update",type:"error",context:"requisitionList"})}},[s.requisitionList,h,l]),E=async()=>{s.requisitionList&&(u({...s,isLoading:!0}),await ii(s.requisitionList.uid).then(async()=>{U.emit("requisitionList/alert",{type:"success",action:"delete",context:"requisitionList"})}).catch(()=>{U.emit("requisitionList/alert",{type:"error",action:"delete",context:"requisitionList"})}).finally(async()=>{await h(),l()}))},g=j({containerTitle:"RequisitionList.containerTitle",updateTitle:"RequisitionList.RequisitionListForm.updateTitle",deleteRequisitionListTitle:"RequisitionList.RequisitionListWrapper.deleteRequisitionListTitle",deleteRequisitionListMessage:"RequisitionList.RequisitionListWrapper.deleteRequisitionListMessage",cancelAction:"RequisitionList.RequisitionListWrapper.cancelAction",confirmAction:"RequisitionList.RequisitionListWrapper.confirmAction",notEnabledTitle:"RequisitionList.RequisitionListsNotEnabled.title",notEnabledMessage:"RequisitionList.RequisitionListsNotEnabled.message",notEnabledActionLabel:"RequisitionList.RequisitionListsNotEnabled.actionLabel"}),m=V(()=>r!=null&&r.Header?e(Li,{name:"Header","aria-label":g.containerTitle,title:g.containerTitle,slot:r.Header}):e(Ye,{"aria-label":g.containerTitle,role:"region",title:g.containerTitle}),[r,g.containerTitle]);return i===!1?e(Te,{title:g.notEnabledTitle,message:g.notEnabledMessage,actionLabel:g.notEnabledActionLabel,onAction:()=>{window.location.href=n}}):R(de,{children:[e(ot,{header:m(),rows:q,skeletonRowCount:10,isLoading:T,pageInfo:k,totalCount:p,handlePageChange:h,handlePageSizeChange:y,defaultPageSize:10,isAdding:b,handleAddNew:D,handleCancelCreate:_}),s.type==="rename"&&s.isOpen&&s.requisitionList&&e(ue,{isOpen:s.isOpen,isLoading:s.isLoading,title:g.updateTitle,modalContent:e(xe,{mode:"update",defaultValues:{name:s.requisitionList.name,description:s.requisitionList.description||""},onSubmit:f,onCancel:l}),handleModalOnClose:l}),s.type==="delete"&&s.isOpen&&e(ue,{isOpen:s.isOpen,isLoading:s.isLoading,title:g.deleteRequisitionListTitle,modalContent:g.deleteRequisitionListMessage,confirmBtnCaption:g.confirmAction,closeBtnCaption:g.cancelAction,handleModalOnClose:l,handleModalOnConfirm:E})]})},Xi=t=>A.createElement("svg",{id:"Icon_Add_Base","data-name":"Icon \\u2013 Add \\u2013 Base",xmlns:"http://www.w3.org/2000/svg",width:24,height:24,viewBox:"0 0 24 24",...t},A.createElement("g",{id:"Large"},A.createElement("rect",{id:"Placement_area","data-name":"Placement area",width:24,height:24,fill:"#fff",opacity:0}),A.createElement("g",{id:"Add_icon","data-name":"Add icon",transform:"translate(9.734 9.737)"},A.createElement("line",{vectorEffect:"non-scaling-stroke",id:"Line_579","data-name":"Line 579",y2:12.7,transform:"translate(2.216 -4.087)",fill:"none",stroke:"currentColor"}),A.createElement("line",{vectorEffect:"non-scaling-stroke",id:"Line_580","data-name":"Line 580",x2:12.7,transform:"translate(-4.079 2.263)",fill:"none",stroke:"currentColor"})))),Ki=t=>A.createElement("svg",{width:24,height:24,viewBox:"0 0 24 24",fill:"none",xmlns:"http://www.w3.org/2000/svg",...t},A.createElement("g",{clipPath:"url(#clip0_102_196)"},A.createElement("path",{vectorEffect:"non-scaling-stroke",d:"M18.3601 18.16H6.5601L4.8801 3H2.3501M19.6701 19.59C19.6701 20.3687 19.0388 21 18.2601 21C17.4814 21 16.8501 20.3687 16.8501 19.59C16.8501 18.8113 17.4814 18.18 18.2601 18.18C19.0388 18.18 19.6701 18.8113 19.6701 19.59ZM7.42986 19.59C7.42986 20.3687 6.79858 21 6.01986 21C5.24114 21 4.60986 20.3687 4.60986 19.59C4.60986 18.8113 5.24114 18.18 6.01986 18.18C6.79858 18.18 7.42986 18.8113 7.42986 19.59Z",stroke:"currentColor",strokeLinejoin:"round"}),A.createElement("path",{vectorEffect:"non-scaling-stroke",d:"M5.25 6.37L20.89 8.06L20.14 14.8H6.19",stroke:"currentColor",strokeLinejoin:"round"})),A.createElement("defs",null,A.createElement("clipPath",{id:"clip0_102_196"},A.createElement("rect",{vectorEffect:"non-scaling-stroke",width:19.29,height:19.5,fill:"white",transform:"translate(2.3501 2.25)"})))),Ji=t=>A.createElement("svg",{width:24,height:24,viewBox:"0 0 24 24",fill:"none",xmlns:"http://www.w3.org/2000/svg",...t},A.createElement("path",{vectorEffect:"non-scaling-stroke",d:"M7.74512 9.87701L12.0001 14.132L16.2551 9.87701",stroke:"currentColor",strokeWidth:1,strokeLinecap:"square",strokeLinejoin:"round"})),ve=t=>A.createElement("svg",{width:24,height:24,viewBox:"0 0 24 24",fill:"none",xmlns:"http://www.w3.org/2000/svg",...t},A.createElement("rect",{x:4.89282,y:2.75,width:14.2143,height:18.5,rx:2.25,stroke:"currentColor",strokeWidth:1}),A.createElement("line",{x1:9.85718,y1:7.67871,x2:16.2857,y2:7.67871,stroke:"currentColor",strokeWidth:1}),A.createElement("line",{x1:9.85718,y1:11.9644,x2:16.2857,y2:11.9644,stroke:"currentColor",strokeWidth:1}),A.createElement("line",{x1:9.85718,y1:16.25,x2:16.2857,y2:16.25,stroke:"currentColor",strokeWidth:1}),A.createElement("circle",{cx:7.71429,cy:7.71429,r:.714286,fill:"currentColor"}),A.createElement("circle",{cx:7.71429,cy:11.9999,r:.714286,fill:"currentColor"}),A.createElement("circle",{cx:7.71429,cy:16.2856,r:.714286,fill:"currentColor"})),Gi=t=>A.createElement("svg",{width:24,height:24,viewBox:"0 0 24 24",fill:"none",xmlns:"http://www.w3.org/2000/svg",...t},A.createElement("path",{d:"M17.3332 11.75H6.6665",strokeWidth:1,strokeLinecap:"square",strokeLinejoin:"round",vectorEffect:"non-scaling-stroke",fill:"none",stroke:"currentColor"})),Yi=t=>A.createElement("svg",{id:"Icon_Search_Base","data-name":"Icon \\u2013 Search \\u2013 Base",xmlns:"http://www.w3.org/2000/svg",width:24,height:24,fill:"none",viewBox:"0 0 24 24",...t},A.createElement("g",{id:"Large"},A.createElement("rect",{id:"Placement_area","data-name":"Placement area",width:24,height:24,fill:"#fff",opacity:0}),A.createElement("g",{id:"Search_icon","data-name":"Search icon",transform:"translate(3.75 3.75)"},A.createElement("circle",{vectorEffect:"non-scaling-stroke",id:"Ellipse_186","data-name":"Ellipse 186",cx:6,cy:6,r:6,fill:"none",stroke:"currentColor"}),A.createElement("line",{vectorEffect:"non-scaling-stroke",id:"Line_556","data-name":"Line 556",x2:6,y2:6,transform:"translate(10.5 10.5)",fill:"none",stroke:"currentColor"})))),si=t=>A.createElement("svg",{xmlns:"http://www.w3.org/2000/svg",width:24,height:24,viewBox:"0 0 24 24",fill:"none",...t},A.createElement("path",{vectorEffect:"non-scaling-stroke",d:"M1 5H23",stroke:"currentColor",strokeWidth:1,strokeMiterlimit:10}),A.createElement("path",{vectorEffect:"non-scaling-stroke",d:"M17.3674 22H6.63446C5.67952 22 4.88992 21.2688 4.8379 20.3338L4 5H20L19.1621 20.3338C19.1119 21.2688 18.3223 22 17.3655 22H17.3674Z",stroke:"currentColor",strokeWidth:1,strokeMiterlimit:10}),A.createElement("path",{vectorEffect:"non-scaling-stroke",d:"M9.87189 2H14.1281C14.6085 2 15 2.39766 15 2.88889V5H9V2.88889C9 2.39912 9.39006 2 9.87189 2Z",stroke:"currentColor",strokeWidth:1,strokeMiterlimit:10}),A.createElement("path",{vectorEffect:"non-scaling-stroke",d:"M8.87402 8.58057L9.39348 17.682",stroke:"currentColor",strokeWidth:1,strokeMiterlimit:10}),A.createElement("path",{vectorEffect:"non-scaling-stroke",d:"M14.6673 8.58057L14.146 17.682",stroke:"currentColor",strokeWidth:1,strokeMiterlimit:10}));let ce=0;const ni=()=>{const[t,n]=I(J.requisitionLists),[r,i]=I(J.requisitionListsLoading),[s,u]=I(ce);return ne(()=>{J.requisitionLists.length===0&&!J.requisitionListsLoading&&(Ze(!0),Se(1,100).then(l=>{const a=(l==null?void 0:l.items)||[];Ce(a),ce++}).catch(l=>{console.error("Error fetching requisition lists:",l),Ce([]),ce++}).finally(()=>{Ze(!1)}))},[]),ne(()=>{const l=U.on("requisitionLists/data",c=>{c&&(Ce(c),ce++,u(ce),n(J.requisitionLists),i(J.requisitionListsLoading))}),a=U.on("requisitionList/data",c=>{c&&(Ni(c),ce++,u(ce),n([...J.requisitionLists]),i(J.requisitionListsLoading))});return()=>{l==null||l.off(),a==null||a.off()}},[]),ne(()=>{s!==ce&&(n(J.requisitionLists),i(J.requisitionListsLoading),u(ce))},[s]),{lists:t,loading:r}};function De(t){const[n,r]=I(null),i=j({errorCreate:"RequisitionList.RequisitionListAlert.errorCreate",successCreate:"RequisitionList.RequisitionListAlert.successCreate",errorDeleteItem:"RequisitionList.RequisitionListAlert.errorDeleteItem",successDeleteItem:"RequisitionList.RequisitionListAlert.successDeleteItem",errorDeleteReqList:"RequisitionList.RequisitionListAlert.errorDeleteReqList",successDeleteReqList:"RequisitionList.RequisitionListAlert.successDeleteReqList",errorAddToCart:"RequisitionList.RequisitionListAlert.errorAddToCart",successAddToCart:"RequisitionList.RequisitionListAlert.successAddToCart",errorUpdateQuantity:"RequisitionList.RequisitionListAlert.errorUpdateQuantity",successUpdateQuantity:"RequisitionList.RequisitionListAlert.successUpdateQuantity",errorUpdate:"RequisitionList.RequisitionListAlert.errorUpdate",successUpdate:"RequisitionList.RequisitionListAlert.successUpdate",errorMove:"RequisitionList.RequisitionListAlert.errorMove",successMove:"RequisitionList.RequisitionListAlert.successMove",errorAddToRequisitionList:"RequisitionList.RequisitionListAlert.errorAddToRequisitionList",successAddToRequisitionList:"RequisitionList.RequisitionListAlert.successAddToRequisitionList",errorMoveToList:"RequisitionList.RequisitionListAlert.errorMoveToList",successMoveToList:"RequisitionList.RequisitionListAlert.successMoveToList",errorCopyToList:"RequisitionList.RequisitionListAlert.errorCopyToList",successCopyToList:"RequisitionList.RequisitionListAlert.successCopyToList",errorImport:"RequisitionList.RequisitionListAlert.errorImport",successImport:"RequisitionList.RequisitionListAlert.successImport"}),s=_e(()=>({...i,...t}),[i,t]),u=_e(()=>({create:{requisitionList:{success:s.successCreate,error:s.errorCreate}},add:{product:{success:s.successAddToRequisitionList,error:s.errorAddToRequisitionList}},update:{product:{success:s.successUpdateQuantity,error:s.errorUpdateQuantity},requisitionList:{success:s.successUpdate,error:s.errorUpdate}},delete:{product:{success:s.successDeleteItem,error:s.errorDeleteItem},requisitionList:{success:s.successDeleteReqList,error:s.errorDeleteReqList}},move:{product:{success:s.successMove,error:s.errorMove}},moveToList:{product:{success:s.successMoveToList,error:s.errorMoveToList}},copyToList:{product:{success:s.successCopyToList,error:s.errorCopyToList}},import:{requisitionList:{success:s.successImport,error:s.errorImport}}}),[s]),l=V(a=>{const{type:c,action:q,context:T,skus:k,message:p,listName:h}=a;let y=p&&p.length>0?p.join(". "):u[q][T][c];h&&(y=y.replace("{listName}",h)),r({type:c,description:y,sku:k==null?void 0:k[0]});const b=setTimeout(()=>{r(null)},5e3);return()=>clearTimeout(b)},[u]);return{alert:n,setAlert:r,handleRequisitionListAlert:l}}function et(){const[t,n]=z(null),[r,i]=z(new Set),s=ae((a,c)=>{i(q=>{const T=new Set(q);return c?T.add(a):T.delete(a),T})},[]),u=ae(()=>{var c;const a=(c=t==null?void 0:t.items)==null?void 0:c.map(q=>q.uid);i(new Set(a))},[t]),l=ae(()=>{i(new Set)},[]);return{currentRequisitionList:t,setCurrentRequisitionList:n,selectedItems:r,setSelectedItems:i,handleItemSelection:s,handleSelectAll:u,handleSelectNone:l}}function it({sourceListUid:t,selectedItems:n,currentPageSize:r,currentPage:i,enrichConfigurableProductsInList:s,fetchAndMergeProducts:u,setCurrentRequisitionList:l,setSelectedItems:a,handleRequisitionListAlert:c}){const[q,T]=I(!1),[k,p]=I(!1),[h,y]=I(!1),[b,D]=I(!1),_=j({successMoveToList:"RequisitionList.RequisitionListAlert.successMoveToList",successCopyToList:"RequisitionList.RequisitionListAlert.successCopyToList"}),f=V(async g=>{var m;if(n.size!==0){p(!0);try{const w=await Mi(t,g,Array.from(n),r,i);if(w!=null&&w.sourceList){const H=await s(w.sourceList),Z=await u(H);l(Z),a(new Set);const X=((m=w.destinationList)==null?void 0:m.name)||"";c({action:"moveToList",type:"success",context:"product",message:[_.successMoveToList.replace("{listName}",X)]})}else c({action:"moveToList",type:"error",context:"product"})}catch{c({action:"moveToList",type:"error",context:"product"})}finally{p(!1),T(!1)}}},[n,t,r,i,s,u,l,a,c,_.successMoveToList]),E=V(async g=>{if(n.size!==0){D(!0);try{const m=await Ai(t,g,Array.from(n));if(m!=null&&m.destinationList){a(new Set);const w=m.destinationList.name||"";c({action:"copyToList",type:"success",context:"product",message:[_.successCopyToList.replace("{listName}",w)]})}else c({action:"copyToList",type:"error",context:"product"})}catch{c({action:"copyToList",type:"error",context:"product"})}finally{D(!1),y(!1)}}},[n,t,a,c,_.successCopyToList]);return{showMoveToListModal:q,setShowMoveToListModal:T,movingToList:k,showCopyToListModal:h,setShowCopyToListModal:y,copyingToList:b,handleMoveToList:f,handleCopyToList:E}}function tt(t,n,r){var l;const i=t.sku??((l=t.product)==null?void 0:l.sku);if(!i||i!==n.sku)return!1;if((r==null?void 0:r.matchBySkuOnly)!==!1)return!0;const s=(t.configurable_options??[]).map(a=>a.value_uid).filter(a=>!!a),u=(n.selectedOptions??[]).filter(a=>!!a);return s.length!==u.length?!1:(s.sort(),u.sort(),s.every((a,c)=>a===u[c]))}const _t=({canCreate:t=!0,sku:n,selectedOptions:r,quantity:i=1,matchBySKU:s,beforeAddProdToReqList:u})=>{const l=j({createTitle:"RequisitionList.RequisitionListForm.createTitle",addToRequisitionList:"RequisitionList.RequisitionListForm.addToRequisitionList",emptyList:"RequisitionList.RequisitionListWrapper.emptyList",addToNewRequisitionList:"RequisitionList.RequisitionListSelector.addToNewRequisitionList",addToSelected:"RequisitionList.RequisitionListSelector.addToSelected"}),[a,c]=I(!1),{lists:q}=ni(),{isEnabled:T}=Ie(),[k,p]=I(null);ne(()=>{const $=o=>{o&&p(o)},d=o=>{o&&p(C=>{const M=C??xi(),F=M.findIndex(L=>L.uid===o.uid);return F>=0?M.map((L,P)=>P===F?o:L):[...M,o]})},N=U.on("requisitionLists/data",$),B=U.on("requisitionList/data",d);return()=>{N==null||N.off(),B==null||B.off()}},[]);const h=k??q,y=_e(()=>{if(!(h!=null&&h.length))return!1;const $={sku:n,selectedOptions:r};return h.some(d=>{var N;return(N=d.items)==null?void 0:N.some(B=>tt(B,$,{matchBySkuOnly:s}))})},[h,n,r,s]),{alert:b,setAlert:D,handleRequisitionListAlert:_}=De(),[f,E]=I({isOpen:!1,isLoading:!1}),g=V(()=>{E({isOpen:!0,isLoading:!1})},[]),m=V(()=>{E({isOpen:!1,isLoading:!1}),c(!1),D(null)},[D]),w=V(async $=>{try{const d={sku:n,quantity:i,...r&&r.length>0?{selected_options:r}:{}};await Ei($,[d])}catch(d){throw console.error("Error adding product to list:",d),d}},[n,i,r]),H=V(async $=>{try{await w($),_({action:"add",type:"success",context:"product",skus:[n]})}catch{_({action:"add",type:"error",context:"product",skus:[n]})}finally{setTimeout(()=>{m()},2e3)}},[n,w,m,_]),Z=V(()=>{if(!u){g();return}Promise.resolve(u()).then(()=>{g()}).catch(()=>{})},[u,g]),X=(q==null?void 0:q.length)>0?a?R("button",{type:"button","aria-label":"Select a requisition list",role:"button",className:"requisition-list-actions","data-testid":"requisition-list-actions-button",onClick:()=>c(!1),children:[e("span",{className:"requisition-list-actions__title","data-testid":"requisition-list-actions-button-text",children:l.addToRequisitionList}),e(le,{source:Ji,size:"32"})]}):e(Ne,{confirmLabel:l.addToSelected,onConfirm:H}):e(Pe,{textContent:l.emptyList}),Y=a?e(be,{variant:"secondary",children:e(ti,{mode:"create",onSuccess:async $=>{await H($.uid)},onError:()=>{_({action:"add",type:"error",context:"product",skus:[n]})},onCancel:()=>{c(!1)}})}):e(oi,{onAddNew:()=>{c(!0)}}),ee=R(de,{children:[b&&e("div",{className:"requisition-list__alert-wrapper",children:e(he,{id:`requisition-list-selector__alert__${n}`,heading:b.description,type:b.type,variant:"primary",className:"requisition-list-selector__alert"},`requisition-list-selector__alert__${n}`)}),!b&&R(de,{children:[X,t&&Y]})]});return T===null||!T?null:R("div",{className:"requisition-list-selector",children:[e(Q,{active:y,activeIcon:e(le,{source:ve}),"aria-label":l.addToRequisitionList,className:y?"requisition-list-selector--active":void 0,"data-testid":"requisition-list-selector",size:"medium",variant:"tertiary",icon:e(le,{source:ve}),onClick:Z}),f.isOpen&&e(ue,{isOpen:!0,isLoading:f.isLoading,title:l.addToRequisitionList,modalContent:ee,handleModalOnClose:m})]})},st=({name:t,description:n,backLink:r,actions:i,className:s,...u})=>R("div",{...u,className:`requisition-list-header ${s||""}`,children:[r&&e("div",{className:"requisition-list-header__back",children:R("a",{href:r.url,className:"requisition-list-header__back-link",onClick:r.onClick,children:[e("span",{className:"requisition-list-header__back-arrow",children:"<"}),r.label]})}),R("div",{className:"requisition-list-header__main",children:[R("div",{className:"requisition-list-header__title-section",children:[e("h1",{className:"requisition-list-header__title",children:t}),n&&e("p",{className:"requisition-list-header__description",children:n})]}),i&&(i.onRename||i.onDelete||i.onShare||i.shareDisabled)&&R("div",{className:"requisition-list-header__actions",children:[(i.onShare||i.shareDisabled)&&e("a",{href:"#",onClick:l=>{var a;l.preventDefault(),!i.shareDisabled&&((a=i.onShare)==null||a.call(i))},className:`requisition-list-header__action-link ${i.shareDisabled?"requisition-list-header__action-link--disabled":""}`,"data-testid":"share-list-btn","aria-disabled":i.shareDisabled?"true":"false","aria-label":i.shareDisabled&&i.shareDisabledReason?`${i.shareLabel} – ${i.shareDisabledReason}`:void 0,"data-disabled-reason":i.shareDisabled?i.shareDisabledReason:void 0,children:i.shareLabel}),i.onRename&&e("a",{href:"#",onClick:l=>{var a;l.preventDefault(),(a=i.onRename)==null||a.call(i)},className:"requisition-list-header__action-link","data-testid":"rename-list-btn",children:i.renameLabel}),i.onDelete&&e("a",{href:"#",onClick:l=>{var a;l.preventDefault(),(a=i.onDelete)==null||a.call(i)},className:"requisition-list-header__action-link","data-testid":"delete-list-btn",children:i.deleteLabel})]})]})]}),ue=({isOpen:t,isLoading:n,title:r,modalContent:i,closeBtnCaption:s,confirmBtnCaption:u,handleModalOnClose:l,handleModalOnConfirm:a})=>t?e(Ri,{className:"requisition-list-modal--overlay","data-testid":"requisition-list-modal",size:"medium",centered:!1,title:r,onClose:l,backgroundDim:!0,clickToDismiss:!0,escapeToDismiss:!0,role:"dialog","aria-label":r,children:R("div",{className:"requisition-list-modal",children:[n?e("div",{className:"requisition-list-modal__spinner","data-testid":"progress-spinner",children:e(fe,{stroke:"4",size:"large"})}):null,e("p",{children:i}),R("div",{className:"requisition-list-modal__buttons",children:[l&&s&&e(Q,{"data-testid":"rl-modal-close-button",type:"button",onClick:l,variant:"secondary",disabled:n,children:s}),a&&u&&e(Q,{"data-testid":"rl-modal-confirm-button",type:"button",onClick:a,disabled:n,children:u})]})]})}):null,nt=({requisitionList:t,routeRequisitionListGrid:n,onUpdate:r,onAlert:i,enrichConfigurableProducts:s,currentCustomerEmail:u,routeSharedRequisitionList:l})=>{var $,d;const[a,c]=z(!1),[q,T]=z(!1),[k,p]=z(!1),[h,y]=z(!1),[b,D]=z(!1),_=($=J.config)==null?void 0:$.requisition_list_sharing_enabled,f=_!==!1&&_!=="0",E=Number(t.items_count??((d=t.items)==null?void 0:d.length)??0),g=E<=0||!J.isCompanyUser,m=j({actionBackToRequisitionLists:"RequisitionList.RequisitionListView.actionBackToRequisitionLists",actionRename:"RequisitionList.RequisitionListView.actionRename",actionDeleteList:"RequisitionList.RequisitionListView.actionDeleteList",actionShare:"RequisitionList.RequisitionListView.actionShare",shareDisabledReason:"RequisitionList.RequisitionListView.shareDisabledReason",shareDisabledNoCompany:"RequisitionList.RequisitionListView.shareDisabledNoCompany",shareListTitle:"RequisitionList.RequisitionListView.shareListTitle",deleteListTitle:"RequisitionList.RequisitionListView.deleteListTitle",deleteListMessage:"RequisitionList.RequisitionListView.deleteListMessage",confirmAction:"RequisitionList.RequisitionListWrapper.confirmAction",cancelAction:"RequisitionList.RequisitionListWrapper.cancelAction",updateTitle:"RequisitionList.RequisitionListForm.updateTitle"}),w=ae(()=>{c(!0)},[]),H=ae(async N=>{var B,o;try{const C=await Ee(t.uid,N.name,N.description,(B=t.page_info)==null?void 0:B.page_size,(o=t.page_info)==null?void 0:o.current_page,s);C?(r==null||r(C),c(!1),i==null||i({action:"update",type:"success",context:"requisitionList"})):i==null||i({action:"update",type:"error",context:"requisitionList"})}catch{i==null||i({action:"update",type:"error",context:"requisitionList"})}},[t.uid,t.page_info,r,i,s]),Z=ae(()=>{T(!0)},[]),X=ae(async()=>{y(!0);try{if(await ii(t.uid)){const B={action:"delete",type:"success",context:"requisitionList"};try{localStorage.setItem("requisitionListPendingAlert",JSON.stringify(B))}catch{}if(n){const o=n();o&&typeof o=="string"&&(window.location.href=o)}}else i==null||i({action:"delete",type:"error",context:"requisitionList"})}catch{i==null||i({action:"delete",type:"error",context:"requisitionList"})}finally{y(!1),T(!1)}},[t.uid,n,i]),Y=ae(()=>{p(!0)},[]),ee=ae(async N=>{D(!0);try{return await Ii(t.uid,N)}catch{return[{code:"SHARE_FAILED",message:"Unable to share list."}]}finally{D(!1)}},[t.uid]);return R(de,{children:[e(st,{name:t.name,description:t.description,backLink:n?{url:"#",label:m.actionBackToRequisitionLists,onClick:N=>{N.preventDefault();const B=n();B&&typeof B=="string"&&(window.location.href=B)}}:void 0,actions:{onRename:w,onDelete:Z,onShare:f&&!g?Y:void 0,renameLabel:m.actionRename,deleteLabel:m.actionDeleteList,shareLabel:f?m.actionShare:void 0,shareDisabled:f&&g,shareDisabledReason:f&&g?E<=0?m.shareDisabledReason:m.shareDisabledNoCompany:void 0}}),a&&e(ue,{isOpen:a,isLoading:!1,title:m.updateTitle,modalContent:e(xe,{mode:"update",defaultValues:{name:t.name,description:t.description||""},onSubmit:H,onCancel:()=>c(!1)}),handleModalOnClose:()=>c(!1)}),q&&e(ue,{isOpen:q,isLoading:h,title:m.deleteListTitle,modalContent:m.deleteListMessage,confirmBtnCaption:m.confirmAction,closeBtnCaption:m.cancelAction,handleModalOnClose:()=>T(!1),handleModalOnConfirm:X}),f&&k&&e(ue,{isOpen:k,isLoading:b,title:m.shareListTitle,modalContent:e($i,{requisitionListUid:t.uid,isSubmitting:b,onSubmit:ee,currentCustomerEmail:u,routeSharedRequisitionList:l}),handleModalOnClose:()=>p(!1)})]})},ot=({className:t,isLoading:n=!1,header:r,rows:i=[],skeletonRowCount:s=10,pageInfo:u,totalCount:l=0,handlePageChange:a,handlePageSizeChange:c,defaultPageSize:q=10,isAdding:T,handleAddNew:k,handleCancelCreate:p,...h})=>{const y=j({name:"RequisitionList.RequisitionListWrapper.name",itemsCount:"RequisitionList.RequisitionListWrapper.itemsCount",lastUpdated:"RequisitionList.RequisitionListWrapper.lastUpdated",actions:"RequisitionList.RequisitionListWrapper.actions",emptyList:"RequisitionList.RequisitionListWrapper.emptyList",show:"RequisitionList.PageSizePicker.show"}),b=((u==null?void 0:u.total_pages)??0)>0,D=!n&&i.length===0&&!b,{alert:_,setAlert:f,handleRequisitionListAlert:E}=De();return ne(()=>{const g=U.on("requisitionList/alert",E);try{const m=localStorage.getItem("requisitionListPendingAlert");if(m){const w=JSON.parse(m);E(w),localStorage.removeItem("requisitionListPendingAlert")}}catch{}return()=>{g==null||g.off()}},[E]),R("div",{...h,className:re(["requisition-list-grid-wrapper",t]),"data-testid":"requisition-list-grid-wrapper",children:[r&&e("div",{className:re(["requisition-list-grid-wrapper__header",t]),"data-testid":"requisition-list-grid-wrapper-header",children:e(Ge,{node:r})}),_&&e("div",{className:"requisition-list__alert-wrapper",children:e(he,{heading:_.description,type:_.type,variant:"primary",onDismiss:()=>f(null)})}),D?e(Pe,{textContent:y.emptyList}):R(de,{children:[e(Me,{columns:[{key:"name",label:y.name},{key:"items_count",label:y.itemsCount},{key:"last_updated",label:y.lastUpdated},{key:"actions",label:y.actions}],rowData:i,loading:n,skeletonRowCount:s}),u&&R("div",{className:re(["requisition-list-grid-wrapper__pagination",t]),children:[e(ri,{pageInfo:u,totalCount:l}),(u.total_pages||0)>1&&e(ei,{totalPages:u.total_pages,currentPage:u.current_page||1,onChange:a,disabled:n}),R("div",{className:"requisition-list-grid-wrapper__pagination-picker",children:[e("span",{children:y.show}),e(ai,{currentPageSize:u.page_size||q,onPageSizeChange:c||(()=>Promise.resolve()),disabled:n})]})]})]}),e("div",{className:re(["requisition-list-grid-wrapper__add-new",t]),children:T?e(be,{variant:"secondary",children:e(ti,{mode:"create",onSuccess:async()=>{await a(),p(),E({type:"success",action:"create",context:"requisitionList"})},onError:()=>{E({type:"error",action:"create",context:"requisitionList"})},onCancel:p})}):e(oi,{onAddNew:k})})]})},oi=({selectable:t,className:n,onAddNew:r})=>{const i=j({addNewReqListBtn:"RequisitionList.AddNewReqList.addNewReqListBtn"});return R("button",{type:"button","aria-label":i.addNewReqListBtn,role:"button",className:re(["requisition-list-actions",["requisition-list-actions--selectable",t],n]),"data-testid":"requisition-list-actions-button",onClick:r,children:[e("span",{className:"requisition-list-actions__title","data-testid":"requisition-list-actions-button-text",children:i.addNewReqListBtn}),e(le,{source:Xi,size:"32"})]})},Pe=({className:t,textContent:n,...r})=>R("div",{className:re(["empty-list",t]),"data-testid":"empty-list",...r,children:[e(le,{source:ve,size:"64",stroke:"2"}),n&&e("h4",{children:n})]}),Te=({className:t,title:n="404 - Not Found",message:r="The requisition list you are looking for does not exist.",actionLabel:i,onAction:s,...u})=>R("div",{className:re(["not-found",t]),"data-testid":"not-found",...u,children:[e(le,{source:Yi,size:"64",stroke:"2"}),e("h2",{children:n}),r&&e("p",{children:r}),i&&s&&e(Q,{variant:"primary",onClick:s,children:i})]}),at=({className:t,items:n,selectedItems:r,currentPage:i,pageSize:s,canEdit:u=!0,handleItemSelection:l,handleUpdateQuantity:a,onAddToCart:c,onDeleteItem:q,...T})=>{const[k,p]=z({}),[h,y]=z({}),b=j({productNameHeader:"RequisitionList.RequisitionListView.productListTable.headers.productName",skuHeader:"RequisitionList.RequisitionListView.productListTable.headers.sku",priceHeader:"RequisitionList.RequisitionListView.productListTable.headers.price",quantityHeader:"RequisitionList.RequisitionListView.productListTable.headers.quantity",subtotalHeader:"RequisitionList.RequisitionListView.productListTable.headers.subtotal",actionsHeader:"RequisitionList.RequisitionListView.productListTable.headers.actions",actionAddToCart:"RequisitionList.RequisitionListView.actionAddToCart",actionDelete:"RequisitionList.RequisitionListView.actionDelete",actionSelect:"RequisitionList.RequisitionListView.actionSelect",itemQuantity:"RequisitionList.RequisitionListView.productListTable.itemQuantity",outOfStock:"RequisitionList.RequisitionListView.productListTable.outOfStock",onlyXLeftInStock:"RequisitionList.RequisitionListView.productListTable.onlyXLeftInStock"}),D=[{label:"#",key:"index"},{label:"item",key:"item"},{label:b.priceHeader,key:"price"},{label:b.quantityHeader,key:"quantity"},{label:b.subtotalHeader,key:"subtotal"}];u&&(D.unshift({label:b.actionSelect,key:"selector"}),D.push({label:b.actionsHeader,key:"actions"}));const _=(o,C)=>{const M=o.target.checked;l(C.uid,M)},f=o=>(i-1)*s+o+1,E=o=>{var C,M;return((C=o.configured_product)==null?void 0:C.name)||((M=o.product)==null?void 0:M.name)||o.sku},g=o=>{var C,M,F,L,P,W;return(M=(C=o.configured_product)==null?void 0:C.images)!=null&&M.length?((F=o.configured_product.images[0])==null?void 0:F.url)||"":(P=(L=o.product)==null?void 0:L.images)!=null&&P.length&&((W=o.product.images[0])==null?void 0:W.url)||""},m=o=>{var C;return((C=o.product)==null?void 0:C.name)||o.sku},w=o=>{var C;return(C=o.configured_product)==null?void 0:C.name},H=o=>o.bundle_options?o.bundle_options.map((M,F)=>R("span",{className:"requisition-list-view-product-list-table__product-configurable-name",children:[M.values[0].label," x ",M.values[0].quantity]},M.values[0].label||F)):void 0,Z=o=>{var C;return((C=o.configured_product)==null?void 0:C.sku)||o.sku},X=o=>{var C,M,F,L,P,W,x,se;return Array.isArray(o.bundle_options)&&o.bundle_options.length>0?o.bundle_options.reduce((pe,K)=>{var ge,Re,Le;return pe+(((Re=(ge=K.values[0])==null?void 0:ge.priceV2)==null?void 0:Re.value)||0)*(((Le=K.values[0])==null?void 0:Le.quantity)||0)},0):((L=(F=(M=(C=o.configured_product)==null?void 0:C.price)==null?void 0:M.final)==null?void 0:F.amount)==null?void 0:L.value)||((se=(x=(W=(P=o.product)==null?void 0:P.price)==null?void 0:W.final)==null?void 0:x.amount)==null?void 0:se.value)||0},Y=o=>{var C,M,F,L,P,W,x,se;return((L=(F=(M=(C=o.configured_product)==null?void 0:C.price)==null?void 0:M.final)==null?void 0:F.amount)==null?void 0:L.currency)||((se=(x=(W=(P=o.product)==null?void 0:P.price)==null?void 0:W.final)==null?void 0:x.amount)==null?void 0:se.currency)},ee=o=>X(o)*o.quantity,$=(o,C)=>{const M=+o.target.value;M>0&&!Number.isNaN(o.target.value)&&y(F=>({...F,[C.uid]:M}))},d=async(o,C)=>{let M=+o.target.value;if(!(M>0&&M!==C.quantity)){y(F=>({...F,[C.uid]:C.quantity}));return}p(F=>({...F,[C.uid]:!0}));try{await a(C.uid,M)}finally{p(F=>({...F,[C.uid]:!1}))}},N=n.map((o,C)=>({selector:e("label",{id:`item-selector-${o.sku}-label`,children:e(yi,{className:"requisition-list-view-product-list-table__checkbox",name:`item-selector-${o.sku}`,"aria-label":`${b.actionSelect} ${m(o)}`,"data-testid":`item-checkbox-${o.sku}`,onChange:M=>_(M,o),value:o.sku,checked:r.has(o.uid)})}),index:e("div",{className:"requisition-list-view-product-list-table__index-container",children:f(C)}),item:R("div",{className:"requisition-list-view-product-list-table__item-container",children:[e(_i,{className:"requisition-list-view-product-list-table__thumbnail",alt:E(o),src:g(o)}),R("div",{className:"requisition-list-view-product-list-table__item-details",children:[e("div",{className:"requisition-list-view-product-list-table__product-name",children:m(o)}),o.stock_status==="OUT_OF_STOCK"&&e("div",{className:"requisition-list-view-product-list-table__out-of-stock",children:b.outOfStock}),o.stock_status!=="OUT_OF_STOCK"&&o.only_x_left_in_stock!==null&&o.only_x_left_in_stock<o.quantity&&e("div",{className:"requisition-list-view-product-list-table__low-stock",children:b.onlyXLeftInStock.replace("{count}",String(o.only_x_left_in_stock))}),e("span",{className:"requisition-list-view-product-list-table__product-configurable-name",children:w(o)}),e("div",{className:"requisition-list-view-product-list-table__sku",children:Z(o)}),H(o)]})]}),price:e(je,{className:"requisition-list-view-product-list-table__price",amount:X(o),currency:Y(o)}),quantity:e("span",{className:"requisition-list-view-product-list-table__quantity",children:e(ye,{disabled:!!k[o.uid],children:e(Ae,{id:`requisition-list-item-quantity-${o.sku}`,"data-testid":`requisition-list-item-quantity-${o.sku}`,name:"quantity",type:"text","aria-label":`${b.itemQuantity} - ${m(o)}`,value:h[o.uid]??o.quantity,onChange:M=>$(M,o),onBlur:M=>d(M,o)})})}),subtotal:e(je,{className:"requisition-list-view-product-list-table__price",amount:ee(o),currency:Y(o)}),actions:R("div",{className:"requisition-list-view__bulk-actions",children:[e(Q,{type:"button",onClick:()=>c([o.uid]),icon:e(le,{source:Ki}),"aria-label":`${b.actionAddToCart} - ${m(o)}`,"data-testid":"product-list-table-add-to-cart-button"}),e(Q,{type:"button",variant:"secondary",icon:e(le,{source:si}),onClick:()=>q([o.uid]),"aria-label":`${b.actionDelete} - ${m(o)}`,"data-testid":"product-list-table-delete-button"})]})})),B=e(Me,{columns:D,rowData:N,"data-testid":"product-list-table",mobileLayout:"stacked"});return e(Ge,{node:Bi("div",{}),className:re(["requisition-list-view-product-list-table-container",t]),"data-testid":"product-list-table-container",...T,children:B})},rt=({selectedItems:t,deletingItemId:n,addingToCartItemId:r,bulkAddingToCart:i,updatingQuantityItemId:s,bulkMovingToList:u=!1,bulkCopyingToList:l=!1,onSelectAll:a,onSelectNone:c,onBulkAddToCart:q,onBulkDelete:T,onBulkMoveToList:k,onBulkCopyToList:p})=>{const h=j({statusDeleting:"RequisitionList.RequisitionListView.statusDeleting",actionSelectAll:"RequisitionList.RequisitionListView.actionSelectAll",actionSelectNone:"RequisitionList.RequisitionListView.actionSelectNone",actionAddToCart:"RequisitionList.RequisitionListView.actionAddToCart",actionDeleteSelectedItems:"RequisitionList.RequisitionListView.actionDeleteSelectedItems",statusBulkAddingToCart:"RequisitionList.RequisitionListView.statusBulkAddingToCart",actionMoveToList:"RequisitionList.RequisitionListView.actionMoveToList",actionCopyToList:"RequisitionList.RequisitionListView.actionCopyToList"}),y=n!==null||r!==null||i||s!==null||u||l,b=t.size>0;return R("div",{className:"requisition-list-view__batch-actions",children:[R("div",{className:"requisition-list-view__batch-actions-left",children:[e("button",{"data-testid":"bulk-actions-select-toggle-btn",type:"button",className:`requisition-list-view__batch-actions-select-toggle ${b?"requisition-list-view__batch-actions-select-toggle--active":""}`,onClick:b?c:a,disabled:y,"aria-label":b?h.actionSelectNone:h.actionSelectAll,children:e(le,{source:Gi})}),e("button",{type:"button",className:"requisition-list-view__batch-actions-select-label",onClick:b?c:a,disabled:y,children:h.actionSelectAll})]}),b&&R("div",{className:"requisition-list-view__batch-actions-buttons",children:[e("span",{className:"requisition-list-view__batch-actions-count-badge","aria-label":`${t.size} items selected`,children:t.size}),k&&e(Q,{"data-testid":"bulk-actions-move-to-list-btn",type:"button",variant:"secondary",onClick:k,disabled:y,children:h.actionMoveToList}),p&&e(Q,{"data-testid":"bulk-actions-copy-to-list-btn",type:"button",variant:"secondary",onClick:p,disabled:y,children:h.actionCopyToList}),e(Q,{"data-testid":"bulk-actions-add-to-cart-btn",type:"button",variant:"secondary",onClick:q,disabled:y,children:i?h.statusBulkAddingToCart:h.actionAddToCart}),e("button",{"data-testid":"bulk-actions-delete-btn",type:"button",className:"requisition-list-view__batch-actions-delete-icon",onClick:T,disabled:y,"aria-label":h.actionDeleteSelectedItems,children:e(le,{source:si})})]})]})},ai=({currentPageSize:t,onPageSizeChange:n,disabled:r=!1,pageSizeOptions:i=[10,25,50,100]})=>{const s=j({itemsPerPage:"RequisitionList.PageSizePicker.itemsPerPage"}),u=a=>{const c=a.target,q=parseInt(c.value,10);n(q)},l=i.map(a=>({value:a.toString(),text:a.toString()}));return e(bi,{disabled:r,"data-testid":"page-size-picker",variant:"primary",size:"medium",value:t.toString(),options:l,handleSelect:u,"aria-label":s.itemsPerPage})},ri=({pageInfo:t,totalCount:n,className:r=""})=>{const i=j({itemsCounter:"RequisitionList.PaginationItemsCounter.itemsCounter"});if(!t||!n)return null;const s=t.page_size??qe,u=t.current_page??1,l=s*u,a=n,c=l-s+1,q=l>a?a:l;return e("span",{className:`pagination-items-counter ${r}`.trim(),children:i.itemsCounter.replace("{from}",c.toString()).replace("{to}",q.toString()).replace("{total}",a.toString())})},Ne=({excludeUid:t,confirmLabel:n,disabled:r=!1,onConfirm:i})=>{const{lists:s}=ni(),[u,l]=I(null),a=t?s.filter(c=>c.uid!==t):s;return e(be,{variant:"secondary",children:R("form",{onSubmit:c=>{c.preventDefault(),u&&i(u)},className:"requisition-list-picker__form",children:[e("div",{className:"requisition-list-picker__available-lists",children:a.map(c=>R(be,{variant:u===c.uid?"primary":"secondary",onClick:()=>l(c.uid),children:[e(le,{source:ve}),e("span",{children:c.name})]},c.uid))}),e("div",{className:"requisition-list-picker__actions",children:e(Q,{type:"submit",disabled:!u||r,children:n})})]})})},yt=({requisitionListUid:t,skipProductLoading:n=!1,pageSize:r=qe,routeRequisitionListGrid:i,fallbackRoute:s="/customer/account",getProductData:u,enrichConfigurableProducts:l,currentCustomerEmail:a,routeSharedRequisitionList:c})=>{var ze,He,$e,We,Ue,Qe;const[q,T]=I(!1),[k,p]=I(null),[h,y]=I(null),[b,D]=I(!1),[_,f]=I(null),[E,g]=I(!1),[m,w]=I(r),[H,Z]=I(!0),[X,Y]=I(!1),[ee,$]=I([]),{currentRequisitionList:d,setCurrentRequisitionList:N,selectedItems:B,setSelectedItems:o,handleItemSelection:C,handleSelectAll:M,handleSelectNone:F}=et(),L=j({emptyRequisitionList:"RequisitionList.RequisitionListView.emptyRequisitionList",errorLoadingProducts:"RequisitionList.RequisitionListView.errorLoadingProducts",errorLoadPage:"RequisitionList.RequisitionListView.errorLoadPage",notFoundTitle:"RequisitionList.RequisitionListView.notFoundTitle",notFoundMessage:"RequisitionList.RequisitionListView.notFoundMessage",notFoundActionLabel:"RequisitionList.RequisitionListView.notFoundActionLabel",notEnabledTitle:"RequisitionList.RequisitionListsNotEnabled.title",notEnabledMessage:"RequisitionList.RequisitionListsNotEnabled.message",partialMoveSuccess:"RequisitionList.RequisitionListAlert.partialMoveSuccess",notEnabledActionLabel:"RequisitionList.RequisitionListsNotEnabled.actionLabel",show:"RequisitionList.PageSizePicker.show",deleteItemsTitle:"RequisitionList.RequisitionListView.deleteItemsTitle",deleteItemsMessage:"RequisitionList.RequisitionListView.deleteItemsMessage",confirmAction:"RequisitionList.RequisitionListView.confirmAction",cancelAction:"RequisitionList.RequisitionListView.cancelAction",moveToListTitle:"RequisitionList.RequisitionListView.moveToListTitle",moveToListConfirm:"RequisitionList.RequisitionListView.moveToListConfirm",copyToListTitle:"RequisitionList.RequisitionListView.copyToListTitle",copyToListConfirm:"RequisitionList.RequisitionListView.copyToListConfirm"}),{alert:P,setAlert:W,handleRequisitionListAlert:x}=De(),{isEnabled:se}=Ie();ne(()=>{const v=U.on("requisitionList/data",O=>{(O==null?void 0:O.uid)===t&&(N(O),o(new Set))}),S=U.on("requisitionList/alert",x);return()=>{v==null||v.off(),S==null||S.off()}},[x,t,N,o]),ne(()=>{w(r)},[r]);const pe=V(async v=>{var O;if(!((O=v.items)!=null&&O.length)||typeof l!="function")return v;const S=await l(v.items);return{...v,items:S}},[l]),K=V(async v=>{var O,G;const S=((O=v.items)==null?void 0:O.map(oe=>oe.sku))||[];if(S.length===0)return v;T(!0);try{const oe=await u(S);if(oe){const ie=new Map;return oe.forEach(te=>{ie.set(te.sku,te)}),{...v,items:(G=v.items)==null?void 0:G.map(te=>{const me=ie.get(te.sku);return{...te,product:me||te.product,stock_status:te.stock_status||(me==null?void 0:me.stock_status)||"IN_STOCK",only_x_left_in_stock:te.only_x_left_in_stock??(me==null?void 0:me.only_x_left_in_stock)??null}})}}return console.warn("No products found"),v}catch(oe){return console.warn(oe instanceof Error?oe.message:L.errorLoadingProducts),v}finally{T(!1)}},[u,L.errorLoadingProducts]),{showMoveToListModal:ge,setShowMoveToListModal:Re,movingToList:Le,showCopyToListModal:Ve,setShowCopyToListModal:Oe,copyingToList:we,handleMoveToList:li,handleCopyToList:ci}=it({sourceListUid:d==null?void 0:d.uid,selectedItems:B,currentPageSize:m,currentPage:((ze=d==null?void 0:d.page_info)==null?void 0:ze.current_page)||1,enrichConfigurableProductsInList:pe,fetchAndMergeProducts:K,setCurrentRequisitionList:N,setSelectedItems:o,handleRequisitionListAlert:x});ne(()=>{if(!t){Z(!1);return}if(d&&d.uid===t)return;Z(!0),(async()=>{try{const S=await ke(t,1,r,l);if(S)if(n)N(S);else{const O=await K(S);N(O)}}catch(S){console.error("Failed to initialize requisition list from UID:",S)}finally{Z(!1)}})()},[t,r,n,l,K,N,d]);const Be=V(async v=>{v&&v.length===1?(D(!1),y(v)):(D(!0),y(null));try{const S=await Di(d.uid,v),O=(v==null?void 0:v.length)||0,G=(S==null?void 0:S.length)||0,oe=O-G;if(S&&S.length>0&&oe>0){const ie={action:"move",type:"error",context:"product",message:[L.partialMoveSuccess.replace("{successCount}",String(oe)).replace("{failedCount}",String(G))]};x(ie),U.emit("requisitionList/alert",ie)}else if(S&&S.length>0){const ie={action:"move",type:"error",context:"product",message:S.map(te=>te.message)};x(ie),U.emit("requisitionList/alert",ie)}else{const ie={action:"move",type:"success",context:"product"};x(ie),U.emit("requisitionList/alert",ie)}}catch{const O={action:"move",type:"error",context:"product"};x(O),U.emit("requisitionList/alert",O)}finally{y(null),D(!1)}},[d,x,L]),Fe=V(v=>{v&&v.length>0&&($(v),Y(!0))},[]),ui=V(async()=>{ee.length===1?p(ee[0]):p("bulk");try{const v=await Pi(d.uid,ee,m,d.page_info.current_page,l);if(v){const S=await pe(v),O=await K(S);N(O),x({action:"delete",type:"success",context:"product"})}else x({action:"delete",type:"error",context:"product"})}catch{x({action:"delete",type:"error",context:"product"})}finally{p(null),Y(!1),$([])}},[ee,d,m,K,N,x,pe,l]),di=V(async(v,S)=>{var O;f(v);try{const G=(O=d.items)==null?void 0:O.find(te=>te.uid===v);if(!G){x({action:"update",type:"error",context:"product"});return}const oe={item_id:G.uid,entered_options:G.entered_options,selected_options:G.selected_options,quantity:S},ie=await Vi(d.uid,[oe],m,d.page_info.current_page,l);if(ie){const te=await K(ie);N(te),x({action:"update",type:"success",context:"product"})}else x({action:"update",type:"error",context:"product"})}catch{x({action:"update",type:"error",context:"product"})}finally{f(null)}},[d,K,N,m,x,l]),pi=V(async v=>{g(!0);try{const S=await ke(d.uid,v,m,l);if(S){const G=await K(S);N(G)}else console.warn(L.errorLoadPage)}catch(S){console.warn(S instanceof Error?S.message:L.errorLoadPage)}finally{g(!1)}},[d==null?void 0:d.uid,K,m,L,N,l]),mi=V(async v=>{w(v),g(!0);try{const S=await ke(d.uid,1,v,l);if(S){const G=await K(S);N(G)}else console.warn(L.errorLoadPage)}catch{console.warn(L.errorLoadPage)}finally{g(!1)}},[d==null?void 0:d.uid,K,L,N,l]),hi=V(async v=>{const S=await pe(v),O=await K(S);N(O)},[N,K,pe]);return R("div",{className:"requisition-list-view__container",children:[H?e("div",{className:"requisition-list-view__loading",children:e(fe,{stroke:"4",size:"large"})}):se===!1?e(Te,{title:L.notEnabledTitle,message:L.notEnabledMessage,actionLabel:L.notEnabledActionLabel,onAction:()=>{window.location.href=s}}):!d||!Oi(d==null?void 0:d.uid)?e(Te,{title:L.notFoundTitle,message:L.notFoundMessage,actionLabel:i?L.notFoundActionLabel:void 0,onAction:i?()=>{const v=i();v&&typeof v=="string"&&(window.location.href=v)}:void 0}):R(de,{children:[e(nt,{requisitionList:d,routeRequisitionListGrid:i,onUpdate:hi,onAlert:x,enrichConfigurableProducts:l,currentCustomerEmail:a,routeSharedRequisitionList:c}),P&&e("div",{className:"requisition-list__alert-wrapper",children:e(he,{heading:P.description,type:P.type,variant:"primary",onDismiss:()=>W(null)})}),d.items_count===0?e(Pe,{textContent:`${d.name} ${L.emptyRequisitionList}`}):R(de,{children:[e(rt,{selectedItems:B,deletingItemId:k,addingToCartItemId:h,bulkAddingToCart:b,updatingQuantityItemId:_,bulkMovingToList:Le,bulkCopyingToList:we,onSelectAll:M,onSelectNone:F,onBulkAddToCart:()=>Be(Array.from(B)),onBulkDelete:()=>Fe(Array.from(B)),onBulkMoveToList:()=>Re(!0),onBulkCopyToList:()=>Oe(!0)}),e(at,{items:d.items,selectedItems:B,currentPage:((He=d.page_info)==null?void 0:He.current_page)||1,pageSize:(($e=d.page_info)==null?void 0:$e.page_size)||qe,handleItemSelection:C,handleUpdateQuantity:di,onAddToCart:Be,onDeleteItem:Fe}),d.page_info&&R("div",{className:"requisition-list-view__pagination",children:[e(ri,{pageInfo:d.page_info,totalCount:d.items_count}),(((We=d.page_info)==null?void 0:We.total_pages)||0)>1&&e(ei,{totalPages:d.page_info.total_pages,currentPage:((Ue=d.page_info)==null?void 0:Ue.current_page)||1,onChange:pi,disabled:E||q}),R("div",{className:"requisition-list-view__pagination-picker",children:[e("span",{children:L.show}),e(ai,{currentPageSize:((Qe=d.page_info)==null?void 0:Qe.page_size)||qe,onPageSizeChange:mi,disabled:E||q})]})]})]})]}),X&&e(ue,{isOpen:X,isLoading:k!==null,title:L.deleteItemsTitle,modalContent:L.deleteItemsMessage,confirmBtnCaption:L.confirmAction,closeBtnCaption:L.cancelAction,handleModalOnClose:()=>{Y(!1),$([])},handleModalOnConfirm:ui}),ge&&e(ue,{isOpen:ge,isLoading:Le,title:L.moveToListTitle,modalContent:e(Ne,{excludeUid:d==null?void 0:d.uid,confirmLabel:L.moveToListConfirm,disabled:Le,onConfirm:li}),handleModalOnClose:()=>Re(!1)}),Ve&&e(ue,{isOpen:Ve,isLoading:we,title:L.copyToListTitle,modalContent:e(Ne,{excludeUid:d==null?void 0:d.uid,confirmLabel:L.copyToListConfirm,disabled:we,onConfirm:ci}),handleModalOnClose:()=>Oe(!1)})]})};export{ti as R,nt as RequisitionListHeader,$i as S,gt as SharedRequisitionList,Rt as a,_t as b,nt as c,yt as d,gt as default};
 //# sourceMappingURL=SharedRequisitionList.js.map
