+++ scripts/__dropins__/storefront-requisition-list/render.js
@@ -1,9 +1,5 @@
 /*! Copyright 2026 Adobe
 All Rights Reserved. */
-(function(o,e){try{if(typeof document<"u"){const i=document.createElement("style"),r=e.styleId;for(const t in e.attributes)i.setAttribute(t,e.attributes[t]);i.setAttribute("data-dropin",r),i.appendChild(document.createTextNode(o));const a=document.querySelector('style[data-dropin="sdk"]');if(a)a.after(i);else{const t=document.querySelector('link[rel="stylesheet"], style');t?t.before(i):document.head.append(i)}}}catch(i){console.error("dropin-styles (injectCodeFunction)",i)}})(`.shared-requisition-list__loading{display:flex;flex-direction:column;justify-content:center;align-items:center;min-height:200px;gap:var(--spacing-small);width:100%}.shared-requisition-list__container{width:100%;margin-top:var(--spacing-medium)}.shared-requisition-list__preview{margin-top:var(--spacing-medium);width:100%}.shared-requisition-list__alert-wrapper{margin-bottom:var(--spacing-medium)}.shared-requisition-list__preview-details{display:flex;flex-direction:column;gap:var(--spacing-small);margin-bottom:var(--spacing-small)}.shared-requisition-list__preview-row{display:flex;gap:var(--spacing-small)}.shared-requisition-list__preview-label{font:var(--type-body-1-strong-font);letter-spacing:var(--type-body-1-strong-letter-spacing);color:var(--color-neutral-800);min-width:120px;flex-shrink:0}.shared-requisition-list__preview-value{font:var(--type-body-1-default-font);letter-spacing:var(--type-body-1-default-letter-spacing);color:var(--color-neutral-900)}.shared-requisition-list__actions{display:flex;justify-content:flex-end;gap:var(--spacing-small);align-items:center}.shared-requisition-list__table-wrapper{margin-bottom:var(--spacing-large)}
-.share-requisition-list-content{display:flex;flex-direction:column;gap:var(--spacing-small)}.requisition-list-modal .share-requisition-list-content__instruction{color:var(--color-neutral-700)}.share-requisition-list-content__success{display:flex;flex-direction:column;gap:var(--spacing-xxsmall)}.share-requisition-list-content__success .share-requisition-list-content__instruction{margin:0}.share-requisition-list-content__recipient-list{display:flex;flex-direction:column}.share-requisition-list-content__recipient-list .share-requisition-list-content__recipient{margin:0;color:var(--color-neutral-800)}.share-requisition-list-content__field{display:flex;flex-direction:column;gap:var(--spacing-xsmall)}.share-requisition-list-content__multi-select{flex:1;min-width:0}.share-requisition-list-content__multi-select .dropin-multi-select__container{min-height:unset}.share-requisition-list-content__field .dropin-field__hint{margin:0}.share-requisition-list-content__loading{display:flex;align-items:center;gap:var(--spacing-xsmall);color:var(--color-neutral-600);padding:var(--spacing-xsmall) 0}.share-requisition-list-content__actions{display:flex;justify-content:end;gap:var(--spacing-small)}.share-requisition-list-content__actions button{min-width:140px}.share-requisition-list-content__divider-secondary{width:100%}.share-requisition-list-content__link-row{display:flex;gap:var(--spacing-small);align-items:center}.share-requisition-list-content__link-row .dropin-input-container{flex:1;min-width:0}
-.requisition-list-header{margin-bottom:var(--spacing-large)}.requisition-list-header__back{margin-bottom:var(--spacing-medium)}.requisition-list-header__back-link{display:inline-block;color:var(--color-neutral-800);font-size:var(--type-body-1-font-size);font-weight:var(--type-body-1-font-weight);line-height:var(--type-body-1-line-height);text-decoration:none!important;cursor:pointer;transition:color .2s ease;position:relative;padding-bottom:1px}.requisition-list-header__back-link *{text-decoration:none!important}.requisition-list-header__back-link:hover{color:var(--color-neutral-600);text-decoration:none!important}.requisition-list-header__back-link:hover:after{content:"";position:absolute;left:0;right:0;bottom:0;height:1px;background-color:var(--color-neutral-800)}.requisition-list-header__back-arrow{margin-right:var(--spacing-xsmall);text-decoration:none!important}.requisition-list-header__main{display:flex;justify-content:space-between;align-items:flex-start;gap:var(--spacing-medium);flex-wrap:wrap}.requisition-list-header__title-section{flex:1;min-width:0}.requisition-list-header__title{margin:0;padding:0;font:var(--type-headline-1-font);letter-spacing:var(--type-headline-1-letter-spacing);color:var(--color-neutral-900)}.requisition-list-header__description{margin:var(--spacing-xsmall) 0 0;padding:0;font-size:var(--type-body-1-font-size);font-weight:var(--type-body-1-font-weight);line-height:var(--type-body-1-line-height);color:var(--color-neutral-700)}.requisition-list-header__actions{display:flex;gap:var(--spacing-medium);flex-wrap:wrap;align-items:center}.requisition-list-header__action-link{color:var(--color-neutral-800);font-size:var(--type-body-2-font-size);font-style:normal;font-weight:var(--type-body-2-font-weight);line-height:var(--type-body-2-line-height);letter-spacing:var(--type-body-2-letter-spacing);text-decoration:underline;cursor:pointer;transition:color .2s ease}.requisition-list-header__action-link:hover{color:var(--color-neutral-600)}.requisition-list-header__action-link--disabled,.requisition-list-header__action-link--disabled:hover{color:var(--color-neutral-500)!important;cursor:not-allowed;text-decoration:none;position:relative}.requisition-list-header__action-link--disabled[data-disabled-reason]:hover:after,.requisition-list-header__action-link--disabled[data-disabled-reason]:focus-visible:after{content:attr(data-disabled-reason);position:absolute;top:calc(100% + 6px);left:0;z-index:5;max-width:260px;padding:6px 8px;border-radius:4px;background:var(--color-neutral-900);color:var(--color-neutral-100);font:var(--type-details-caption-2-font);line-height:1.3;white-space:normal}@media (max-width: 768px){.requisition-list-header__main{flex-direction:column;align-items:stretch}.requisition-list-header__actions{justify-content:flex-start}}@supports (height: 100dvh){.requisition-list-modal .dropin-modal{height:100dvh}}.dropin-modal__body--medium>.dropin-modal__header-title,.dropin-modal__body--full>.dropin-modal__header-title,.requisition-list-modal .dropin-modal__content{align-items:center;margin:0}.requisition-list-modal .dropin-modal__header-title-content h3{font:var(--type-headline-1-font);letter-spacing:var(--type-headline-1-letter-spacing);margin:0;padding:0}.requisition-list-modal .requisition-list-modal__spinner{align-items:center;background-color:var(--color-neutral-200);display:flex;height:100%;left:0;justify-content:center;opacity:var(--color-opacity-24);position:absolute;top:0;width:100%;z-index:1}.requisition-list-modal p{margin:var(--spacing-small) 0}.requisition-list-modal .requisition-list-modal__buttons{align-items:center;display:flex;gap:0 var(--spacing-small);justify-content:right}.requisition-list-modal--overlay{margin:auto;max-width:375px;padding:var(--spacing-small);position:relative;width:100%}@media (min-width: 768px){.requisition-list-modal--overlay{margin:auto;max-width:768px;padding:var(--spacing-xxbig);position:relative;width:100%}.requisition-list-modal p{color:var(--color-neutral-800);font:var(--type-body-2-strong-font);letter-spacing:var(--type-body-2-strong-letter-spacing);margin-bottom:var(--spacing-medium)}}
-.requisition-list-form{background-color:var(--color-neutral-50);box-sizing:border-box}.requisition-list-form__title{color:var(--color-neutral-800);font:var(--type-headline-2-strong-font);letter-spacing:var(--type-headline-2-strong-letter-spacing);margin-bottom:var(--spacing-medium)}.requisition-list-form__notification{margin-bottom:var(--spacing-medium)}.requisition-list-form__form{display:grid;gap:var(--spacing-medium);grid-template-columns:1fr}.requisition-list-form__actions{align-items:center;display:flex;gap:0 var(--grid-2-gutters);justify-content:flex-end}.requisition-list-form_progress-spinner{float:right}
-.requisition-list-modal .requisition-list-actions{margin:var(--spacing-medium) auto}.requisition-list-grid-wrapper__content{display:grid;grid-template-columns:repeat(4,1fr);row-gap:var(--spacing-xsmall);width:100%}.requisition-list-grid-wrapper__list-header{display:contents}.requisition-list-grid-wrapper__pagination{grid-column:1 / -1;margin-top:var(--spacing-small);display:flex;justify-content:space-between;align-items:center}.requisition-list-grid-wrapper__pagination-picker{display:flex;align-items:center;gap:var(--spacing-xsmall);font:var(--type-body-1-default-font);letter-spacing:var(--type-body-1-default-letter-spacing);color:var(--color-neutral-800)}.requisition-list-grid-wrapper__add-new{margin:var(--spacing-medium) 0}.requisition-list-grid-wrapper__name__title{display:block;font:var(--type-body-1-emphasized-font);margin-bottom:2px}.requisition-list-grid-wrapper__name__description{font:var(--type-details-caption-2-font)}.requisition-list-grid-wrapper__actions{display:flex;gap:var(--spacing-medium)}.requisition-list-grid-wrapper__actions .dropin-button--tertiary{padding:0}.requisition-list-grid-wrapper__actions .dropin-button--tertiary:hover{text-decoration:underline}.requisition-list-empty-list{margin-bottom:var(--spacing-small);text-align:center}.requisition-list__alert-wrapper{margin-bottom:var(--spacing-small)}.requisition-list-actions{align-items:center;background-color:var(--color-neutral-50);border:var(--shape-border-width-2) solid var(--color-neutral-400);border-radius:var(--shape-border-radius-2);cursor:pointer;display:flex;justify-content:space-between;padding:var(--spacing-medium);width:100%}.requisition-list-actions--selectable{max-width:100%;width:auto}.requisition-list-actions__title{font:var(--type-button-2-font);letter-spacing:var(--type-button-2-letter-spacing);-moz-appearance:none;appearance:none;-webkit-appearance:none;color:var(--color-neutral-800)}.requisition-list-actions svg{color:var(--color-neutral-800);stroke:var(--shape-icon-stroke-4)}.empty-list{text-align:center;margin:auto 0}.not-found{display:flex;flex-direction:column;align-items:center;justify-content:center;padding:var(--spacing-large);text-align:center;min-height:400px}.not-found svg{color:var(--color-neutral-400);margin-bottom:var(--spacing-medium)}.not-found h2{font-size:var(--type-headline-3-font-size);font-weight:var(--type-headline-3-font-weight);margin:0 0 var(--spacing-small) 0;color:var(--color-neutral-800)}.not-found p{font-size:var(--type-body-1-font-size);color:var(--color-neutral-600);margin:0 0 var(--spacing-medium) 0;max-width:500px}.not-found button{margin-top:var(--spacing-small)}.requisition-list-view-product-list-table-container{display:flex;flex-direction:column;gap:var(--spacing-small)}.requisition-list-view-product-list-table__checkbox{align-self:center;justify-self:center}.requisition-list-view-product-list-table__index-container{display:inline-block;line-height:1}.requisition-list-view-product-list-table__item-container{display:flex;align-items:flex-start;gap:var(--spacing-small)}.requisition-list-view-product-list-table__item-details{display:flex;flex-direction:column;gap:var(--spacing-xxsmall)}.requisition-list-view-product-list-table__product-name{font:var(--type-body-1-font);letter-spacing:var(--type-body-1-letter-spacing);font-weight:700}.requisition-list-view-product-list-table__out-of-stock{font:var(--type-body-1-default-font);color:var(--color-error, #d32f2f)}.requisition-list-view-product-list-table__low-stock{font:var(--type-body-1-default-font);color:var(--color-warning, #ed6c02)}.requisition-list-view-product-list-table__product-configurable-name{font:var(--type-body-1-default-font)}.requisition-list-view-product-list-table__sku{font:var(--type-details-caption-1-font);letter-spacing:var(--type-details-caption-1-letter-spacing)}.requisition-list-view-product-list-table__thumbnail{width:44px;flex-shrink:0}.requisition-list-view-product-list-table__quantity{font:var(--type-body-1-font);letter-spacing:var(--type-body-1-letter-spacing);display:block;justify-self:center}.requisition-list-view-product-list-table__discount-container{display:flex;flex-direction:row;flex-wrap:wrap;gap:var(--spacing-xxsmall);align-items:baseline}.requisition-list-view-product-list-table-container__submit-container{display:flex;justify-content:flex-end}.requisition-list-view__batch-actions{--batch-actions-background: #f0f4f8;--batch-actions-toggle-border: #b3b3b3;--batch-actions-toggle-border-hover: #666666;--batch-actions-toggle-background: #ffffff;--batch-actions-toggle-background-hover: #f5f5f5;--batch-actions-toggle-active-background: #505050;--batch-actions-toggle-active-background-hover: #3d3d3d;--batch-actions-toggle-icon-active: #ffffff;--batch-actions-label-color: #2d2d2d;--batch-actions-badge-background: #378ef0;--batch-actions-badge-color: #ffffff;--batch-actions-button-background: #fafafa;--batch-actions-button-background-hover: #f0f0f0;--batch-actions-button-border: #d3d3d3;--batch-actions-delete-color: #4a4a4a;--batch-actions-delete-color-hover: #d7373f;display:flex;align-items:center;justify-content:space-between;padding:var(--spacing-medium) 0;margin-bottom:var(--spacing-small);background-color:var(--batch-actions-background);border-radius:0}.requisition-list-view__batch-actions-left{display:flex;align-items:center;gap:var(--spacing-small);padding-left:var(--spacing-medium)}.requisition-list-view__batch-actions-select-toggle{display:flex;align-items:center;justify-content:center;width:20px;height:20px;padding:0;border:1px solid var(--batch-actions-toggle-border);border-radius:0;background-color:var(--batch-actions-toggle-background);cursor:pointer;transition:all .2s ease}.requisition-list-view__batch-actions-select-toggle:hover:not(:disabled){border-color:var(--batch-actions-toggle-border-hover);background-color:var(--batch-actions-toggle-background-hover)}.requisition-list-view__batch-actions-select-toggle--active{background-color:var(--batch-actions-toggle-active-background);border-color:var(--batch-actions-toggle-active-background)}.requisition-list-view__batch-actions-select-toggle--active svg,.requisition-list-view__batch-actions-select-toggle--active path{fill:var(--batch-actions-toggle-icon-active);color:var(--batch-actions-toggle-icon-active)}.requisition-list-view__batch-actions-select-toggle--active:hover:not(:disabled){background-color:var(--batch-actions-toggle-active-background-hover);border-color:var(--batch-actions-toggle-active-background-hover)}.requisition-list-view__batch-actions-select-toggle:disabled{opacity:.5;cursor:not-allowed}.requisition-list-view__batch-actions-select-label{padding:0;border:none;background:none;font-size:var(--type-base-font-size);font-weight:var(--type-base-font-weight);color:var(--batch-actions-label-color);cursor:pointer;transition:opacity .2s ease}.requisition-list-view__batch-actions-select-label:hover:not(:disabled){opacity:.8}.requisition-list-view__batch-actions-select-label:disabled{opacity:.5;cursor:not-allowed}.requisition-list-view__batch-actions-count-badge{display:inline-flex;align-items:center;justify-content:center;width:24px;height:24px;padding:0;border-radius:50%;background-color:var(--batch-actions-badge-background);color:var(--batch-actions-badge-color);font-size:var(--type-detail-2-font-size);font-weight:var(--type-detail-2-font-weight);line-height:1;flex-shrink:0}.requisition-list-view__batch-actions-buttons{display:flex;align-items:center;gap:var(--spacing-small);padding-right:var(--spacing-medium)}.requisition-list-view__batch-actions-buttons button:not(.requisition-list-view__batch-actions-delete-icon){border-radius:var(--shape-border-radius-1);background-color:var(--batch-actions-button-background);border:1px solid var(--batch-actions-button-border)}.requisition-list-view__batch-actions-buttons button:not(.requisition-list-view__batch-actions-delete-icon):hover:not(:disabled){background-color:var(--batch-actions-button-background-hover)}.requisition-list-view__batch-actions-delete-icon{display:flex;align-items:center;justify-content:center;width:40px;height:40px;padding:0;border:none;background:transparent;color:var(--batch-actions-delete-color);cursor:pointer;transition:color .2s ease}.requisition-list-view__batch-actions-delete-icon:hover:not(:disabled){color:var(--batch-actions-delete-color-hover);background:none!important}.requisition-list-view__batch-actions-delete-icon:disabled{opacity:.5;cursor:not-allowed}.requisition-list-view__bulk-actions{margin-bottom:var(--spacing-xsmall);display:flex;gap:var(--spacing-xsmall);align-items:center}.page-size-picker{display:flex;align-items:center;gap:var(--spacing-xsmall);font:var(--type-body-1-default-font);letter-spacing:var(--type-body-1-default-letter-spacing);color:var(--color-neutral-800)}.page-size-picker__label{font-size:var(--font-size-small);color:var(--color-text-secondary);white-space:nowrap}.page-size-picker__select{min-width:60px}.pagination-items-counter{font:var(--type-body-1-default-font);letter-spacing:var(--type-body-1-default-letter-spacing);color:var(--color-neutral-800)}.requisition-list-picker__form .dropin-card__content{align-items:center;cursor:pointer;display:flex;flex-direction:row;flex-wrap:wrap;gap:var(--spacing-xxsmall);padding:var(--spacing-xsmall)}.requisition-list-picker__form .dropin-card--secondary{border:none}.requisition-list-picker__form .requisition-list-picker__actions{display:flex;justify-content:flex-end;margin-top:var(--spacing-medium)}.requisition-list-picker__available-lists{max-height:300px;min-height:200px;overflow-y:scroll}.requisition-list-picker__available-lists::-webkit-scrollbar{width:var(--spacing-small);display:block}.requisition-list-picker__available-lists::-webkit-scrollbar-track{background:var(--color-neutral-100);border-radius:var(--shape-border-radius-2)}.requisition-list-picker__available-lists::-webkit-scrollbar-thumb{background:var(--color-neutral-300);border-radius:var(--shape-border-radius-2)}.requisition-list-picker__available-lists::-webkit-scrollbar-thumb:hover{background:var(--color-neutral-500)}.requisition-list-view__container{width:100%}.requisition-list-view__loading{display:flex;justify-content:center;align-items:center;min-height:400px;width:100%}.requisition-list-view__pagination{margin-top:var(--spacing-small);display:flex;justify-content:space-between;align-items:center}.requisition-list-view__pagination-picker{display:flex;align-items:center;gap:var(--spacing-xsmall);font:var(--type-body-1-default-font);letter-spacing:var(--type-body-1-default-letter-spacing);color:var(--color-neutral-800)}`,{styleId:"RequisitionList"});
-import{jsx as i}from"@dropins/tools/preact-jsx-runtime.js";import{deepmerge as u,Render as c}from"@dropins/tools/lib.js";import{useState as d,useEffect as m}from"@dropins/tools/preact-hooks.js";import{UIProvider as p}from"@dropins/tools/components.js";import{events as L}from"@dropins/tools/event-bus.js";import{c as y}from"./chunks/initialize.js";import"./chunks/state.js";import"./chunks/fetch-error.js";import"./chunks/fetch-graphql.js";import"@dropins/tools/fetch-graphql.js";const q={containerTitle:"Requisition Lists",RequisitionListWrapper:{name:"Name & Description",itemsCount:"Items",lastUpdated:"Latest activity",actions:"Actions",loginMsg:"Please login",deleteRequisitionListTitle:"Are you sure you want to delete this Requisition List?",deleteRequisitionListMessage:"Requisition List will be permanently deleted. This action can not be undone.",confirmAction:"Confirm",cancelAction:"Cancel",emptyList:"No Requisition Lists found"},AddNewReqList:{addNewReqListBtn:"Add new Requisition List"},RequisitionListItem:{actionUpdate:"Update",actionDelete:"Delete"},RequisitionListForm:{actionCancel:"Cancel",actionSave:"Save",requiredField:"This is a required field.",nameMinLength:"Name must be at least {min} characters long.",nameInvalidCharacters:"Name contains invalid characters. Only letters, numbers, spaces, and basic punctuation are allowed.",floatingLabel:"Requisition List Name *",placeholder:"Requisition List Name",label:"Description",updateTitle:"Update Requisition List",createTitle:"Create Requisition List",addToRequisitionList:"Add to Requisition List"},RequisitionListSelector:{addToNewRequisitionList:"Add to New Requisition List",addToSelected:"Add to Selected List"},RequisitionListAlert:{errorCreate:"Error creating requisition list.",successCreate:"Requisition list created successfully.",errorAddToCart:"Error adding item to cart.",successAddToCart:"Item(s) added to cart successfully.",errorUpdateQuantity:"Error updating quantity.",successUpdateQuantity:"Item quantity updated successfully.",errorUpdate:"Error updating requisition list.",successUpdate:"Requisition list updated successfully.",errorDeleteItem:"Error deleting item.",successDeleteItem:"Item(s) deleted successfully.",errorDeleteReqList:"Error deleting requisition list.",successDeleteReqList:"Requisition list deleted successfully.",errorMove:"Error moving item(s) to cart.",successMove:"Item(s) successfully moved to cart.",partialMoveSuccess:"{successCount} product(s) successfully added and {failedCount} product(s) couldn't be added to your shopping cart.",errorAddToRequisitionList:"Error adding item(s) to requisition list.",successAddToRequisitionList:"Item(s) successfully added to requisition list.",errorMoveToList:"Error moving item(s) to requisition list.",successMoveToList:"Item(s) successfully moved to {listName}.",errorCopyToList:"Error copying item(s) to requisition list.",successCopyToList:"Item(s) successfully copied to {listName}.",errorImport:"Error importing requisition list.",successImport:'Requisition list "{listName}" has been added to your account.'},SharedRequisitionList:{loading:"Loading shared requisition list...",previewTitle:"Shared Requisition List",senderLabel:"Shared by",listNameLabel:"List name",descriptionLabel:"Description",itemsCountLabel:"Items",importButton:"Import List",importingButton:"Importing...",errorPreview:"Unable to load the shared requisition list.",skuHeader:"SKU",qtyHeader:"Qty",optionsHeader:"Options"},RequisitionListView:{actionDelete:"Delete",statusDeleting:"Deleting...",actionDeleteSelected:"Delete Selected",actionDeleteSelectedItems:"Delete selected items",actionSelect:"Select",actionSelectAll:"Select All",actionSelectNone:"Select None",actionAddToCart:"Add to Cart",statusAddingToCart:"Adding...",actionAddSelectedToCart:"Add Selected to Cart",statusBulkAddingToCart:"Adding to Cart...",actionUpdateQuantity:"Update",statusUpdatingQuantity:"Updating...",errorUpdateQuantity:"Error updating quantity",successUpdateQuantity:"Item quantity updated successfully.",actionBackToRequisitionListsOverview:"Back to requisition lists overview",actionBackToRequisitionLists:"Back to Requisition Lists",actionRename:"Rename",actionDeleteList:"Delete List",actionShare:"Share",shareDisabledReason:"You cannot share an empty requisition list.",shareDisabledNoCompany:"Sharing is available only for company account users.",shareListTitle:"Share Requisition List",deleteListTitle:"Delete Requisition List?",deleteListMessage:"Are you sure you want to delete this requisition list? This action cannot be undone.",deleteItemsTitle:"Delete Item(s)?",deleteItemsMessage:"Are you sure you want to delete the selected item(s) from this requisition list? This action cannot be undone.",confirmAction:"Delete",cancelAction:"Cancel",actionMoveToList:"Move to List",moveToListTitle:"Move to Requisition List",moveToListConfirm:"Move to requisition list",actionCopyToList:"Copy to List",copyToListTitle:"Copy to Requisition List",copyToListConfirm:"Copy to requisition list",emptyRequisitionList:" Requisition List is empty",productListTable:{headers:{productName:"Product name",sku:"SKU",price:"Price",quantity:"Quantity",subtotal:"Subtotal",actions:"Actions"},itemQuantity:"Item quantity",outOfStock:"Out of stock",onlyXLeftInStock:"Only {count} left in stock"},errorLoadPage:"Failed to load page",errorLoadingProducts:"Failed to load product data",notFoundTitle:"Requisition List Not Found",notFoundMessage:"The requisition list you are looking for does not exist or you do not have access to it.",notFoundActionLabel:"Back to Requisition Lists"},ShareRequisitionListContent:{emailInstruction:"Select below the email addresses of the people with whom you'd like to share your list. Your name, email address and the details of your list will be shared with recipients. You can send your list with multiple recipients. Please note only registered users will be able to import lists.",emailLabel:"Email addresses",emailPlaceholder:"Select company users",submitLabel:"Submit",linkInstruction:"Copy the link below and send it to people with whom you'd like to share your list. Please note only registered users will be able to import lists.",copyLink:"Copy Link",linkCopied:"Copied!",loadingUsers:"Loading users...",loadingLink:"Generating share link...",noUsersAvailable:"No company colleagues available to share with.",usersLoadError:"Unable to load company users. Please try again.",maxRecipientsValidation:"You can select up to {max} recipients.",shareSuccessMessage:"You have shared your list with the below email addresses:"},RequisitionListsNotEnabled:{title:"Requisition Lists Not Available",message:"Requisition Lists are not available. Please contact your administrator for more information.",actionLabel:"Go to My Account"},PageSizePicker:{show:"Show",itemsPerPage:"Items per page"},PaginationItemsCounter:{itemsCounter:"Items {from}-{to} of {total}"}},g={RequisitionList:q},h={default:g},R=({children:s})=>{var t;const[o,n]=d("en_US");m(()=>{const e=L.on("locale",l=>{n(l)},{eager:!0});return()=>{e==null||e.off()}},[]);const a=(t=y.getConfig())==null?void 0:t.langDefinitions,r=u(h,a??{});return i(p,{lang:o,langDefinitions:r,children:s})},k=new c(i(R,{}));export{k as render};
+(function(o,e){try{if(typeof document<"u"){const i=document.createElement("style"),r=e.styleId;for(const t in e.attributes)i.setAttribute(t,e.attributes[t]);i.setAttribute("data-dropin",r),i.appendChild(document.createTextNode(o));const a=document.querySelector('style[data-dropin="sdk"]');if(a)a.after(i);else{const t=document.querySelector('link[rel="stylesheet"], style');t?t.before(i):document.head.append(i)}}}catch(i){console.error("dropin-styles (injectCodeFunction)",i)}})('.shared-requisition-list__loading{display:flex;flex-direction:column;justify-content:center;align-items:center;min-height:200px;gap:var(--spacing-small);width:100%}.shared-requisition-list__container{width:100%;margin-top:var(--spacing-medium)}.shared-requisition-list__preview{margin-top:var(--spacing-medium);width:100%}.shared-requisition-list__alert-wrapper{margin-bottom:var(--spacing-medium)}.shared-requisition-list__preview-details{display:flex;flex-direction:column;gap:var(--spacing-small);margin-bottom:var(--spacing-small)}.shared-requisition-list__preview-row{display:flex;gap:var(--spacing-small)}.shared-requisition-list__preview-label{font:var(--type-body-1-strong-font);letter-spacing:var(--type-body-1-strong-letter-spacing);color:var(--color-neutral-800);min-width:120px;flex-shrink:0}.shared-requisition-list__preview-value{font:var(--type-body-1-default-font);letter-spacing:var(--type-body-1-default-letter-spacing);color:var(--color-neutral-900)}.shared-requisition-list__actions{display:flex;justify-content:flex-end;gap:var(--spacing-small);align-items:center}.shared-requisition-list__table-wrapper{margin-bottom:var(--spacing-large)}.share-requisition-list-content{display:flex;flex-direction:column;gap:var(--spacing-small)}.requisition-list-modal .share-requisition-list-content__instruction{color:var(--color-neutral-700)}.share-requisition-list-content__success{display:flex;flex-direction:column;gap:var(--spacing-xxsmall)}.share-requisition-list-content__success .share-requisition-list-content__instruction{margin:0}.share-requisition-list-content__recipient-list{display:flex;flex-direction:column}.share-requisition-list-content__recipient-list .share-requisition-list-content__recipient{margin:0;color:var(--color-neutral-800)}.share-requisition-list-content__field{display:flex;flex-direction:column;gap:var(--spacing-xsmall)}.share-requisition-list-content__multi-select{flex:1;min-width:0}.share-requisition-list-content__multi-select .dropin-multi-select__container{min-height:unset}.share-requisition-list-content__field .dropin-field__hint{margin:0}.share-requisition-list-content__loading{display:flex;align-items:center;gap:var(--spacing-xsmall);color:var(--color-neutral-600);padding:var(--spacing-xsmall) 0}.share-requisition-list-content__actions{display:flex;justify-content:end;gap:var(--spacing-small)}.share-requisition-list-content__actions button{min-width:140px}.share-requisition-list-content__divider-secondary{width:100%}.share-requisition-list-content__link-row{display:flex;gap:var(--spacing-small);align-items:center}.share-requisition-list-content__link-row .dropin-input-container{flex:1;min-width:0}.requisition-list-form{background-color:var(--color-neutral-50);box-sizing:border-box}.requisition-list-form__title{color:var(--color-neutral-800);font:var(--type-headline-2-strong-font);letter-spacing:var(--type-headline-2-strong-letter-spacing);margin-bottom:var(--spacing-medium)}.requisition-list-form__notification{margin-bottom:var(--spacing-medium)}.requisition-list-form__form{display:grid;gap:var(--spacing-medium);grid-template-columns:1fr}.requisition-list-form__actions{align-items:center;display:flex;gap:0 var(--grid-2-gutters);justify-content:flex-end}.requisition-list-form_progress-spinner{float:right}.requisition-list-modal .requisition-list-actions{margin:var(--spacing-medium) auto}.requisition-list-header{margin-bottom:var(--spacing-large)}.requisition-list-header__back{margin-bottom:var(--spacing-medium)}.requisition-list-header__back-link{display:inline-block;color:var(--color-neutral-800);font-size:var(--type-body-1-font-size);font-weight:var(--type-body-1-font-weight);line-height:var(--type-body-1-line-height);text-decoration:none!important;cursor:pointer;transition:color .2s ease;position:relative;padding-bottom:1px}.requisition-list-header__back-link *{text-decoration:none!important}.requisition-list-header__back-link:hover{color:var(--color-neutral-600);text-decoration:none!important}.requisition-list-header__back-link:hover:after{content:"";position:absolute;left:0;right:0;bottom:0;height:1px;background-color:var(--color-neutral-800)}.requisition-list-header__back-arrow{margin-right:var(--spacing-xsmall);text-decoration:none!important}.requisition-list-header__main{display:flex;justify-content:space-between;align-items:flex-start;gap:var(--spacing-medium);flex-wrap:wrap}.requisition-list-header__title-section{flex:1;min-width:0}.requisition-list-header__title{margin:0;padding:0;font:var(--type-headline-1-font);letter-spacing:var(--type-headline-1-letter-spacing);color:var(--color-neutral-900)}.requisition-list-header__description{margin:var(--spacing-xsmall) 0 0;padding:0;font-size:var(--type-body-1-font-size);font-weight:var(--type-body-1-font-weight);line-height:var(--type-body-1-line-height);color:var(--color-neutral-700)}.requisition-list-header__actions{display:flex;gap:var(--spacing-medium);flex-wrap:wrap;align-items:center}.requisition-list-header__action-link{color:var(--color-neutral-800);font-size:var(--type-body-2-font-size);font-style:normal;font-weight:var(--type-body-2-font-weight);line-height:var(--type-body-2-line-height);letter-spacing:var(--type-body-2-letter-spacing);text-decoration:underline;cursor:pointer;transition:color .2s ease}.requisition-list-header__action-link:hover{color:var(--color-neutral-600)}.requisition-list-header__action-link--disabled,.requisition-list-header__action-link--disabled:hover{color:var(--color-neutral-500)!important;cursor:not-allowed;text-decoration:none;position:relative}.requisition-list-header__action-link--disabled[data-disabled-reason]:hover:after,.requisition-list-header__action-link--disabled[data-disabled-reason]:focus-visible:after{content:attr(data-disabled-reason);position:absolute;top:calc(100% + 6px);left:0;z-index:5;max-width:260px;padding:6px 8px;border-radius:4px;background:var(--color-neutral-900);color:var(--color-neutral-100);font:var(--type-details-caption-2-font);line-height:1.3;white-space:normal}@media (max-width: 768px){.requisition-list-header__main{flex-direction:column;align-items:stretch}.requisition-list-header__actions{justify-content:flex-start}}@supports (height: 100dvh){.requisition-list-modal .dropin-modal{height:100dvh}}.dropin-modal__body--medium>.dropin-modal__header-title,.dropin-modal__body--full>.dropin-modal__header-title,.requisition-list-modal .dropin-modal__content{align-items:center;margin:0}.requisition-list-modal .dropin-modal__header-title-content h3{font:var(--type-headline-1-font);letter-spacing:var(--type-headline-1-letter-spacing);margin:0;padding:0}.requisition-list-modal .requisition-list-modal__spinner{align-items:center;background-color:var(--color-neutral-200);display:flex;height:100%;left:0;justify-content:center;opacity:var(--color-opacity-24);position:absolute;top:0;width:100%;z-index:1}.requisition-list-modal p{margin:var(--spacing-small) 0}.requisition-list-modal .requisition-list-modal__buttons{align-items:center;display:flex;gap:0 var(--spacing-small);justify-content:right}.requisition-list-modal--overlay{margin:auto;max-width:375px;padding:var(--spacing-small);position:relative;width:100%}@media (min-width: 768px){.requisition-list-modal--overlay{margin:auto;max-width:768px;padding:var(--spacing-xxbig);position:relative;width:100%}.requisition-list-modal p{color:var(--color-neutral-800);font:var(--type-body-2-strong-font);letter-spacing:var(--type-body-2-strong-letter-spacing);margin-bottom:var(--spacing-medium)}}.requisition-list-grid-wrapper__content{display:grid;grid-template-columns:repeat(4,1fr);row-gap:var(--spacing-xsmall);width:100%}.requisition-list-grid-wrapper__list-header{display:contents}.requisition-list-grid-wrapper__pagination{grid-column:1 / -1;margin-top:var(--spacing-small);display:flex;justify-content:space-between;align-items:center}.requisition-list-grid-wrapper__pagination-picker{display:flex;align-items:center;gap:var(--spacing-xsmall);font:var(--type-body-1-default-font);letter-spacing:var(--type-body-1-default-letter-spacing);color:var(--color-neutral-800)}.requisition-list-grid-wrapper__add-new{margin:var(--spacing-medium) 0}.requisition-list-grid-wrapper__name__title{display:block;font:var(--type-body-1-emphasized-font);margin-bottom:2px}.requisition-list-grid-wrapper__name__description{font:var(--type-details-caption-2-font)}.requisition-list-grid-wrapper__actions{display:flex;gap:var(--spacing-medium)}.requisition-list-grid-wrapper__actions .dropin-button--tertiary{padding:0}.requisition-list-grid-wrapper__actions .dropin-button--tertiary:hover{text-decoration:underline}.requisition-list-empty-list{margin-bottom:var(--spacing-small);text-align:center}.requisition-list__alert-wrapper{margin-bottom:var(--spacing-small)}.requisition-list-actions{align-items:center;background-color:var(--color-neutral-50);border:var(--shape-border-width-2) solid var(--color-neutral-400);border-radius:var(--shape-border-radius-2);cursor:pointer;display:flex;justify-content:space-between;padding:var(--spacing-medium);width:100%}.requisition-list-actions--selectable{max-width:100%;width:auto}.requisition-list-actions__title{font:var(--type-button-2-font);letter-spacing:var(--type-button-2-letter-spacing);-moz-appearance:none;appearance:none;-webkit-appearance:none;color:var(--color-neutral-800)}.requisition-list-actions svg{color:var(--color-neutral-800);stroke:var(--shape-icon-stroke-4)}.empty-list{text-align:center;margin:auto 0}.not-found{display:flex;flex-direction:column;align-items:center;justify-content:center;padding:var(--spacing-large);text-align:center;min-height:400px}.not-found svg{color:var(--color-neutral-400);margin-bottom:var(--spacing-medium)}.not-found h2{font-size:var(--type-headline-3-font-size);font-weight:var(--type-headline-3-font-weight);margin:0 0 var(--spacing-small) 0;color:var(--color-neutral-800)}.not-found p{font-size:var(--type-body-1-font-size);color:var(--color-neutral-600);margin:0 0 var(--spacing-medium) 0;max-width:500px}.not-found button{margin-top:var(--spacing-small)}.requisition-list-view-product-list-table-container{display:flex;flex-direction:column;gap:var(--spacing-small)}.requisition-list-view-product-list-table__checkbox{align-self:center;justify-self:center}.requisition-list-view-product-list-table__index-container{display:inline-block;line-height:1}.requisition-list-view-product-list-table__item-container{display:flex;align-items:flex-start;gap:var(--spacing-small)}.requisition-list-view-product-list-table__item-details{display:flex;flex-direction:column;gap:var(--spacing-xxsmall)}.requisition-list-view-product-list-table__product-name{font:var(--type-body-1-font);letter-spacing:var(--type-body-1-letter-spacing);font-weight:700}.requisition-list-view-product-list-table__out-of-stock{font:var(--type-body-1-default-font);color:var(--color-error, #d32f2f)}.requisition-list-view-product-list-table__low-stock{font:var(--type-body-1-default-font);color:var(--color-warning, #ed6c02)}.requisition-list-view-product-list-table__product-configurable-name{font:var(--type-body-1-default-font)}.requisition-list-view-product-list-table__sku{font:var(--type-details-caption-1-font);letter-spacing:var(--type-details-caption-1-letter-spacing)}.requisition-list-view-product-list-table__thumbnail{width:44px;flex-shrink:0}.requisition-list-view-product-list-table__quantity{font:var(--type-body-1-font);letter-spacing:var(--type-body-1-letter-spacing);display:block;justify-self:center}.requisition-list-view-product-list-table__discount-container{display:flex;flex-direction:row;flex-wrap:wrap;gap:var(--spacing-xxsmall);align-items:baseline}.requisition-list-view-product-list-table-container__submit-container{display:flex;justify-content:flex-end}.requisition-list-view__batch-actions{--batch-actions-background: #f0f4f8;--batch-actions-toggle-border: #b3b3b3;--batch-actions-toggle-border-hover: #666666;--batch-actions-toggle-background: #ffffff;--batch-actions-toggle-background-hover: #f5f5f5;--batch-actions-toggle-active-background: #505050;--batch-actions-toggle-active-background-hover: #3d3d3d;--batch-actions-toggle-icon-active: #ffffff;--batch-actions-label-color: #2d2d2d;--batch-actions-badge-background: #378ef0;--batch-actions-badge-color: #ffffff;--batch-actions-button-background: #fafafa;--batch-actions-button-background-hover: #f0f0f0;--batch-actions-button-border: #d3d3d3;--batch-actions-delete-color: #4a4a4a;--batch-actions-delete-color-hover: #d7373f;display:flex;align-items:center;justify-content:space-between;padding:var(--spacing-medium) 0;margin-bottom:var(--spacing-small);background-color:var(--batch-actions-background);border-radius:0}.requisition-list-view__batch-actions-left{display:flex;align-items:center;gap:var(--spacing-small);padding-left:var(--spacing-medium)}.requisition-list-view__batch-actions-select-toggle{display:flex;align-items:center;justify-content:center;width:20px;height:20px;padding:0;border:1px solid var(--batch-actions-toggle-border);border-radius:0;background-color:var(--batch-actions-toggle-background);cursor:pointer;transition:all .2s ease}.requisition-list-view__batch-actions-select-toggle:hover:not(:disabled){border-color:var(--batch-actions-toggle-border-hover);background-color:var(--batch-actions-toggle-background-hover)}.requisition-list-view__batch-actions-select-toggle--active{background-color:var(--batch-actions-toggle-active-background);border-color:var(--batch-actions-toggle-active-background)}.requisition-list-view__batch-actions-select-toggle--active svg,.requisition-list-view__batch-actions-select-toggle--active path{fill:var(--batch-actions-toggle-icon-active);color:var(--batch-actions-toggle-icon-active)}.requisition-list-view__batch-actions-select-toggle--active:hover:not(:disabled){background-color:var(--batch-actions-toggle-active-background-hover);border-color:var(--batch-actions-toggle-active-background-hover)}.requisition-list-view__batch-actions-select-toggle:disabled{opacity:.5;cursor:not-allowed}.requisition-list-view__batch-actions-select-label{padding:0;border:none;background:none;font-size:var(--type-base-font-size);font-weight:var(--type-base-font-weight);color:var(--batch-actions-label-color);cursor:pointer;transition:opacity .2s ease}.requisition-list-view__batch-actions-select-label:hover:not(:disabled){opacity:.8}.requisition-list-view__batch-actions-select-label:disabled{opacity:.5;cursor:not-allowed}.requisition-list-view__batch-actions-count-badge{display:inline-flex;align-items:center;justify-content:center;width:24px;height:24px;padding:0;border-radius:50%;background-color:var(--batch-actions-badge-background);color:var(--batch-actions-badge-color);font-size:var(--type-detail-2-font-size);font-weight:var(--type-detail-2-font-weight);line-height:1;flex-shrink:0}.requisition-list-view__batch-actions-buttons{display:flex;align-items:center;gap:var(--spacing-small);padding-right:var(--spacing-medium)}.requisition-list-view__batch-actions-buttons button:not(.requisition-list-view__batch-actions-delete-icon){border-radius:var(--shape-border-radius-1);background-color:var(--batch-actions-button-background);border:1px solid var(--batch-actions-button-border)}.requisition-list-view__batch-actions-buttons button:not(.requisition-list-view__batch-actions-delete-icon):hover:not(:disabled){background-color:var(--batch-actions-button-background-hover)}.requisition-list-view__batch-actions-delete-icon{display:flex;align-items:center;justify-content:center;width:40px;height:40px;padding:0;border:none;background:transparent;color:var(--batch-actions-delete-color);cursor:pointer;transition:color .2s ease}.requisition-list-view__batch-actions-delete-icon:hover:not(:disabled){color:var(--batch-actions-delete-color-hover);background:none!important}.requisition-list-view__batch-actions-delete-icon:disabled{opacity:.5;cursor:not-allowed}.requisition-list-view__bulk-actions{margin-bottom:var(--spacing-xsmall);display:flex;gap:var(--spacing-xsmall);align-items:center}.page-size-picker{display:flex;align-items:center;gap:var(--spacing-xsmall);font:var(--type-body-1-default-font);letter-spacing:var(--type-body-1-default-letter-spacing);color:var(--color-neutral-800)}.page-size-picker__label{font-size:var(--font-size-small);color:var(--color-text-secondary);white-space:nowrap}.page-size-picker__select{min-width:60px}.pagination-items-counter{font:var(--type-body-1-default-font);letter-spacing:var(--type-body-1-default-letter-spacing);color:var(--color-neutral-800)}.requisition-list-picker__form .dropin-card__content{align-items:center;cursor:pointer;display:flex;flex-direction:row;flex-wrap:wrap;gap:var(--spacing-xxsmall);padding:var(--spacing-xsmall)}.requisition-list-picker__form .dropin-card--secondary{border:none}.requisition-list-picker__form .requisition-list-picker__actions{display:flex;justify-content:flex-end;margin-top:var(--spacing-medium)}.requisition-list-picker__available-lists{max-height:300px;min-height:200px;overflow-y:scroll}.requisition-list-picker__available-lists::-webkit-scrollbar{width:var(--spacing-small);display:block}.requisition-list-picker__available-lists::-webkit-scrollbar-track{background:var(--color-neutral-100);border-radius:var(--shape-border-radius-2)}.requisition-list-picker__available-lists::-webkit-scrollbar-thumb{background:var(--color-neutral-300);border-radius:var(--shape-border-radius-2)}.requisition-list-picker__available-lists::-webkit-scrollbar-thumb:hover{background:var(--color-neutral-500)}.requisition-list-view__container{width:100%}.requisition-list-view__loading{display:flex;justify-content:center;align-items:center;min-height:400px;width:100%}.requisition-list-view__pagination{margin-top:var(--spacing-small);display:flex;justify-content:space-between;align-items:center}.requisition-list-view__pagination-picker{display:flex;align-items:center;gap:var(--spacing-xsmall);font:var(--type-body-1-default-font);letter-spacing:var(--type-body-1-default-letter-spacing);color:var(--color-neutral-800)}',{styleId:"RequisitionList"});
+import{jsx as i}from"@dropins/tools/preact-jsx-runtime.js";import{deepmerge as u,Render as c}from"@dropins/tools/lib.js";import{useState as d,useEffect as m}from"@dropins/tools/preact-hooks.js";import{UIProvider as p}from"@dropins/tools/components.js";import{events as L}from"@dropins/tools/event-bus.js";import{config as y}from"./api.js";import"@dropins/tools/fetch-graphql.js";const q={containerTitle:"Requisition Lists",RequisitionListWrapper:{name:"Name & Description",itemsCount:"Items",lastUpdated:"Latest activity",actions:"Actions",loginMsg:"Please login",deleteRequisitionListTitle:"Are you sure you want to delete this Requisition List?",deleteRequisitionListMessage:"Requisition List will be permanently deleted. This action can not be undone.",confirmAction:"Confirm",cancelAction:"Cancel",emptyList:"No Requisition Lists found"},AddNewReqList:{addNewReqListBtn:"Add new Requisition List"},RequisitionListItem:{actionUpdate:"Update",actionDelete:"Delete"},RequisitionListForm:{actionCancel:"Cancel",actionSave:"Save",requiredField:"This is a required field.",nameMinLength:"Name must be at least {min} characters long.",nameInvalidCharacters:"Name contains invalid characters. Only letters, numbers, spaces, and basic punctuation are allowed.",floatingLabel:"Requisition List Name *",placeholder:"Requisition List Name",label:"Description",updateTitle:"Update Requisition List",createTitle:"Create Requisition List",addToRequisitionList:"Add to Requisition List"},RequisitionListSelector:{addToNewRequisitionList:"Add to New Requisition List",addToSelected:"Add to Selected List"},RequisitionListAlert:{errorCreate:"Error creating requisition list.",successCreate:"Requisition list created successfully.",errorAddToCart:"Error adding item to cart.",successAddToCart:"Item(s) added to cart successfully.",errorUpdateQuantity:"Error updating quantity.",successUpdateQuantity:"Item quantity updated successfully.",errorUpdate:"Error updating requisition list.",successUpdate:"Requisition list updated successfully.",errorDeleteItem:"Error deleting item.",successDeleteItem:"Item(s) deleted successfully.",errorDeleteReqList:"Error deleting requisition list.",successDeleteReqList:"Requisition list deleted successfully.",errorMove:"Error moving item(s) to cart.",successMove:"Item(s) successfully moved to cart.",partialMoveSuccess:"{successCount} product(s) successfully added and {failedCount} product(s) couldn't be added to your shopping cart.",errorAddToRequisitionList:"Error adding item(s) to requisition list.",successAddToRequisitionList:"Item(s) successfully added to requisition list.",errorMoveToList:"Error moving item(s) to requisition list.",successMoveToList:"Item(s) successfully moved to {listName}.",errorCopyToList:"Error copying item(s) to requisition list.",successCopyToList:"Item(s) successfully copied to {listName}.",errorImport:"Error importing requisition list.",successImport:'Requisition list "{listName}" has been added to your account.'},SharedRequisitionList:{loading:"Loading shared requisition list...",previewTitle:"Shared Requisition List",senderLabel:"Shared by",listNameLabel:"List name",descriptionLabel:"Description",itemsCountLabel:"Items",importButton:"Import List",importingButton:"Importing...",errorPreview:"Unable to load the shared requisition list.",skuHeader:"SKU",qtyHeader:"Qty",optionsHeader:"Options"},RequisitionListView:{actionDelete:"Delete",statusDeleting:"Deleting...",actionDeleteSelected:"Delete Selected",actionDeleteSelectedItems:"Delete selected items",actionSelect:"Select",actionSelectAll:"Select All",actionSelectNone:"Select None",actionAddToCart:"Add to Cart",statusAddingToCart:"Adding...",actionAddSelectedToCart:"Add Selected to Cart",statusBulkAddingToCart:"Adding to Cart...",actionUpdateQuantity:"Update",statusUpdatingQuantity:"Updating...",errorUpdateQuantity:"Error updating quantity",successUpdateQuantity:"Item quantity updated successfully.",actionBackToRequisitionListsOverview:"Back to requisition lists overview",actionBackToRequisitionLists:"Back to Requisition Lists",actionRename:"Rename",actionDeleteList:"Delete List",actionShare:"Share",shareDisabledReason:"You cannot share an empty requisition list.",shareDisabledNoCompany:"Sharing is available only for company account users.",shareListTitle:"Share Requisition List",deleteListTitle:"Delete Requisition List?",deleteListMessage:"Are you sure you want to delete this requisition list? This action cannot be undone.",deleteItemsTitle:"Delete Item(s)?",deleteItemsMessage:"Are you sure you want to delete the selected item(s) from this requisition list? This action cannot be undone.",confirmAction:"Delete",cancelAction:"Cancel",actionMoveToList:"Move to List",moveToListTitle:"Move to Requisition List",moveToListConfirm:"Move to requisition list",actionCopyToList:"Copy to List",copyToListTitle:"Copy to Requisition List",copyToListConfirm:"Copy to requisition list",emptyRequisitionList:" Requisition List is empty",productListTable:{headers:{productName:"Product name",sku:"SKU",price:"Price",quantity:"Quantity",subtotal:"Subtotal",actions:"Actions"},itemQuantity:"Item quantity",outOfStock:"Out of stock",onlyXLeftInStock:"Only {count} left in stock"},errorLoadPage:"Failed to load page",errorLoadingProducts:"Failed to load product data",notFoundTitle:"Requisition List Not Found",notFoundMessage:"The requisition list you are looking for does not exist or you do not have access to it.",notFoundActionLabel:"Back to Requisition Lists"},ShareRequisitionListContent:{emailInstruction:"Select below the email addresses of the people with whom you'd like to share your list. Your name, email address and the details of your list will be shared with recipients. You can send your list with multiple recipients. Please note only registered users will be able to import lists.",emailLabel:"Email addresses",emailPlaceholder:"Select company users",submitLabel:"Submit",linkInstruction:"Copy the link below and send it to people with whom you'd like to share your list. Please note only registered users will be able to import lists.",copyLink:"Copy Link",linkCopied:"Copied!",loadingUsers:"Loading users...",loadingLink:"Generating share link...",noUsersAvailable:"No company colleagues available to share with.",usersLoadError:"Unable to load company users. Please try again.",maxRecipientsValidation:"You can select up to {max} recipients.",shareSuccessMessage:"You have shared your list with the below email addresses:"},RequisitionListsNotEnabled:{title:"Requisition Lists Not Available",message:"Requisition Lists are not available. Please contact your administrator for more information.",actionLabel:"Go to My Account"},PageSizePicker:{show:"Show",itemsPerPage:"Items per page"},PaginationItemsCounter:{itemsCounter:"Items {from}-{to} of {total}"}},g={RequisitionList:q},h={default:g},R=({children:s})=>{var t;const[o,n]=d("en_US");m(()=>{const e=L.on("locale",l=>{n(l)},{eager:!0});return()=>{e==null||e.off()}},[]);const a=(t=y.getConfig())==null?void 0:t.langDefinitions,r=u(h,a??{});return i(p,{lang:o,langDefinitions:r,children:s})},w=new c(i(R,{}));export{w as render};
 //# sourceMappingURL=render.js.map
